### 2023-01-16 #73 @Zeitsperre
Xclim : Xarray-based climate data analytics
Submitting Author: Trevor James Smith (@Zeitsperre)
All current maintainers: (@Zeitsperre, @tlogan2000, @aulemahal)
Package Name: xclim
One-Line Description of Package: Climate indices computation package based on Xarray
Repository Link:  https://github.com/Ouranosinc/xclim
Version submitted: v0.40.0
Version accepted: TBD

---

`xclim` is an operational Python library for climate services, providing numerous climate-related indicator tools with an extensible framework for constructing custom climate indicators, statistical downscaling and bias adjustment of climate model simulations, as well as climate model ensemble analysis tools.

xclim is built using `xarray`_ and can seamlessly benefit from the parallelization handling provided by `dask`. Its objective is to make it as simple as possible for users to perform typical climate services data treatment workflows. Leveraging xarray and dask, users can easily bias-adjust climate simulations over large spatial domains or compute indices from large climate datasets.

> *Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.*

  - Who is the target audience, and what are scientific applications of this package?

`xclim` aims to position itself as a climate services tool for any researchers interested in using [Climate and Forecast Conventions](http://cfconventions.org/) compliant datasets to perform climate analyses. This tool is optimized for working with Big Data in the climate science domain and can function as an independent library for one-off analyses in Jupyter notebooks or as a backend engine for performing climate data analyses over PyWPS (e.g. [Finch](https://github.com/bird-house/finch)). It was primarily developed targeting earth and environmental science audiences and researchers, originally for calculating climate indicators for the Canadian government web service [ClimateData.ca](https://climatedata.ca/).

The primary domains that xclim is built for are in calculating climate indicators, performing statistical correction / bias adjustment of climate model output variables/simulations, and in performing climate model simulation ensemble statistics.

[icclim](https://icclim.readthedocs.io/) is another library for the computation of climate indices. Starting with version 5.0 of icclim, some of the core computations rely on xclim. See [explanations](https://icclim.readthedocs.io/en/stable/explanation/xclim_and_icclim.html) about differences between xclim and icclim.

[scikit-downscale](https://scikit-downscale.readthedocs.io/) is a library offering algorithms for statistical downscaling. `xclim` drew inspiration for its fit-predict architecture. The suite of downscaling algorithms offered differ.

- [x] has an [OSI approved license][OsiApprovedLicense].
- [x] contains a README with instructions for installing the development version. (*Under CONTRIBUTING*)

- [x] The package is not a "minor utility" as defined by JOSS's [submission requirements][JossSubmissionRequirements]: "Minor 'utility' packages, including 'thin' API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.
- [x] The package is deposited in a long-term repository with the DOI: https://doi.org/10.5281/zenodo.2795043

- [x] Yes, I am OK with reviewers submitting requested changes as issues to my repo. Reviewers will then link to the issues in their submitted review.

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

- [x] [Last but not least, please fill out our pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us track

[Editor and review templates can be found here][Templates]

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
