### 2023-01-16 #73 @Zeitsperre
To whom it may concern, our paper is available on a separate branch within xclim. The Pull Request containing the paper can be found here: https://github.com/Ouranosinc/xclim/pull/250
