### 2023-01-17 #73 @lwasser
hi @Zeitsperre welcome to pyOpenSci!! are you a part of the pangeo community by chance?  i think i saw xclim discussed in a few pangeo threads.
