### 2023-01-17 #73 @lwasser
awesome! nice to meet you as well @Zeitsperre ! I asked because we are working on a partnership with pangeo to curate that list via peer review. as a part of that curation, we have a small set of checks that are specific to the pangeo community including:

- [ ] ​​Consume and produce high-level data structures (e.g. xarray datasets / pandas dataframes) wherever feasible
- [ ] ​​Operate lazily when called on dask data structure
- [ ] ​​Avoid file I/O unless specifically an I/O package

So we'd want to include you in that given your are in the list of tools. i haven't done a deep dive into your package but these are universal pangeo standards.If the review goes smoothly, we'd then tag your package as being pangeo community related / vetted.

do this package adhere to the above pangeo standards? Many thanks - i'm just trying to figure out the best path here to support our partnership :) i just started working on this and have an [open pr here](https://github.com/pyOpenSci/software-review/pull/70/files) to update our review submission template!
