### 2023-01-17 #73 @Zeitsperre
I believe we tick all the boxes you listed:

* [x]  ​​Consume and produce high-level data structures (e.g. xarray datasets / pandas dataframes) wherever feasible  (**all functions operate on either xr.DataArray or xr.Dataset objects**)
* [x]  ​​Operate lazily when called on dask data structure (**both indices and indicators seemlessly integrate with `dask`**)
* [x]  ​​Avoid file I/O unless specifically an I/O package (**`xarray` is left to handle all file loading or writing**)

I know that the core developers of xclim would gladly welcome feedback on how to better organize/structure our code base, and xclim being considered high enough in quality to be formally Pangeo-endorsed would be great!
