### 2023-01-17 #73 @lwasser
perfect! Thank you @Zeitsperre for the speedy replies!  I am going to followup and also get our editor in chief involved @NickleDave I stepped in only because i'm actively working with pangeo now so i just wanted to check in to see if this review could be supported via that collaboration as well. And it does sound like it could. more soon!
