### 2023-01-18 657583e
Add snow cover indices for snw variable
