+++ xclim/testing/tests/test_indices.py
@@ -2438,45 +2438,73 @@ def test_days_with_snow(prsn_series):
     assert out.units == "d"
 
 
-def test_snow_cover_duration(snd_series):
-    a = np.ones(366) / 100.0
-    a[10:20] = 0.3
-    snd = snd_series(a)
-    out = xci.snow_cover_duration(snd)
-    assert len(out) == 2
-    assert out[0] == 10
+class TestSnowCover:
+    def test_snow_cover_duration(self, snd_series, snw_series):
+        a = np.ones(366) / 100.0
+        a[10:20] = 0.3
 
+        snd = snd_series(a)
+        snw = snw_series(a)
 
-def test_continous_snow_cover_start(snd_series):
-    snd = snd_series(np.arange(366) / 100.0)
-    out = xci.continuous_snow_cover_start(snd)
-    assert len(out) == 2
-    np.testing.assert_array_equal(out, [snd.time.dt.dayofyear[0].data + 2, np.nan])
-    for attr in ["units", "is_dayofyear", "calendar"]:
-        assert attr in out.attrs.keys()
-    assert out.attrs["units"] == ""
-    assert out.attrs["is_dayofyear"] == 1
+        out = xci.snow_depth_cover_duration(snd)
+        assert len(out) == 2
+        assert out[0] == 10
 
+        out = xci.snow_water_cover_duration(snw)
+        assert len(out) == 2
+        assert out[0] == 10
 
-def test_continuous_snow_cover_end(snd_series):
-    a = np.concatenate(
-        [
-            np.zeros(100),
-            np.arange(10),
-            10 * np.ones(100),
-            10 * np.arange(10)[::-1],
-            np.zeros(146),
-        ]
-    )
-    snd = snd_series(a / 100.0)
-    out = xci.continuous_snow_cover_end(snd)
-    assert len(out) == 2
-    doy = snd.time.dt.dayofyear[0].data
-    np.testing.assert_array_equal(out, [(doy + 219) % 366, np.nan])
-    for attr in ["units", "is_dayofyear", "calendar"]:
-        assert attr in out.attrs.keys()
-    assert out.attrs["units"] == ""
-    assert out.attrs["is_dayofyear"] == 1
+    def test_continous_snow_cover_start(self, snd_series, snw_series):
+        a = np.arange(366) / 100.0
+        snd = snd_series(a)
+        snw = snw_series(a)
+
+        out = xci.continuous_snow_depth_cover_start(snd)
+        assert len(out) == 2
+        np.testing.assert_array_equal(out, [snd.time.dt.dayofyear[0].data + 2, np.nan])
+        for attr in ["units", "is_dayofyear", "calendar"]:
+            assert attr in out.attrs.keys()
+        assert out.attrs["units"] == ""
+        assert out.attrs["is_dayofyear"] == 1
+
+        out = xci.continuous_snow_water_cover_start(snw)
+        assert len(out) == 2
+        np.testing.assert_array_equal(out, [snw.time.dt.dayofyear[0].data + 2, np.nan])
+        for attr in ["units", "is_dayofyear", "calendar"]:
+            assert attr in out.attrs.keys()
+        assert out.attrs["units"] == ""
+        assert out.attrs["is_dayofyear"] == 1
+
+    def test_continuous_snow_cover_end(self, snd_series, snw_series):
+        a = np.concatenate(
+            [
+                np.zeros(100),
+                np.arange(10),
+                10 * np.ones(100),
+                10 * np.arange(10)[::-1],
+                np.zeros(146),
+            ]
+        )
+        snd = snd_series(a / 100.0)
+        snw = snw_series(a / 100.0)
+
[…]
