+++ xclim/testing/tests/test_snow.py
@@ -16,26 +16,45 @@ class TestSnowDepth:
 
 
 class TestSnowCoverDuration:
-    def test_simple(self, snd_series):
+    def test_simple(self, snd_series, snw_series):
         snd = snd_series(np.ones(110), start="2001-01-01")
-        out = land.snow_cover_duration(snd, freq="M")
+        snw = snw_series(np.ones(110), start="2001-01-01")
+
+        out = land.snow_depth_cover_duration(snd, freq="M")
+        assert out.units == "days"
+        np.testing.assert_array_equal(out, [31, 28, 31, np.nan])
+
+        out = land.snow_water_cover_duration(snw, freq="M")
         assert out.units == "days"
         np.testing.assert_array_equal(out, [31, 28, 31, np.nan])
 
 
 class TestContinuousSnowCoverStartEnd:
-    def test_simple(self, snd_series):
+    def test_simple(self, snd_series, snw_series):
         a = np.zeros(365)
+
+        # snow depth
         a[100:200] = 0.03
         snd = snd_series(a, start="2001-07-01")
         snd = snd.expand_dims(lat=[0, 1, 2])
-        out = land.continuous_snow_cover_start(snd)
+        snw = snw_series(a, start="2001-07-01")
+        snw = snw.expand_dims(lat=[0, 1, 2])
+
+        out = land.continuous_snow_depth_cover_start(snd)
         assert out.units == ""
         np.testing.assert_array_equal(out.isel(lat=0), snd.time.dt.dayofyear[100])
 
-        out = land.continuous_snow_cover_end(snd)
+        out = land.continuous_snow_depth_cover_end(snd)
         assert out.units == ""
         np.testing.assert_array_equal(out.isel(lat=0), snd.time.dt.dayofyear[200])
+
+        out = land.continuous_snow_water_cover_start(snw)
+        assert out.units == ""
+        np.testing.assert_array_equal(out.isel(lat=0), snw.time.dt.dayofyear[100])
+
+        out = land.continuous_snow_water_cover_end(snw)
+        assert out.units == ""
+        np.testing.assert_array_equal(out.isel(lat=0), snw.time.dt.dayofyear[200])
 
 
 class TestSndMaxDoy:

[…]
