### 2023-01-18 157c0b1
Snow cover indices for var `snw`
