+++ xclim/indices/_threshold.py
@@ -1,6 +1,8 @@
 # noqa: D100
 from __future__ import annotations
 
+import warnings
+
 import numpy as np
 import xarray
 
@@ -39,6 +41,10 @@ __all__ = [
     "cooling_degree_days",
     "continuous_snow_cover_end",
     "continuous_snow_cover_start",
+    "continuous_snow_depth_cover_end",
+    "continuous_snow_depth_cover_start",
+    "continuous_snow_water_cover_end",
+    "continuous_snow_water_cover_start",
     "days_with_snow",
     "growing_degree_days",
     "growing_season_start",
@@ -58,6 +64,8 @@ __all__ = [
     "hot_spell_frequency",
     "hot_spell_max_length",
     "snow_cover_duration",
+    "snow_depth_cover_duration",
+    "snow_water_cover_duration",
     "tn_days_above",
     "tn_days_below",
     "tg_days_above",
@@ -233,14 +241,59 @@ def continuous_snow_cover_end(
     window: int = 14,
     freq: str = "AS-JUL",
 ) -> xarray.DataArray:
-    r"""End date of continuous snow cover.
+    r"""End date of continuous snow depth cover.
 
-    First day after the start of the continuous snow cover when snow depth is below a threshold (default: 2 cm)
+    First day after the start of the continuous snow depth cover when snow depth is below a threshold (default: 2 cm)
     for at least `N` (default: 14) consecutive days.
 
     Warnings
     --------
-    The default `freq` is valid for the northern hemisphere.
+    * The default `freq` is valid for the northern hemisphere.
+    * This index will be removed in a future version of xclim.
+
+    Parameters
+    ----------
+    snd : xarray.DataArray
+        Surface snow thickness.
+    thresh : str
+        Threshold snow thickness.
+    window : int
+        Minimum number of days with snow depth below threshold.
+    freq : str
+        Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray, [dimensionless]
+        First day after the start of the continuous snow depth cover when the snow depth
+        goes below a threshold for a minimum duration.
+        If there is no such day, returns np.nan.
+
+    References
+    ----------
+    :cite:cts:`chaumont_elaboration_2017`
+    """
+    warnings.warn(
+        "The `continuous_snow_cover_end` is being deprecated in favour of `continuous_snow_depth_cover_end`"
+        "This indice will be removed in `xclim>=0.42.0`. Please update your scripts accordingly.",
+        DeprecationWarning,
+        stacklevel=3,
+    )
+    return continuous_snow_depth_cover_end(snd, thresh, window, freq)
+
+
+@declare_units(snd="[length]", thresh="[length]")
+def continuous_snow_depth_cover_end(
+    snd: xarray.DataArray, thresh: str = "2 cm", window: int = 14, freq: str = "AS-JUL"
+) -> xarray.DataArray:
+    r"""End date of continuous snow depth cover.
+
+    First day after the start of the continuous snow depth cover when snow depth is below a threshold (default: 2 cm)
+    for at least `N` (default: 14) consecutive days.
+
+    Warnings
+    --------
+    * The default `freq` is valid for the northern hemisphere.
 
     Parameters
     ----------
@@ -256,7 +309,7 @@ def continuous_snow_cover_end(
     Returns
     -------
     xarray.DataArray, [dimensionless]
-        First day after the start of the continuous snow cover when the snow depth
+        First day after the start of the continuous snow depth cover when the snow depth
         goes below a threshold for a minimum duration.
         If there is no such day, returns np.nan.
 
@@ -276,6 +329,56 @@ def continuous_snow_cover_end(
[…]
