### 2023-01-18 #73 @NickleDave
Hi again @Zeitsperre -- Happy to report that `xclim` passed initial editor-in-chief checks (below) with flying colors.

I have one suggestion, which we strongly recommend, but do not require:
- [ ] switch to pyproject.toml with static metadata, to adhere to pep 517 + 621, away from legacy setup.py file
  - looks like you mainly use setup.py to clean up README for build if I'm not mistaken?
  - I wonder if you could use other tools to the same ends? E.g., if you use hatch in place of setuptools there's a plug-in that will "prettify" the README for you, see https://talkpython.fm/episodes/show/395/tools-for-readme.md-creation-and-maintenance
  - I know you have a lot of config stuff in the setup.cfg, so switching to another dev tool might be a hassle, but having metadata in the pyproject.toml instead of a setup.py would be ideal

We expect to know how we will go about selecting editors by the *end of day Monday** (Jan. 23rd).
@lwasser needs to talk a little more with people about the Pangeo collaboration.
We will report back to you then with a timeline for the review.

  - [x] The package imports properly into a standard Python environment `import package-name`.
    - verified for both pip (version 0.40.0)
    - two notes:
      - I wonder if it would help to move some text from about.md to the index? So a user immediately knows what `xclim` is and does. Getting hit with a long table of contents can be daunting
      - "Basic usage" is great although looks like it could use maybe use some very minor revision of text just for flow + readability
    - plenty of tutorials, looks excellent
  - [x] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---
