### 2023-01-19 #73 @Zeitsperre
Thanks, @NickleDave! I'm glad to see that all the hours spent reading PEPs and adopting good practices has been helpful here!

I'll be opening a PR later today to address the PEP 517 and 621 compliance comment. We don't do anything too fancy in our `setup.py`, so I don't imagine it should be difficult to adopt `pyproject.toml`. I'll link that here once it's underway.
