### 2023-01-19 e90996a
Ensure all-NaN slices yield NaN outputs
