### 2023-01-19 6f93b0b
Support all-NaN slices in {snw,snd}_max_doy
