+++ xclim/indices/_hydrology.py
@@ -136,7 +136,9 @@ def snd_max_doy(snd: xarray.DataArray, freq: str = "AS-JUL") -> xarray.DataArray
     valid = at_least_n_valid(snd.where(snd > 0), n=1, freq=freq)
 
     # Compute doymax. Will return first time step if all snow depths are 0.
-    out = generic.select_resample_op(snd, op=generic.doymax, freq=freq)
+    out = generic.select_resample_op(
+        snd.where(snd > 0, 0), op=generic.doymax, freq=freq
+    )
     out.attrs.update(units="", is_dayofyear=np.int32(1), calendar=get_calendar(snd))
 
     # Mask arrays that miss at least one non-null snd.
@@ -190,7 +192,9 @@ def snw_max_doy(snw: xarray.DataArray, freq: str = "AS-JUL") -> xarray.DataArray
     valid = at_least_n_valid(snw.where(snw > 0), n=1, freq=freq)
 
     # Compute doymax. Will return first time step if all snow depths are 0.
-    out = generic.select_resample_op(snw, op=generic.doymax, freq=freq)
+    out = generic.select_resample_op(
+        snw.where(snw > 0, 0), op=generic.doymax, freq=freq
+    )
     out.attrs.update(units="", is_dayofyear=np.int32(1), calendar=get_calendar(snw))
 
     # Mask arrays that miss at least one non-null snd.
