### 2023-01-19 3ea77fa
remove code-autolink in README.rst
