### 2023-01-19 ab14a40
WIP - use pyproject.toml and migrate some configs from setup.cfg
