+++ setup.cfg
@@ -3,13 +3,13 @@ current_version = 0.40.0
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+).(?P<patch>\d+)(\-(?P<release>[a-z]+))?
-serialize = 
+serialize =
 	{major}.{minor}.{patch}-{release}
 	{major}.{minor}.{patch}
 
 [bumpversion:part:release]
 optional_value = gamma
-values = 
+values =
 	beta
 	gamma
 
@@ -21,19 +21,15 @@ replace = VERSION = "{new_version}"
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
-[coverage:run]
-relative_files = True
-omit = */tests/*.py
-
 [flake8]
-exclude = 
+exclude =
 	.git,
 	docs,
 	build,
 	.eggs,
 max-line-length = 88
 max-complexity = 12
-ignore = 
+ignore =
 	C901
 	E203
 	E231
@@ -43,12 +39,12 @@ ignore =
 	F403
 	W503
 	W504
-per-file-ignores = 
+per-file-ignores =
 	tests/*:E402
-rst-directives = 
+rst-directives =
 	bibliography
 	autolink-skip
-rst-roles = 
+rst-roles =
 	doc,
 	mod,
 	py:attr,
@@ -67,7 +63,7 @@ rst-roles =
 	cite:p
 	cite:t
 	cite:ts
-extend-ignore = 
+extend-ignore =
 	RST399,
 	RST201,
 	RST203,
@@ -75,31 +71,10 @@ extend-ignore =
 	RST301,
 	RST306
 
-[aliases]
-test = pytest
-
-[tool:pytest]
-addopts = 
-	--verbose
-	--cov=xclim
-	--cov-report=term-missing
-	--numprocesses=auto
-	--maxprocesses=6
-	--dist=loadscope
-norecursedirs = docs/notebooks/*
-filterwarnings = 
-	ignore::UserWarning
-testpaths = xclim/testing/tests xclim/testing/tests/test_sdba
-usefixtures = xdoctest_namespace
-doctest_optionflags = NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL NUMBER ELLIPSIS
-markers = 
-	slow: marks tests as slow (deselect with '-m "not slow"')
-	requires_docs: mark tests that can only be run with documentation present
-
 [pycodestyle]
 count = False
 exclude = xclim/testing/tests
-ignore = 
+ignore =
 	E226,
 	E402,
 	E501,
@@ -109,8 +84,3 @@ statistics = True
 [pydocstyle]
 convention = numpy
 match = ((?!(test_|conftest|conf|rstjinja|autodoc_indicator)).)*\.py
-
-[isort]
-profile = black
-py_version = 38
-append_only = true
