### 2023-01-19 98ebd64
move dev recipe and MANIFEST configuration into pyproject.toml
