### 2023-01-20 de3be7e
add toml checks to pre-commit
