+++ .pre-commit-config.yaml
@@ -21,13 +21,10 @@ repos:
   - id: debug-statements
   - id: pretty-format-json
     args: ['--autofix', '--no-ensure-ascii', '--no-sort-keys']
-- repo: https://github.com/macisamuele/language-formatters-pre-commit-hooks
-  rev: v2.5.0
+- repo: https://github.com/pappasam/toml-sort
+  rev: v0.22.2
   hooks:
-  - id: pretty-format-toml
-    args: ['--autofix']
-#  - id: pretty-format-yaml
-#    args: ['--autofix', '--indent', '2', '--preserve-quotes']
+  - id: toml-sort-fix
 - repo: https://github.com/pycqa/flake8
   rev: 6.0.0
   hooks:
@@ -44,7 +41,7 @@ repos:
   - id: isort
     args: ['--settings-file=pyproject.toml', "--add_imports='from __future__ import annotations'"]
 - repo: https://github.com/nbQA-dev/nbQA
-  rev: 1.6.0
+  rev: 1.6.1
   hooks:
   - id: nbqa-black
     args: ['--target-version=py38']
@@ -58,7 +55,7 @@ repos:
   - id: nbstripout
     files: '.ipynb'
 - repo: https://github.com/pycqa/pydocstyle
-  rev: 6.2.0
+  rev: 6.3.0
   hooks:
   - id: pydocstyle
     args: ['--config=setup.cfg']
@@ -70,7 +67,7 @@ repos:
     additional_dependencies:
     - 'black==22.12.0'
 - repo: https://github.com/adrienverge/yamllint.git
-  rev: v1.28.0
+  rev: v1.29.0
   hooks:
   - id: yamllint
     args: ['--config-file=.yamllint.yaml']
