### 2023-01-20 5e7688c
move dependencies to pyproject.toml, remove versioning in setup.cfg
