+++ setup.cfg
@@ -13,10 +13,6 @@ values =
 	beta
 	gamma
 
-[bumpversion:file:setup.py]
-search = VERSION = "{current_version}"
-replace = VERSION = "{new_version}"
-
 [bumpversion:file:xclim/__init__.py]
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
