+++ pyproject.toml
@@ -5,115 +5,122 @@ build-backend = "flit_core.buildapi"
 [project]
 name = "xclim"
 authors = [
-    { name="Travis Logan", email="logan.travis@ouranos.ca" }
-    ]
+  {name = "Travis Logan", email = "logan.travis@ouranos.ca"}
+]
 maintainers = [
-    { name="Trevor James Smith", email="smith.trevorj@ouranos.ca" },
-    { name="Pascal Bourgault", email="bourgault.pascal@ouranos.ca" }
-    ]
+  {name = "Trevor James Smith", email = "smith.trevorj@ouranos.ca"},
+  {name = "Pascal Bourgault", email = "bourgault.pascal@ouranos.ca"}
+]
 requires-python = ">=3.8.0"
 keywords = ["xclim", "xarray", "climate", "climatology", "bias correction", "ensemble", "indicators", "analysis"]
-license = { file="LICENSE" }
-classifiers=[
-        "Development Status :: 4 - Beta",
-        "Intended Audience :: Developers",
-        "Intended Audience :: Science/Research",
-        "License :: OSI Approved :: Apache Software License",
-        "Natural Language :: English",
-        "Operating System :: OS Independent",
-        "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.8",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Topic :: Scientific/Engineering :: Atmospheric Science",
-    ]
+license = {file = "LICENSE"}
+classifiers = [
+  "Development Status :: 4 - Beta",
+  "Intended Audience :: Developers",
+  "Intended Audience :: Science/Research",
+  "License :: OSI Approved :: Apache Software License",
+  "Natural Language :: English",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.8",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Topic :: Scientific/Engineering :: Atmospheric Science"
+]
 dynamic = ["description", "version"]
-
-[project.urls]
-"About Ouranos" =  "https://www.ouranos.ca/en/"
-"Changelog" = "https://xclim.readthedocs.io/en/stable/history.html"
-"Issue tracker" = "https://github.com/Ouranosinc/xclim/issues"
+dependencies = [
+  "boltons>=20.1",
+  "bottleneck>=1.3.1",
+  "cf-xarray>=0.6.1",
+  "cftime>=1.4.1",
+  "Click>=8.1",
+  "dask[array]>=2.6",
+  "jsonpickle",
+  "numba",
+  "numpy>=1.16",
+  "pandas>=0.23",
+  "pint>=0.10",
+  "pyyaml",
+  "scikit-learn>=0.21.3",
+  "scipy>=1.2,<1.9", # see: https://github.com/Ouranosinc/xclim/issues/1142
+  "statsmodels",
+  "xarray>=2022.06.0"
+]
 
 [project.optional-dependencies]
 dev = [
-    # dev tools and testing
-    "black >=22.12",
-    "blackdoc",
-    "bump2version",
-    "coverage",
-    "flake8",
-    "flake8-rst-docstrings",
-    "ipython",
-    "mypy",
-    "nbqa",
-    "nbval",
-    "pre-commit >=2.9",
[…]
