+++ tox.ini
@@ -27,12 +27,13 @@ deps =
     yamllint
 commands =
     pydocstyle --config=setup.cfg xclim
-    flake8 xclim
-    black --check --target-version py38 xclim
-    nbqa black --check docs --target-version py38
-    blackdoc --check --target-version py38 xclim --exclude xclim/indices/__init__.py
-    isort --check --settings-file=setup.cfg xclim
-    yamllint --config-file .yamllint.yaml xclim
+    flake8 --config=setup.cfg xclim
+    black --check xclim
+    nbqa black --check docs
+    blackdoc --check --exclude=xclim/indices/__init__.py xclim
+    blackdoc --check docs
+    isort --check xclim
+    yamllint --config-file=.yamllint.yaml xclim
 
 [testenv:docs]
 description = Build the documentation with makefile under {basepython}
