### 2023-01-20 a96b626
address formatting checks
