+++ xclim/testing/tests/test_generic_indicators.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import numpy as np
 
 from xclim import generic, set_options
