### 2023-01-20 35ddfd6
remove setup.py, adapt tox to use flit/pyproject.toml, fix coverage configuration error
