+++ pyproject.toml
@@ -105,7 +105,7 @@ target-version = [
 
 [tool.coverage.run]
 relative_files = true
-omit = "*/tests/*.py"
+omit = ["*/tests/*.py"]
 
 [tool.flit.sdist]
 include = ["AUTHORS.rst", "CONTRIBUTING.rst", "HISTORY.rst", "LICENSE", "README.rst", "requirements_upstream.txt", "xclim/**/*.py", "xclim/**/*.json", "xclim/**/*.yml", "xclim/**/*.txt", "docs/**/*.rst", "docs/conf.py", "docs/Makefile", "docs/make.bat", "docs/**/*.jpg", "docs/**/*.png", "docs/**/*.gif"]
