+++ tox.ini
@@ -7,9 +7,9 @@ env_list =
     py38-upstream-doctest
     py39-slow
     py310-lm3
+isolated_build = True
 requires =
     pip >= 21.0
-    setuptools
 opts = -v
 
 [testenv:black]
