### 2023-01-20 9a11652
add a long_description from pre-build script to pyproject.toml
