### 2023-01-20 80ff09d
static typing changes
