### 2023-01-20 90f2899
delete pre-build script
