+++ pyproject.toml
@@ -93,9 +93,11 @@ dev = [
 xclim = "xclim.cli:cli"
 
 [project.urls]
-"About Ouranos" = "https://www.ouranos.ca/en/"
+"Homepage" = "https://xclim.readthedocs.io/"
+"Source" = "https://github.com/Ouranosinc/xclim/"
 "Changelog" = "https://xclim.readthedocs.io/en/stable/history.html"
 "Issue tracker" = "https://github.com/Ouranosinc/xclim/issues"
+"About Ouranos" = "https://www.ouranos.ca/en/"
 
 [tool]
 
