### 2023-01-20 f26aad3
delete duplicate logos, move logos to a permanently situated folder, add exceptions to sdist generation
