+++ pyproject.toml
@@ -127,55 +127,21 @@ warn_return_any = true
 warn_unused_configs = true
 
 [[tool.mypy.overrides]]
-module = "xarray.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "numpy.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "scipy.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "boltons.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "lmoments3.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "dask.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "numba.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "pint.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "cftime.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "pandas.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "bottleneck.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "sklearn.cluster.*"
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "clisops.core.subset.*"
+module = [
+  "xarray.*",
+  "numpy.*",
+  "scipy.*",
+  "boltons.*",
+  "lmoments3.*",
+  "dask.*",
+  "numba.*",
+  "pint.*",
+  "cftime.*",
+  "pandas.*",
+  "bottleneck.*",
+  "sklearn.cluster.*",
+  "clisops.core.subset.*"
+]
 ignore_missing_imports = true
 
 [tool.pytest.ini_options]
