### 2023-01-23 6306fcc
update CI to reflect changes and update CONTRIBUTING.rst for flit
