### 2023-01-23 28743c9
add indicators and translations, update history
