+++ xclim/indicators/land/_snow.py
@@ -5,15 +5,17 @@ from xclim.core.indicator import Daily, ResamplingIndicatorWithIndexing
 
 __all__ = [
     "blowing_snow",
-    "snow_cover_duration",
-    "continuous_snow_cover_start",
     "continuous_snow_cover_end",
+    "continuous_snow_cover_start",
     "snd_max_doy",
+    "snow_amount_approximation",
+    "snow_cover_duration",
+    "snow_depth",
+    "snow_depth_approximation",
     "snow_melt_we_max",
     "snw_max",
     "snw_max_doy",
     "winter_storm",
-    "snow_depth",
 ]
 
 
@@ -155,4 +157,24 @@ snow_depth = SnowWithIndexing(
     abstract="Mean of daily snow depth.",
     cell_methods="time: mean over days",
     compute=xci.snow_depth,
+)
+
+snow_amount_approximation = SnowWithIndexing(
+    title="Surface snow amount",
+    identifier="snow_amount_approximation",
+    units="kg m-2",
+    standard_name="surface_snow_amount",
+    long_name="Approximation of daily snow amount from snow depth and density",
+    description="The approximation of daily snow amount from snow depth and density.",
+    compute=xci.snow_amount_approximation,
+)
+
+snow_depth_approximation = SnowWithIndexing(
+    title="Surface snow depth",
+    identifier="snow_depth_approximation",
+    units="m",
+    standard_name="surface_snow_thickness",
+    long_name="Approximation of daily snow depth from snow amount and density",
+    description="The approximation of daily snow depth from snow amount and density.",
+    compute=xci.snow_depth_approximation,
 )
