+++ HISTORY.rst
@@ -2,6 +2,18 @@
 History
 =======
 
+0.41.0 (unreleased)
+-------------------
+Contributors to this version: Trevor James Smith (:user:`Zeitsperre`)
+
+New indicators
+^^^^^^^^^^^^^^
+* New indices and indicators for converting from snow water equivalent to snow depth (``snow_depth_approximation``) and snow depth to snow water equivalent (``snow_amount_approximation``) using snow density [kg/m^3].
+
+Internal changes
+^^^^^^^^^^^^^^^^
+* Some grammar and spelling corrections were applied to dostrings. (:pull:`1271`).
+
 0.40.0 (2023-01-13)
 -------------------
 Contributors to this version: Trevor James Smith (:user:`Zeitsperre`), Pascal Bourgault (:user:`aulemahal`), David Huard (:user:`huard`), Juliette Lavoie (:user:`juliettelavoie`).
