### 2023-01-23 b52be66
add missing variable snr
