### 2023-01-23 69ab47b
remove unnecessary treatments in atmosds
