### 2023-01-24 9c8f46a
exclude tests from flake8 explicitly
