+++ setup.cfg
@@ -23,6 +23,7 @@ exclude =
 	docs,
 	build,
 	.eggs,
+    xclim/testing/tests
 max-line-length = 88
 max-complexity = 12
 ignore =
