### 2023-01-24 ae51e5e
Apply suggestions from code review
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
