### 2023-01-24 5f82fe6
use loadscope for doctests
