+++ Makefile
@@ -66,7 +66,7 @@ lint: ## check style with flake8 and black
 
 test: ## run tests quickly with the default Python
 	pytest xclim/testing/tests
-	pytest --nbval docs/notebooks
+	pytest --nbval --dist=loadscope docs/notebooks
 	pytest --rootdir xclim/testing/tests/ --xdoctest xclim
 
 test-all: ## run tests on every Python version with tox
@@ -94,11 +94,10 @@ servedocs: docs ## compile the docs watching for changes
 	watchmedo shell-command -p '*.rst' -c '$(MAKE) -C docs html' -R -D .
 
 release: dist ## package and upload a release
-	twine upload dist/*
+	flit publish dist/*
 
 dist: clean ## builds source and wheel package
-	python setup.py sdist
-	python setup.py bdist_wheel
+	flit build
 	ls -l dist
 
 install: clean ## install the package to the active Python's site-packages
