### 2023-01-24 d4622dd
update workflows and tox for coveralls support of pyproject.toml
