+++ tox.ini
@@ -82,6 +82,7 @@ setenv =
 passenv =
     CI
     CONDA_EXE
+    COVERALLS_*
     GITHUB_*
     LD_LIBRARY_PATH
 extras = dev
