### 2023-01-24 6c16fe1
configure coveralls service name
