+++ tox.ini
@@ -77,7 +77,6 @@ commands =
 description = Run tests with pytest under {basepython}
 setenv =
     COV_CORE_SOURCE =
-    PYTEST_ADDOPTS = "--color=yes"
     PYTHONPATH = {toxinidir}
 passenv =
     CI
