### 2023-01-24 c59b4fd
try unpinning netcdf4
