### 2023-01-24 1b9367e
mark test as xfailing due to multithreading with netcdf4
