### 2023-01-24 e19473a
Bump version: 0.40.0 → 0.40.1-beta
