### 2023-01-24 #73 @Zeitsperre
@NickleDave

I just merged changes to address some of your comments, primarily for `pyproject.toml` and the ReadMe suggestion. I opted to adopt `flit` as it seemed to be a very stable and no-nonsense setup engine. It took a bit of time to migrate some configurations around, but it looks/works much better.

Agreed on your point about `Basic Usage`. I'll be opening a ticket to address that down-the-line. There are a fair amount of Pull Requests currently open, but we tend to address things in a reasonable amount of time.

Let me know if you have any other suggestions, thanks!
