### 2023-01-24 Ouranosinc/xclim#1282 @Zeitsperre
[PyOpenSci] Basic Usage notebook needs cleanup / better flow
### Generic Issue

* xclim version: 0.40.0
* Python version: Any
* Operating System: All

### Description

One of the suggestions coming from the PyOpenSci review process was to clean up the `Basic Usage` notebook to make it a bit more consistent with the rest of the documentation (https://github.com/pyOpenSci/software-submission/issues/73). The flow of the text needs a bit of work.

For context, this issue is not blocking the PyOpenSci review process, but would be nice to have.

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
