### 2023-01-24 #73 @NickleDave
Hi @Zeitsperre!

> I just merged changes to address some of your comments

Awesome! Glad to hear it.

> Agreed on your point about Basic Usage.

Thank you for even making note of that. The docs look really great over all and I'm sure they'll look better after the review.

I'm sorry for not getting back to you by the end of the day yesterday--totally my fault.

I want to let you know we do have an editor, @Batalex! 🎉🎉🎉 who will introduce themselves and officially start the review in the next couple of days. We are actively recruiting reviewers right now as well.
