+++ xclim/testing/tests/data.py
@@ -13,7 +13,7 @@ from xclim.testing import get_local_testdata as _get_local_testdata
 from xclim.testing import open_dataset as _open_dataset
 from xclim.testing.utils import _default_cache_dir  # noqa
 
-MAIN_TESTDATA_BRANCH = os.getenv("MAIN_TESTDATA_BRANCH", "main")
+MAIN_TESTDATA_BRANCH = os.getenv("MAIN_TESTDATA_BRANCH", "fix_era5_evspsblpot")
 TD = Path(__file__).parent / "data"
 
 
