### 2023-01-25 68b9726
Thresholds changed from `str` to `Quantified`
