+++ xclim/testing/tests/test_precip.py
@@ -619,8 +619,12 @@ def test_dry_spell_frequency_op(open_dataset):
         pr, thresh="3 mm", window=7, freq="MS", op="max"
     )
 
-    np.testing.assert_allclose(test_sum[0, 2], [1], rtol=1e-1)
-    np.testing.assert_allclose(test_max[0, 2], [2], rtol=1e-1)
+    np.testing.assert_allclose(
+        test_sum[0, :14], [1, 0, 1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0], rtol=1e-1
+    )
+    np.testing.assert_allclose(
+        test_max[0, :14], [1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 2, 1], rtol=1e-1
+    )
     assert (
         "The monthly number of dry periods of 7 day(s) or more, "
         "during which the total precipitation on a window of 7 day(s) is below 3 mm."
