+++ xclim/testing/tests/test_run_length.py
@@ -368,7 +368,7 @@ def test_keep_longest_run_synthetic():
 
 def test_keep_longest_run_data(open_dataset):
     era5 = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc")
-    cond = era5.swe > 0.001
+    cond = era5.swe > 0.002
     lrun = rl.keep_longest_run(cond, "time")
     np.testing.assert_array_equal(
         lrun.isel(time=slice(651, 658), location=2),
