### 2023-01-27 6b34822
First fix and simplify func
