### 2023-01-30 84314ae
update CI, try testing against Python3.11
