### 2023-01-30 42adcba
reconfigure tox and CI
