+++ tox.ini
@@ -4,9 +4,9 @@ env_list =
     docs
     doctests_only
 ;    opt-slow
-    py38-upstream-doctest
-    py39-slow
-    py310-lm3
+    py39-upstream-doctest
+    py310-slow
+    py311-lm3
 isolated_build = True
 requires =
     pip >= 21.0
