+++ .github/workflows/main.yml
@@ -58,7 +58,7 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
-  test-py311-upstream-lm3:
+  test-py39-upstream-lm3:
     needs: black
     if: |
       contains(github.event.pull_request.labels.*.name, 'approved') ||
@@ -70,18 +70,18 @@ jobs:
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
-          python-version: "3.11"
+          python-version: "3.9"
       - name: Install tox
         run: pip install tox
       - name: Test with tox
         run: tox -e py38-upstream-lm3-coverage
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          COVERALLS_FLAG_NAME: run-py311-upstream-lm3
+          COVERALLS_FLAG_NAME: run-py39-upstream-lm3
           COVERALLS_PARALLEL: true
           COVERALLS_SERVICE_NAME: github
 
-  test-py39-opt-slow:
+  test-py310-opt-slow:
     needs: black
     if: |
       contains(github.event.pull_request.labels.*.name, 'approved') ||
@@ -100,7 +100,7 @@ jobs:
           micromamba-version: 'latest'
           environment-file: ci/environment.yml
           extra-specs: |
-            python-version=3.9
+            python-version="3.10"
       - name: Conda and Mamba versions
         run: |
           mamba update -n base conda mamba
@@ -111,7 +111,6 @@ jobs:
           python -m pip install --no-user .
       - name: Test
         run: |
-          conda activate xclim39
           pip check
           pytest xclim --durations=10
         env:
@@ -135,11 +134,11 @@ jobs:
           coveralls
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          COVERALLS_FLAG_NAME: run-py39-opt-slow
+          COVERALLS_FLAG_NAME: run-py310-opt-slow
           COVERALLS_PARALLEL: true
           COVERALLS_SERVICE_NAME: github
 
-  test-py38-lm3:
+  test-py311-lm3:
     needs: black
     if: |
       contains(github.event.pull_request.labels.*.name, 'approved') ||
@@ -151,18 +150,18 @@ jobs:
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
-          python-version: "3.8"
+          python-version: "3.11"
       - name: Install tox
         run: pip install tox~=3.0
       - name: Test with tox
-        run: tox -e py310-lm3-coverage
+        run: tox -e py311-lm3-coverage
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          COVERALLS_FLAG_NAME: run-py38-lm3
+          COVERALLS_FLAG_NAME: run-py311-lm3
           COVERALLS_PARALLEL: true
           COVERALLS_SERVICE_NAME: github
 
-  doctests-py310:
+  doctests-py38:
     needs: black
     if: |
       contains(github.event.pull_request.labels.*.name, 'approved') ||
@@ -173,7 +172,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
         with:
-          python-version: "3.10"
+          python-version: "3.8"
       - name: Install pylint and tox
         run: pip install pylint tox
       - name: Run pylint
@@ -182,15 +181,15 @@ jobs:
         run: tox -e doctests_only-coverage
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
[…]
