### 2023-01-30 e754fea
python3.11 not supported by numba, adjust configurations
