+++ .github/workflows/main.yml
@@ -100,7 +100,7 @@ jobs:
           micromamba-version: 'latest'
           environment-file: environment.yml
           extra-specs: |
-            python-version="3.10"
+            python="3.10"
       - name: Conda and Mamba versions
         run: |
           mamba update -n base conda mamba
