+++ .github/workflows/main.yml
@@ -104,7 +104,8 @@ jobs:
       - name: Conda and Mamba versions
         run: |
           micromamba update -n base conda micromamba
-          micromamba --version
+          conda --version
+          echo "micromamba: $(micromamba --version)"
       - name: Install dependencies
         run: |
           python -m pip install --no-user "git+https://github.com/Ouranosinc/SBCK.git@easier-install#egg=sbck&subdirectory=python"
