+++ xclim/ensembles/_robustness.py
@@ -47,7 +47,7 @@ def change_significance(
         Weights to apply along the 'realization' dimension. This array cannot contain missing values.
         Note: 'ttest' and 'welch-ttest' are not currently supported with weighted arrays.
     realization: str
-        Name of the realization dimension. If you don't have multiple cliamte realizations or models, chosse None.
+        Name of the realization dimension. If you don't have multiple climate realizations or models, choose None.
     **kwargs
         Other arguments specific to the statistical test.
 
