### 2023-01-31 15cf0de
If 'realization' is None, create a dummy realization
