+++ xclim/ensembles/_robustness.py
@@ -265,6 +265,15 @@ def change_significance(
         f"Significant change was tested with test {test} with parameters {kwargs_str}."
     )
     das = {"fut": fut} if ref is None else {"fut": fut, "ref": ref}
+    changed.attrs.update(
+        description="Change significance. " + test_str,
+        units="",
+        test=str(test),
+        histroy=update_history(
+            f"change_significance(fut=fut, ref=ref, test={test}, {kwargs_str})",
+            **das,
+        )
+    )
     pos_frac.attrs.update(
         description="Fraction of members showing significant change that agree on a positive change. "
         + test_str,
@@ -284,7 +293,7 @@ def change_significance(
             **das,
         ),
     )
-    return change_frac, pos_frac
+    return changed, change_frac, pos_frac
 
 
 def robustness_coefficient(
