+++ xclim/ensembles/_partitioning.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import numpy as np
 import pandas as pd
 import xarray as xr
