+++ docs/references.bib
@@ -413,19 +413,20 @@ Section: Bulletin of the American Meteorological Society},
 }
 
 @article{thornthwaite_approach_1948,
-	title = {An {Approach} {Toward} a {Rational} {Classification} of {Climate}},
-	volume = {66},
-	issn = {0038-075X},
-	url = {https://journals.lww.com/soilsci/Citation/1948/07000/An_Approach_Toward_a_Rational_Classification_of.7.aspx},
-	language = {en-US},
+	title = {An {Approach} toward a {Rational} {Classification} of {Climate}},
+	volume = {38},
+	issn = {0016-7428},
+	url = {https://www.jstor.org/stable/210739},
+	doi = {10.2307/210739},
 	number = {1},
-	urldate = {2022-07-29},
-	journal = {Soil Science},
+	urldate = {2023-01-31},
+	journal = {Geographical Review},
 	author = {Thornthwaite, C. W.},
-	month = jul,
 	year = {1948},
-	pages = {77},
+	note = {Publisher: [American Geographical Society, Wiley]},
+	pages = {55--94},
 }
+
 
 @misc{brode_utci_2009,
 	title = {{UTCI}},
