+++ docs/references.bib
@@ -427,7 +427,6 @@ Section: Bulletin of the American Meteorological Society},
 	pages = {55--94},
 }
 
-
 @misc{brode_utci_2009,
 	title = {{UTCI}},
 	copyright = {MIT License},
