### 2023-01-31 337b97b
cleanup commands, use latest ubuntu image
