+++ .github/workflows/main.yml
@@ -21,7 +21,7 @@ on:
 
 jobs:
   black:
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     if: |
       ((github.event_name == 'pull_request') && (github.event.action != 'labeled')) ||
       (github.event.review.state == 'approved') ||
@@ -32,7 +32,8 @@ jobs:
         with:
           access_token: ${{ github.token }}
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
+      - name: Set up Python3.8
+        uses: actions/setup-python@v4
         with:
           python-version: "3.8"
       - name: Install tox
@@ -44,7 +45,7 @@ jobs:
     needs: black
     if: |
       (github.event_name == 'pull_request') && !contains(github.event.pull_request.labels.*.name, 'approved')
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python3.8
@@ -52,7 +53,7 @@ jobs:
         with:
           python-version: "3.8"
       - name: Install tox
-        run: pip install tox
+        run: pip install tox~=3.0
       - name: Test with tox
         run: tox -e py38
         env:
@@ -64,16 +65,16 @@ jobs:
       contains(github.event.pull_request.labels.*.name, 'approved') ||
       (github.event.review.state == 'approved') ||
       (github.event_name == 'push')
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }}
+      - name: Set up Python3.9
         uses: actions/setup-python@v4
         with:
           python-version: "3.9"
       - name: Install tox
         run: pip install tox
-      - name: Test with tox
+      - name: Test with tox~=3.0
         run: tox -e py39-upstream-lm3-coverage
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -87,13 +88,13 @@ jobs:
       contains(github.event.pull_request.labels.*.name, 'approved') ||
       (github.event.review.state == 'approved') ||
       (github.event_name == 'push')
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     defaults:
       run:
         shell: bash -l {0}
     steps:
       - uses: actions/checkout@v3
-      - name: Setup conda
+      - name: Setup Conda (Micromamba) with Python3.10
         uses: mamba-org/provision-with-micromamba@main
         with:
           cache-downloads: true
@@ -103,12 +104,11 @@ jobs:
             python="3.10"
       - name: Conda and Mamba versions
         run: |
-          micromamba update -n base conda micromamba
           conda --version
           echo "micromamba: $(micromamba --version)"
       - name: Install dependencies
         run: |
-          python -m pip install --no-user "git+https://github.com/Ouranosinc/SBCK.git@easier-install#egg=sbck&subdirectory=python"
+          python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git#egg=sbck"
           python -m pip install --no-user .
       - name: Test
         run: |
@@ -145,10 +145,10 @@ jobs:
       contains(github.event.pull_request.labels.*.name, 'approved') ||
       (github.event.review.state == 'approved') ||
       (github.event_name == 'push')
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }}
+      - name: Set up Python3.8
         uses: actions/setup-python@v4
         with:
[…]
