+++ .github/workflows/main.yml
@@ -108,7 +108,7 @@ jobs:
           echo "micromamba: $(micromamba --version)"
       - name: Install dependencies
         run: |
-          python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git#egg=sbck"
+          python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck"
           python -m pip install --no-user .
       - name: Test
         run: |
