### 2023-01-31 d69a020
split doctests and coverage steps
