+++ docs/conf.py
@@ -124,6 +124,7 @@ linkcheck_ignore = [
     r"http://www.utci.org/.*",  # Added on 2022-12-08: site appears to be down (timeout)
     r"https://ui.adsabs.harvard.edu/abs/2018AGUFMIN33A..06A*",  # Added on 2023-01-24: bad ssl certificate
     r"https://www.jstor.org/.*",  # Added on 2023-01-31: does not allow linkcheck requests (error 403)
+    r"https://doi.org/10.2307/210739",  # Added on 2023-01-31: does not allow linkcheck requests (error 403)
 ]
 linkcheck_exclude_documents = [r"readme"]
 
