+++ .pre-commit-config.yaml
@@ -32,7 +32,7 @@ repos:
     additional_dependencies: ['flake8-rst-docstrings']
     args: ['--config=setup.cfg']
 - repo: https://github.com/psf/black
-  rev: 22.12.0
+  rev: 23.1.0
   hooks:
   - id: black
 - repo: https://github.com/PyCQA/isort
@@ -65,7 +65,7 @@ repos:
   - id: blackdoc
     exclude: '(xclim/indices/__init__.py|docs/installation.rst)'
     additional_dependencies:
-    - 'black==22.12.0'
+    - 'black==23.1.0'
 - repo: https://github.com/adrienverge/yamllint.git
   rev: v1.29.0
   hooks:
