+++ tox.ini
@@ -10,6 +10,7 @@ env_list =
 isolated_build = True
 requires =
     pip >= 21.0
+    tox ~=3.0
 opts = -v
 
 [testenv:black]
