### 2023-02-01 6dd4d8b
Bump version: 0.40.2-beta → 0.40.3-beta
