### 2023-02-01 72a6365
add net radiation variables
