+++ xclim/data/variables.yml
@@ -79,6 +79,16 @@ variables:
     data_flags:
       - values_repeating_for_n_or_more_days:
           n: 5
+  rls:
+    canonical_units: W m-2
+    cell_methods: "time: mean"
+    description: Net longwave radiation.
+    standard_name: surface_net_downward_longwave_flux
+  rss:
+    canonical_units: W m-2
+    cell_methods: "time: mean"
+    description: Net shortwave radiation.
+    standard_name: surface_net_downward_shortwave_flux
   rlds:
     canonical_units: W m-2
     cell_methods: "time: mean"
