### 2023-02-01 1b67d72
add radiation converters as indicators
