+++ xclim/indices/_conversion.py
@@ -27,24 +27,26 @@ from xclim.indices.helpers import (
 )
 
 __all__ = [
-    "humidex",
+    "clausius_clapeyron_scaled_precipitation",
     "heat_index",
-    "tas",
-    "uas_vas_2_sfcwind",
-    "sfcwind_2_uas_vas",
-    "saturation_vapor_pressure",
+    "humidex",
+    "longwave_radiation_upwards_from_net_downwards",
+    "mean_radiant_temperature",
+    "potential_evapotranspiration",
+    "rain_approximation",
     "relative_humidity",
-    "specific_humidity",
-    "specific_humidity_from_dewpoint",
-    "snowfall_approximation",
+    "saturation_vapor_pressure",
+    "sfcwind_2_uas_vas",
+    "shortwave_radiation_upwards_from_net_downwards",
     "snow_amount_approximation",
     "snow_depth_approximation",
-    "rain_approximation",
-    "wind_chill_index",
-    "clausius_clapeyron_scaled_precipitation",
-    "potential_evapotranspiration",
+    "snowfall_approximation",
+    "specific_humidity",
+    "specific_humidity_from_dewpoint",
+    "tas",
+    "uas_vas_2_sfcwind",
     "universal_thermal_climate_index",
-    "mean_radiant_temperature",
+    "wind_chill_index",
 ]
 
 
