+++ xclim/indicators/atmos/_conversion.py
@@ -28,6 +28,8 @@ __all__ = [
     "corn_heat_units",
     "universal_thermal_climate_index",
     "mean_radiant_temperature",
+    "shortwave_radiation_upwards_from_net_downwards",
+    "longwave_radiation_upwards_from_net_downwards",
 ]
 
 
@@ -381,4 +383,31 @@ mean_radiant_temperature = Converter(
     cell_methods="",
     var_name="mrt",
     compute=indices.mean_radiant_temperature,
+)
+
+
+shortwave_radiation_upwards_from_net_downwards = Converter(
+    title="Upwards shortwave radiation",
+    identifier="shortwave_radiation_upwards_from_net_downward",
+    units="W m-2",
+    standard_name="surface_net_downward_shortwave_fluxt",
+    long_name="Calculation of upwards shortwave radiative flux from net surface shortwave "
+    "and downward surface shortwave fluxes.",
+    description="The calculation of upwards shortwave radiative flux from net surface shortwave "
+    "and downward surface shortwave fluxes.",
+    var_name="rsus",
+    compute=indices.shortwave_radiation_upwards_from_net_downwards,
+)
+
+longwave_radiation_upwards_from_net_downwards = Converter(
+    title="Upwards longwave radiation",
+    identifier="longwave_radiation_upwards_from_net_downwards",
+    units="W m-2",
+    standard_name="surface_net_downward_longwave_flux",
+    long_name="Calculation of upwards longwave radiative flux from net surface longwave "
+    "and downward surface longwave fluxes.",
+    description="The calculation of upwards longwave radiative flux from net surface longwave "
+    "and downward surface longwave fluxes.",
+    var_name="rlus",
+    compute=indices.longwave_radiation_upwards_from_net_downwards,
 )
