### 2023-02-02 c0d09c0
Only if 'changed' is not None, expand attributes
