+++ xclim/testing/tests/test_ensembles.py
@@ -563,63 +563,145 @@ def robust_data(request):
 
 
 @pytest.mark.parametrize(
-    "test,exp_chng,exp_sign,kws",
+    "test,exp_chng_frac,exp_pos_frac,exp_chng,kws",
     [
-        ("ttest", [0.25, 1, 1, 1], [1, 0.5, 1, 1], {}),
-        ("welch-ttest", [0.25, 1, 1, 1], [1, 0.5, 1, 1], {}),
-        ("mannwhitney-utest", [0.25, 1, 1, 1], [1, 0.5, 1, 1], {}),
-        ("threshold", [0.25, 1, 1, 1], [1, 0.5, 1, 1], {"rel_thresh": 0.002}),
+        (
+            "ttest",
+            [0.25, 1, 1, 1],
+            [1, 0.5, 1, 1],
+            [
+                [False, False, False, True],
+                [True, True, True, True],
+                [True, True, True, True],
+                [False, False, True, True],
+            ],
+            {},
+        ),
+        (
+            "welch-ttest",
+            [0.25, 1, 1, 1],
+            [1, 0.5, 1, 1],
+            [
+                [False, False, False, True],
+                [True, True, True, True],
+                [True, True, True, True],
+                [False, False, True, True],
+            ],
+            {},
+        ),
+        (
+            "mannwhitney-utest",
+            [0.25, 1, 1, 1],
+            [1, 0.5, 1, 1],
+            [
+                [False, False, False, True],
+                [True, True, True, True],
+                [True, True, True, True],
+                [False, False, True, True],
+            ],
+            {},
+        ),
+        (
+            "threshold",
+            [0.25, 1, 1, 1],
+            [1, 0.5, 1, 1],
+            [
+                [False, False, False, True],
+                [True, True, True, True],
+                [True, True, True, True],
+                [False, False, True, True],
+            ],
+            {"rel_thresh": 0.002},
+        ),
         (
             "threshold",
             [0, 0, 0.5, 0],
             [np.nan, np.nan, 1, np.nan],
+            [
+                [False, False, False, False],
+                [False, False, False, False],
+                [False, False, True, True],
+                [False, False, False, False],
+            ],
             {"abs_thresh": 2},
         ),
-        (None, [1, 1, 1, 1], [0.5, 0.5, 1, 1], {}),
+        (
+            None,
+            [1, 1, 1, 1],
+            [0.5, 0.5, 1, 1],
+            [],
+            {},
+        ),
     ],
 )
-def test_change_significance(robust_data, test, exp_chng, exp_sign, kws):
+def test_change_significance(robust_data, test, exp_chng_frac, exp_pos_frac, exp_chng, kws):
     ref, fut = robust_data
-    chng, sign = ensembles.change_significance(fut, ref, test=test, **kws)
-    assert chng.attrs["test"] == str(test)
+    chng_frac, pos_frac, chng = ensembles.change_significance(fut, ref, test=test, **kws)
+    assert chng_frac.attrs["test"] == str(test)
     if isinstance(ref, xr.Dataset):
-        chng = chng.tas
-        sign = sign.tas
+        chng_frac = chng_frac.tas
+        pos_frac = pos_frac.tas
+        if chng is not None:
+            chng = chng.tas
+    np.testing.assert_array_almost_equal(chng_frac, exp_chng_frac)
+    np.testing.assert_array_almost_equal(pos_frac, exp_pos_frac)
     np.testing.assert_array_almost_equal(chng, exp_chng)
[…]
