### 2023-02-02 #73 @Batalex
:wave: Hi @jmunroe and @aguspesce! Thank you for volunteering to review Xclim for pyOpenSci!

Meet @Zeitsperre, our lovely submission author, and feel free to introduce yourselves 🥰

1. Here is the **[reviewers guide](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html/)**. This guide contains all the steps and information needed to complete your review.
2. You will need the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** to fill out and submit in this issue as a comment once your review is complete.

Please get in touch with any questions or concerns! I just wanted to let you know that your review is due: Feb 26th.

Also, I'd like to point out that the next PR should be part of your review: https://github.com/Ouranosinc/xclim/pull/250.

PS: I'm sorry for the information dump. Please make sure that you review the [submitted version](https://github.com/Ouranosinc/xclim/releases/tag/v0.40.0). However, some issues you could point out might have already been solved in the meantime, but that's ok.
