### 2023-02-02 b3d5db2
return pvals. remove realization argument
