+++ xclim/ensembles/_robustness.py
@@ -12,6 +12,7 @@ from __future__ import annotations
 from typing import Union
 
 import numpy as np
+import scipy
 import scipy.stats as spstats  # noqa
 import xarray as xr
 
@@ -23,7 +24,7 @@ def change_significance(
     ref: xr.DataArray | xr.Dataset = None,
     test: str = "ttest",
     weights: xr.DataArray = None,
-    realization="realization",
+    p_vals: bool = False,
     **kwargs,
 ) -> tuple[xr.DataArray | xr.Dataset, xr.DataArray | xr.Dataset]:
     """Robustness statistics qualifying how the members of an ensemble agree on the existence of change and on its sign.
@@ -46,12 +47,12 @@ def change_significance(
     weights : xr.DataArray
         Weights to apply along the 'realization' dimension. This array cannot contain missing values.
         Note: 'ttest' and 'welch-ttest' are not currently supported with weighted arrays.
-    realization: str
-        Name of the realization dimension. If you don't have multiple climate realizations or models, choose None.
+    p_vals : bool
+        If True, return the estimated p-values.
     **kwargs
         Other arguments specific to the statistical test.
 
-        For 'ttest' and 'welch-ttest':
+        For 'ttest', 'welch-ttest' and 'mannwhitney-utest':
             p_change : float (default : 0.05)
                 p-value threshold for rejecting the hypothesis of no significant change.
         For 'threshold': (Only one of those must be given.)
@@ -69,6 +70,9 @@ def change_significance(
     pos_frac : xr.DataArray or xr.Dataset
         The fraction of members showing significant change that show a positive change ]0, 1].
         Null values are returned where no members show significant change.
+    pvals : xr.DataArray or xr.Dataset or None
+        The p-values estimated by the significance tests. Only returned if `p_vals` is True and `test` is one of
+        'ttest', 'welch-ttest', 'mannwhitney-utest'.
 
         The table below shows the coefficient needed to retrieve the number of members
         that have the indicated characteristics, by multiplying it to the total
@@ -126,20 +130,26 @@ def change_significance(
     ...     delta, test="threshold", abs_thresh=2
     ... )
     """
-    if realization is None:
-        realization = "realization"
-        fut = fut.assign_coords({"realization": "dummy"})
-        fut = fut.expand_dims("realization")
-        if ref is not None:
-            ref = ref.assign_coords({"realization": "dummy"})
-            ref = ref.expand_dims("realization")
+    # Realization dimension name
+    realization = "realization"
 
+    # Assign dummy realization dimension if not present.
+    if realization not in fut.dims:
+        fut = fut.assign_coords({realization: "dummy"})
+        fut = fut.expand_dims(realization)
+    if ref is not None and realization not in ref.dims:
+        ref = ref.assign_coords({realization: "dummy"})
+        ref = ref.expand_dims(realization)
+
+    # Significance tests parameter names
     test_params = {
         "ttest": ["p_change"],
         "welch-ttest": ["p_change"],
         "mannwhitney-utest": ["p_change"],
         "threshold": ["abs_thresh", "rel_thresh"],
     }
+
+    # Get delta, either from fut or from fut - ref
     changed = None
     if ref is None:
         delta = fut
@@ -158,6 +168,7 @@ def change_significance(
         else:
             n_valid_real = weights.where(fut.notnull().all("time")).sum(realization)
 
+    pvals = None
     if test == "ttest":
         if weights is not None:
             raise NotImplementedError(
@@ -178,6 +189,7 @@ def change_significance(
         )
         # When p < p_change, the hypothesis of no significant change is rejected.
         changed = pvals < p_change
+
     elif test == "welch-ttest":
         if weights is not None:
[…]
