+++ HISTORY.rst
@@ -13,19 +13,21 @@ Bug fixes
 New indicators
 ^^^^^^^^^^^^^^
 * New indices and indicators for converting from snow water equivalent to snow depth (``snow_depth_approximation``) and snow depth to snow water equivalent (``snow_amount_approximation``) using snow density [kg/m^3]. (:pull:`1271`).
+* New indices and indicators for determining upwards radiation (CF variables `rsus` and `rlus`) from net and downwards longwave radiation (shortwave: `rss` and `rsds`; longwave: `rls` and `rlds`). (:pull:`1271`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^
 * `xclim` has adopted `PEP 517 <https://peps.python.org/pep-0517/>`_ and `PEP 621 <https://peps.python.org/pep-0621/>`_ (``pyproject.toml`` using the `flit <https://flit.pypa.io/en/stable/>`_ backend) to replace the legacy ``setup.py`` used to manage package organisation and building. Many tooling configurations that already supported the ``pyproject.toml`` standard have been migrated to this file. CI and development tooling documentation has been updated to reflect these changes. (:pull:`1278`, suggested from `PyOpenSci Software Review <https://github.com/pyOpenSci/software-review/issues/73>`_).
 * Documentation source files have been moved around to remove some duplicated image files. (:pull:`1278`).
 * Coveralls GitHub Action removed as it did not support ``pyproject.toml``-based configurations. (:pull:`1278`).
-* Add a remark about how xclim's CFFWIS is different from the original 1982 implementation. (:issue:`1104`, :pull:`1284`).
+* Add a remark about how `xclim`'s CFFWIS is different from the original 1982 implementation. (:issue:`1104`, :pull:`1284`).
 * Update CI runs to use Python3.9 when examining upstream dependencies. Replace `setup-conda` action with `provision-with-micromamba` action. (:pull:`1286`).
 * Update CI runs to always use `tox~=3.0` and the `latest` Ubuntu images (now `v22.04`). (:pull:`1288`).
 * `SBCK` installation command now points to the official development repository. (:pull:`1288`).
 * Some references in the BibTeX were updated to point to better resources. (:pull:`1288`).
 * Add a GitHub CI workflow for performing dependency security review scanning. (:pull:`1287`).
-* Grammar and spelling corrections were applied to some dostrings. (:pull:`1271`).
+* Grammar and spelling corrections were applied to some docstrings. (:pull:`1271`).
+* Added `[radiation]` (`[power] / [area]`) to list of defined acceptable units. (:pull:`1271`).
 
 0.40.0 (2023-01-13)
 -------------------
