### 2023-02-02 c84e5e2
Apply suggestions from code review
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
