+++ xclim/indicators/atmos/_conversion.py
@@ -390,8 +390,8 @@ shortwave_radiation_upwards_from_net_downwards = Converter(
     title="Upwards shortwave radiation",
     identifier="shortwave_radiation_upwards_from_net_downward",
     units="W m-2",
-    standard_name="surface_net_upward_shortwave_flux",
-    long_name="Upwards shortwave radiative flux",
+    standard_name="surface_upwelling_shortwave_flux",
+    long_name="Upwelling shortwave flux",
     description="The calculation of upwards shortwave radiative flux from net surface shortwave "
     "and downward surface shortwave fluxes.",
     var_name="rsus",
@@ -402,8 +402,8 @@ longwave_radiation_upwards_from_net_downwards = Converter(
     title="Upwards longwave radiation",
     identifier="longwave_radiation_upwards_from_net_downwards",
     units="W m-2",
-    standard_name="surface_net_upward_longwave_flux",
-    long_name="Upwards longwave radiative flux",
+    standard_name="surface_upwelling_longwave_flux",
+    long_name="Upwelling longwave flux",
     description="The calculation of upwards longwave radiative flux from net surface longwave "
     "and downward surface longwave fluxes.",
     var_name="rlus",
