### 2023-02-03 37f72ed
scipy version request with parse_version
