+++ xclim/ensembles/_robustness.py
@@ -88,8 +88,6 @@ def change_significance(
         +-----------------+------------------------------+                        |
         | Negative change | (1 - pos_frac) * change_frac |                        |
         +-----------------+------------------------------+------------------------+
-    changed :  xr.DataArray or xr.Dataset
-        Significant changes for each `realization` member.
 
     Notes
     -----
