### 2023-02-03 32e0ea9
Add missing argument of `use_ufunc` in `windowed_run_count`
