### 2023-02-03 c99a632
move references to bibtex. Add module and function to api docs
