### 2023-02-03 abf8532
rename indices and indicators to use cf-like naming conventions
