+++ xclim/indicators/atmos/_conversion.py
@@ -28,8 +28,8 @@ __all__ = [
     "corn_heat_units",
     "universal_thermal_climate_index",
     "mean_radiant_temperature",
-    "shortwave_radiation_upwards_from_net_downwards",
-    "longwave_radiation_upwards_from_net_downwards",
+    "shortwave_upwelling_radiation_from_net_downwelling",
+    "longwave_upwelling_radiation_from_net_downwelling",
 ]
 
 
@@ -386,26 +386,26 @@ mean_radiant_temperature = Converter(
 )
 
 
-shortwave_radiation_upwards_from_net_downwards = Converter(
-    title="Upwards shortwave radiation",
-    identifier="shortwave_radiation_upwards_from_net_downward",
+shortwave_upwelling_radiation_from_net_downwelling = Converter(
+    title="Upwelling shortwave radiation",
+    identifier="shortwave_upwelling_radiation_from_net_downwelling",
     units="W m-2",
     standard_name="surface_upwelling_shortwave_flux",
     long_name="Upwelling shortwave flux",
-    description="The calculation of upwards shortwave radiative flux from net surface shortwave "
-    "and downward surface shortwave fluxes.",
+    description="The calculation of upwelling shortwave radiative flux from net surface shortwave "
+    "and downwelling surface shortwave fluxes.",
     var_name="rsus",
-    compute=indices.shortwave_radiation_upwards_from_net_downwards,
+    compute=indices.shortwave_upwelling_radiation_from_net_downwelling,
 )
 
-longwave_radiation_upwards_from_net_downwards = Converter(
-    title="Upwards longwave radiation",
-    identifier="longwave_radiation_upwards_from_net_downwards",
+longwave_upwelling_radiation_from_net_downwelling = Converter(
+    title="Upwelling longwave radiation",
+    identifier="longwave_upwelling_radiation_from_net_downwelling",
     units="W m-2",
     standard_name="surface_upwelling_longwave_flux",
     long_name="Upwelling longwave flux",
-    description="The calculation of upwards longwave radiative flux from net surface longwave "
-    "and downward surface longwave fluxes.",
+    description="The calculation of upwelling longwave radiative flux from net surface longwave "
+    "and downwelling surface longwave fluxes.",
     var_name="rlus",
-    compute=indices.longwave_radiation_upwards_from_net_downwards,
+    compute=indices.longwave_upwelling_radiation_from_net_downwelling,
 )
