### 2023-02-03 dcb3e0e
Bump version: 0.40.4-beta → 0.40.5-beta
