+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.40.4-beta
+current_version = 0.40.5-beta
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+).(?P<patch>\d+)(\-(?P<release>[a-z]+))?
