### 2023-02-03 05940ee
add rsus to atmosds generator
