### 2023-02-04 016af77
Bump version: 0.40.5-beta → 0.40.6-beta
