+++ tox.ini
@@ -85,6 +85,7 @@ passenv =
     COVERALLS_*
     GITHUB_*
     LD_LIBRARY_PATH
+    MAIN_TESTDATA_BRANCH
 extras = dev
 deps =
     coverage: coveralls
