### 2023-02-06 9e7523b
fix previously uncaught error in doctest
