### 2023-02-06 d1bbe69
use hydro context for precip rate to length conversion
