+++ tox.ini
@@ -60,7 +60,7 @@ description = Run documentation linters and doctests with pytest under {basepyth
 commands =
 ;    - mypy xclim
 ;    pylint --rcfile=pylintrc --exit-zero xclim  # pylint and tox do not play well together at the moment
-    pytest --no-cov --nbval --dist=loadscope docs/notebooks --durations=10 --ignore=docs/notebooks/example.ipynb
+;    pytest --no-cov --nbval --dist=loadscope docs/notebooks --durations=10 --ignore=docs/notebooks/example.ipynb
     pytest --rootdir xclim/testing/tests/ --xdoctest xclim --ignore=xclim/testing/tests/ --durations=10
 
 # Requires tox-conda compatible with tox@v4.0
