+++ tox.ini
@@ -2,7 +2,7 @@
 env_list =
     black
     docs
-    doctests_only
+    notebooks_doctests
 ;    opt-slow
     py38-lm3
     py39-upstream-doctest
@@ -55,12 +55,10 @@ allowlist_externals =
 ;conda_env = environment-dev.yml
 ;extras =
 
-[testenv:doctests_only]
-description = Run documentation linters and doctests with pytest under {basepython}
+[testenv:notebooks_doctests]
+description = Run notebooks and doctests with pytest under {basepython}
 commands =
-;    - mypy xclim
-;    pylint --rcfile=pylintrc --exit-zero xclim  # pylint and tox do not play well together at the moment
-;    pytest --no-cov --nbval --dist=loadscope docs/notebooks --durations=10 --ignore=docs/notebooks/example.ipynb
+    pytest --no-cov --nbval --dist=loadscope docs/notebooks --durations=10 --ignore=docs/notebooks/example.ipynb
     pytest --rootdir xclim/testing/tests/ --xdoctest xclim --ignore=xclim/testing/tests/ --durations=10
 
 # Requires tox-conda compatible with tox@v4.0
