### 2023-02-06 88c7527
Bump version: 0.40.6-beta → 0.40.7-beta
