### 2023-02-07 #73 @NickleDave
I might be reading their docs wrong, but I understood it to say that the paper files can be on some other branch that never gets merged with main.
