### 2023-02-08 69d1861
Update xclim/ensembles/_robustness.py
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
