### 2023-02-08 e1d2ed3
get rid of 'realization' in 'apply_ufunc' calls
