+++ xclim/ensembles/_robustness.py
@@ -185,8 +185,8 @@ def change_significance(
             lambda f, r: spstats.ttest_1samp(f, r, axis=-1, nan_policy="omit")[1],
             fut,
             ref.mean("time"),
-            input_core_dims=[[realization, "time"], [realization]],
-            output_core_dims=[[realization]],
+            input_core_dims=[["time"], []],
+            output_core_dims=[[]],
             vectorize=True,
             dask="parallelized",
             output_dtypes=[float],
@@ -209,8 +209,8 @@ def change_significance(
             )[1],
             fut,
             ref,
-            input_core_dims=[[realization, "time"], [realization, "time"]],
-            output_core_dims=[[realization]],
+            input_core_dims=[["time"], ["time"]],
+            output_core_dims=[[]],
             exclude_dims={"time"},
             vectorize=True,
             dask="parallelized",
@@ -236,8 +236,8 @@ def change_significance(
             lambda f, r: spstats.mannwhitneyu(f, r, axis=-1, nan_policy="omit")[1],
             fut,
             ref,
-            input_core_dims=[[realization, "time"], [realization, "time"]],
-            output_core_dims=[[realization]],
+            input_core_dims=[["time"], ["time"]],
+            output_core_dims=[[]],
             exclude_dims={"time"},
             vectorize=True,
             dask="parallelized",
