### 2023-02-08 bb7c72c
Apply suggestions from code review
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
