+++ xclim/indicators/atmos/_precip.py
@@ -188,7 +188,7 @@ dryness_index = Precip(
     "the precipitation and evapotranspiration factors without deduction for surface runoff or drainage. "
     "Metric originally published in Riou et al. (1994). ",
     cell_methods="",
-    src_freq=["D", "MS"],
+    src_freq=["D", "M"],
     compute=indices.dryness_index,
 )
 
