### 2023-02-08 a63ebaa
Update xclim/indicators/atmos/_precip.py
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
