### 2023-02-08 6e52bdc
explicitly set inputkind
