### 2023-02-08 e6a9234
add tests (WIP)
