### 2023-02-08 5cf14e4
adjust units conversion and add working tests
