+++ xclim/testing/tests/test_indices.py
@@ -3080,18 +3080,15 @@ def test_mean_radiant_temperature(
 
 class TestDrynessIndex:
     def test_dryness_index(self, atmosds):
-        ds = atmosds
+        ds = atmosds.isel(location=3)
 
         evspsblpot = ds.evspsblpot
         pr = ds.pr
 
         di = xci.dryness_index(pr, evspsblpot)
         di_wet = xci.dryness_index(pr, evspsblpot, wo="300 mm")
-
-        # FIXME: This test needs to be adjusted once the converter is working.
+        di_plus_100 = di + 100
         np.testing.assert_allclose(
-            di.isel(location=3)[:10], np.array([1, 2, 3, 4, 5, 6, 7, 8, 9])
+            di, np.array([13.355, 102.426, 65.576, 158.078]), rtol=1e-03
         )
-        np.testing.assert_allclose(
-            di_wet.isel(location=3)[:10], np.array([1, 2, 3, 4, 5, 6, 7, 8, 9])
-        )
+        np.testing.assert_allclose(di_wet, di_plus_100)
