### 2023-02-08 e396b0a
Update xclim/indices/_agro.py
Co-authored-by: Éric Dupuis <71575674+coxipi@users.noreply.github.com>
