### 2023-02-08 4b2aabc
set wo in declared units, calculate rlus and rsus on-the-fly to prevent thread issues
