+++ xclim/testing/tests/test_indicators.py
@@ -464,7 +464,6 @@ def test_all_parameters_understood(official_indicators):
     if problems - {
         ("COOL_NIGHT_INDEX", "lat"),
         ("DRYNESS_INDEX", "lat"),
-        ("DRYNESS_INDEX", "wo"),
     }:
         raise ValueError(
             f"The following indicator/parameter couple {problems} use types not listed in InputKind."
