### 2023-02-08 0cc2bb9
add a "thread_unsafe" marker
