### 2023-02-09 0dbf365
Add Dryness Index (#1235) and Simplify atmosds dataset (#1269)
### What kind of change does this PR introduce?

* Adds the agroclimatic metric Dryness Index from Riou et al. 1994
(https://hal.inrae.fr/hal-02843898)
* Adds the changes to the atmosds dataset creation #1269 to the main
development branch
* Updates `pytest-xdist` to use the `--dist=worksteal` scheduler

### Does this PR introduce a breaking change?

**Yes.** Testing datasets have been modified, so changes to expected
results apply for all tests based on the ERA5 Canadian cities' dataset.

### Other information:

This PR lead to th
[…]
