+++ xclim/testing/tests/test_atmos.py
@@ -2,6 +2,8 @@
 # Expected values might be the same as for the indices tests, see test_indices comments.
 from __future__ import annotations
 
+import sys
+
 import numpy as np
 import pytest
 import xarray as xr
@@ -232,21 +234,53 @@ def test_wind_chill_index(atmosds):
     out = atmos.wind_chill_index(ds=atmosds)
 
     np.testing.assert_allclose(
-        out.isel(time=0), [np.nan, -6.116, -36.064, -7.153, np.nan], rtol=1e-3
+        out.isel(time=0), [np.nan, -6.716, -35.617, -8.486, np.nan], rtol=1e-3
     )
 
     out_us = atmos.wind_chill_index(ds=atmosds, method="US")
 
     np.testing.assert_allclose(
-        out_us.isel(time=0), [-1.041, -6.116, -36.064, -7.153, 2.951], rtol=1e-3
+        out_us.isel(time=0), [-1.429, -6.716, -35.617, -8.486, 2.781], rtol=1e-3
     )
 
 
-class TestPotentialEvapotranspiration:
-    def test_convert_units(self, open_dataset):
-        ds = open_dataset(
-            "ERA5/daily_surface_cancities_1990-1993.nc",
+class TestDrynessIndex:
+    def test_simple(self, atmosds):
+        ds = atmosds.isel(location=3)
+
+        pr = ds.pr
+        evspsblpot = ds.evspsblpot
+
+        di = atmos.dryness_index(pr, evspsblpot)
+        np.testing.assert_allclose(
+            di, np.array([13.355, 102.426, 65.576, 158.078]), rtol=1e-03
         )
+        assert di.attrs["long_name"] == "Growing season humidity"
+
+    def test_variable_initial_conditions(self, atmosds):
+        ds = atmosds
+
+        pr = ds.pr
+        evspsblpot = ds.evspsblpot
+
+        di = atmos.dryness_index(pr, evspsblpot)
+        di_wet = atmos.dryness_index(pr, evspsblpot, wo="250 mm")
+        di_dry = atmos.dryness_index(pr, evspsblpot, wo="100 mm")
+
+        assert np.all(di_wet > di_dry)
+        di_plus_50 = di + 50
+        np.testing.assert_allclose(di_wet, di_plus_50, rtol=1e-03)
+        di_minus_100 = di - 100
+        np.testing.assert_allclose(di_dry, di_minus_100, rtol=1e-03)
+
+        for value, array in {"200 mm": di, "250 mm": di_wet, "100 mm": di_dry}.items():
+            assert array.attrs["long_name"] == "Growing season humidity"
+            assert value in array.attrs["description"]
+
+
+class TestPotentialEvapotranspiration:
+    def test_convert_units(self, atmosds):
+        ds = atmosds
 
         tn = ds.tasmin
         tx = ds.tasmax
@@ -304,10 +338,8 @@ class TestPotentialEvapotranspiration:
         np.testing.assert_allclose(pet_mb05, pet_mb05C, atol=1)
         np.testing.assert_allclose(pet_fao_pm98, pet_fao_pm98C, atol=1)
 
-    def test_nan_values(self, open_dataset):
-        ds = open_dataset(
-            "ERA5/daily_surface_cancities_1990-1993.nc",
-        )
+    def test_nan_values(self, atmosds):
+        ds = atmosds
 
         tn = ds.tasmin
         tx = ds.tasmax
@@ -357,10 +389,8 @@ class TestPotentialEvapotranspiration:
 
 
 class TestWaterBudget:
-    def test_convert_units(self, open_dataset):
-        ds = open_dataset(
[…]
