+++ xclim/testing/tests/test_indices.py
@@ -2184,7 +2184,7 @@ class TestWinterRainRatio:
 class TestTG:
     @pytest.mark.parametrize(
         "ind,exp",
-        [(xci.tg_mean, 283.1391), (xci.tg_min, 266.1117), (xci.tg_max, 292.1250)],
+        [(xci.tg_mean, 283.0615), (xci.tg_min, 266.1208), (xci.tg_max, 291.5018)],
     )
     def test_simple(self, open_dataset, ind, exp):
         ds = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc")
@@ -3076,3 +3076,19 @@ def test_mean_radiant_temperature(
     )
 
     np.testing.assert_allclose(mrt, expected, rtol=1e-03)
+
+
+class TestDrynessIndex:
+    def test_dryness_index(self, atmosds):
+        ds = atmosds.isel(location=3)
+
+        evspsblpot = ds.evspsblpot
+        pr = ds.pr
+
+        di = xci.dryness_index(pr, evspsblpot)
+        di_wet = xci.dryness_index(pr, evspsblpot, wo="300 mm")
+        di_plus_100 = di + 100
+        np.testing.assert_allclose(
+            di, np.array([13.355, 102.426, 65.576, 158.078]), rtol=1e-03
+        )
+        np.testing.assert_allclose(di_wet, di_plus_100)
