### 2023-02-09 5e11d5b
Bump version: 0.40.7-beta → 0.40.8-beta
