+++ xclim/testing/tests/test_ensembles.py
@@ -604,6 +604,18 @@ def robust_data(request):
             {},
         ),
         (
+            "brownforsythe-test",
+            [0.5, 0.25, 0.5, 0.5],
+            [0.5, 0., 1, 1],
+            [
+                [False, True, False, True],
+                [True, False, False, False],
+                [False, True, False, True],
+                [False, False, False, True],
+            ],
+            {},
+        ),
+        (
             "threshold",
             [0.25, 1, 1, 1],
             [1, 0.5, 1, 1],
