### 2023-02-10 1cad62e
Added missing conversion factor snd -> snw in tests
