### 2023-02-10 5a98c03
snow_water -> snow_amount
