### 2023-02-10 ed1e706
WIP - better cosza missing two cases
