### 2023-02-14 dc6c75e
add svgconverter for badges in documentation LaTeX
