### 2023-02-14 0f6eee3
implement `XCLIM_TESTDATA_BRANCH` and `XCLIM_PREFETCH_TESTING_DATA` environment variables, remove obsolete subset-specific testing data, adjust testing data fetching logic
