+++ pyproject.toml
@@ -70,6 +70,7 @@ dev = [
   "pytest-xdist[psutil] >=3.2",
   "tokenize-rt",
   "tox",
+  # "tox-conda",  # Will be added when a tox@v4.0+ compatible plugin is released.
   "xdoctest",
   "yamale",
   "yamllint",
