+++ .github/workflows/main.yml
@@ -40,7 +40,7 @@ jobs:
         with:
           python-version: "3.8"
       - name: Install tox
-        run: pip install tox~=3.0
+        run: pip install tox~=4.0
       - name: Run linting suite
         run: tox -e black
 
@@ -56,7 +56,7 @@ jobs:
         with:
           python-version: "3.8"
       - name: Install tox
-        run: pip install tox~=3.0
+        run: pip install tox~=4.0
       - name: Test with tox
         run: tox -e py38
         env:
@@ -76,7 +76,7 @@ jobs:
         with:
           python-version: "3.9"
       - name: Install tox
-        run: pip install tox~=3.0
+        run: pip install tox~=4.0
       - name: Test with tox
         run: tox -e py39-upstream-lm3-coverage
         env:
@@ -156,7 +156,7 @@ jobs:
         with:
           python-version: "3.8"
       - name: Install tox
-        run: pip install tox~=3.0
+        run: pip install tox~=4.0
       - name: Test with tox
         run: tox -e py38-lm3-coverage
         env:
