### 2023-02-14 b69b867
upgrade to tox>=4.0, adjust tox configuration to set pytest-xdist and coverage options as needed, adjust pyproject.toml to not use pytest-xdist/report coverage by default
