+++ .github/workflows/main.yml
@@ -20,7 +20,7 @@ on:
       - submitted
 
 env:
-  MAIN_TESTDATA_BRANCH: main
+  XCLIM_TESTDATA_BRANCH: main
 
 jobs:
   black:
@@ -115,8 +115,9 @@ jobs:
           python -m pip install --no-user .
       - name: Test
         run: |
+          pip list
           pip check
-          pytest xclim --durations=10
+          pytest xclim --num-workers=logical --durations=10 --cov=xclim --cov-report=term-missing
         env:
           CONDA_EXE: mamba
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -184,10 +185,6 @@ jobs:
         run: pylint --rcfile=pylintrc --disable=import-error --exit-zero xclim
       - name: Test with tox
         run: tox -e notebooks_doctests
-      - name: Report coverage
-        run: |
-          pip install --upgrade coveralls
-          coveralls
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           COVERALLS_FLAG_NAME: run-notebooks_doctests-py38
