### 2023-02-20 d62a6b0
default constant snow density value
