### 2023-02-20 7227de2
Sturm et, al 2010
