+++ docs/references.bib
@@ -1880,3 +1880,17 @@ Publisher: Ubiquity Press},
 	author = {Vrac, M.},
 	year = {2018},
 }
+
+@article {sturm_swe_2010,
+      author = "Matthew  Sturm and Brian  Taras and Glen E.  Liston and Chris  Derksen and Tobias  Jonas and Jon  Lea",
+      title = "Estimating Snow Water Equivalent Using Snow Depth Data and Climate Classes",
+      journal = "Journal of Hydrometeorology",
+      year = "2010",
+      publisher = "American Meteorological Society",
+      address = "Boston MA, USA",
+      volume = "11",
+      number = "6",
+      doi = "10.1175/2010JHM1202.1",
+      pages=      "1380 - 1394",
+      url = "https://journals.ametsoc.org/view/journals/hydr/11/6/2010jhm1202_1.xml"
+}
