+++ xclim/indices/_conversion.py
@@ -906,7 +906,7 @@ def rain_approximation(
 @declare_units(snw="[mass]/[area]", snr="[mass]/[volume]", const="[mass]/[volume]")
 def snow_depth_from_amount(
     snw: xr.DataArray,
-    snr: xr.DataArray = None,
+    snr: xr.DataArray | None = None,
     const: Quantified = "312 kg m-3",
 ) -> xr.DataArray:
     """Snow depth from snow amount and density.
@@ -918,7 +918,7 @@ def snow_depth_from_amount(
         If snow water equivalent (`swe` [m]) is provided instead, will be converted to `snw` before calculating.
     snr : xr.DataArray, optional
         Snow density [kg/m^3].
-    const: Quantified, optional
+    const: Quantified
         Constant snow density [kg/m^3]
         `const` is only used if `snr` is None.
 
@@ -929,7 +929,7 @@ def snow_depth_from_amount(
 
     Notes
     -----
-    The estimated mean snow density value of 312 kg m-3 is taken from :cite:p:`sturm_swe_2010`.
+    The estimated mean snow density value of 312 kg m-3 is taken from :cite:t:`sturm_swe_2010`.
 
     References
     ----------
@@ -950,7 +950,7 @@ def snow_depth_from_amount(
 @declare_units(snd="[length]", snr="[mass]/[volume]", const="[mass]/[volume]")
 def snow_amount_from_depth(
     snd: xr.DataArray,
-    snr: xr.DataArray = None,
+    snr: xr.DataArray | None = None,
     const: Quantified = "312 kg m-3",
 ) -> xr.DataArray:
     """Snow amount from snow depth and density.
@@ -961,7 +961,7 @@ def snow_amount_from_depth(
         Snow depth [m].
     snr : xr.DataArray, optional
         Snow density [kg/m^3].
-    const: Quantified, optional
+    const: Quantified
         Constant snow density [kg/m^3]
         `const` is only used if `snr` is None.
 
@@ -972,7 +972,7 @@ def snow_amount_from_depth(
 
     Notes
     -----
-    The estimated mean snow density value of 312 kg m-3 is taken from :cite:p:`sturm_swe_2010`.
+    The estimated mean snow density value of 312 kg m-3 is taken from :cite:t:`sturm_swe_2010`.
 
     References
     ----------
