+++ docs/references.bib
@@ -1891,6 +1891,6 @@ Publisher: Ubiquity Press},
       volume = "11",
       number = "6",
       doi = "10.1175/2010JHM1202.1",
-      pages=      "1380 - 1394",
+      pages = "1380 - 1394",
       url = "https://journals.ametsoc.org/view/journals/hydr/11/6/2010jhm1202_1.xml"
 }
