### 2023-02-22 b8023e7
Fix: pre-commit conda-forge package
`pre_commit` (underscored) was used but it is oudated.
See: https://github.com/search?q=org%3Aconda-forge+pre_commit
