+++ environment.yml
@@ -43,7 +43,7 @@ dependencies:
     - nc-time-axis
     - netCDF4>=1.4
     - pooch
-    - pre_commit
+    - pre-commit
     - pybtex
     - pydocstyle
     - pylint<2.15
