### 2023-02-24 0610641
Bump version: 0.40.8-beta → 0.40.9-beta
