### 2023-02-27 Ouranosinc/xclim#1303 @Zeitsperre
[PyOpenSci] `test_mean_radiant_temperature` sometimes fails with segfault
### Setup Information

- Xclim version: 0.40.0
- Python version: Any
- Operating System: All

### Description

The `test_mean_radiant_temperature` test seems to sometimes fail and sometimes pass. This might be due to some backend issues (asynchronous read/write? Caching?) but all the same, the intermittent failures need to be resolved.

This issue was raised in the pyOpenSci review process (https://github.com/pyOpenSci/software-submission/issues/73#issuecomment-1445193038)

### Steps To Reproduce

```
$ tox
```

### Additional context

```
 ______________________ xclim/testing/tests/test_atmos.py _______________________
[gw0] linux -- Python 3.10.9 /home/runner/micromamba-root/envs/xclim/bin/python3.10
worker 'gw0' crashed while running 'xclim/testing/tests/test_atmos.py::TestMeanRadiantTemperature::test_mean_radiant_temperature'
```

### Contribution

- [ ] I would be willing/able to open a Pull Request to address this bug.

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
