### 2023-02-27 Ouranosinc/xclim#1304 @Zeitsperre
[PyOpenSci] More descriptive repo badges
### Generic Issue

* xclim version: 0.40.0
* Python version: Any
* Operating System: All

### Description

We are missing a few useful repository badges in the README:
 - A [repostatus.org](https://www.repostatus.org/) badge,
 - Python versions supported,
 - Current package version (on PyPI / Conda).

This issue is linked to comments from reviewers in PyOpenSci: https://github.com/pyOpenSci/software-submission/issues/73#issuecomment-1445193038

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
