+++ xclim/indicators/land/_snow.py
@@ -62,7 +62,7 @@ snw_season_length = SnowWithIndexing(
     title="Snow cover duration (amount)",
     identifier="snw_season_length",
     units="days",
-    long_name="Number of days with snow amount at or above threshold",
+    long_name="Snow cover duration",
     description="The {freq} number of days with snow amount greater than or equal to {thresh}.",
     abstract="Number of days when the snow amount is greater than or equal to a given threshold.",
     compute=xci.snw_season_length,
