+++ HISTORY.rst
@@ -48,6 +48,7 @@ Internal changes
 * Updated testing data used to generate the `atmosds` dataset to use more reproducibly-converted ERA5 data, generated with the `miranda` Python package. (:pull:`1269`).
 * Updated testing dependencies to use `pytest-xdist>=3.2`, allowing for the new `--dist=worksteal` scheduler for distributing the pool of remaining tests across workers after individual workers have exhausted their own queues. (:pull:`1235`).
 * Adding infer context to the unit conversion in of the training of ExtremeValues. (:pull:`1299`).
+* README badges for supported Python versions and repository health have been added. (:issue:`1304`, :pull:`1307`).
 
 0.40.0 (2023-01-13)
 -------------------
