+++ docs/contributing.rst
@@ -1,1 +1,1 @@
-.. include:: ../.github/CONTRIBUTING.rst
+.. include:: ../CONTRIBUTING.rst
