### 2023-02-27 e130bf8
add license information to README.rst
