+++ README.rst
@@ -52,13 +52,21 @@ xclim is in active development and is being used in production by climate servic
 * If you would like to contribute code or documentation (which is greatly appreciated!), check out the `Contributing Guidelines`_ before you begin!
 
 .. _issue tracker: https://github.com/Ouranosinc/xclim/issues
-.. _Contributing Guidelines: https://github.com/Ouranosinc/xclim/blob/master/.github/CONTRIBUTING.rst
+.. _Contributing Guidelines: https://github.com/Ouranosinc/xclim/blob/master/CONTRIBUTING.rst
 
 How to cite this library
 ------------------------
 If you wish to cite xclim in a research publication, we kindly ask that you use the bibliographical reference information available through `Zenodo`_
 
-.. _Zenodo: https://doi.org/10.5281/zenodo.2795043
+.. _Zenodo: https://doi.org/10.5281/zenodo.2795043S
+
+License
+-------
+This is free software: you can redistribute it and/or modify it under the terms of the `Apache License 2.0`_. A copy of this license is provided in the code repository (`LICENSE`_).
+
+
+.. _Apache License 2.0: https://opensource.org/license/apache-2-0/
+.. _LICENSE: https://github.com/Ouranosinc/xclim/blob/master/LICENSE
 
 Credits
 -------
