+++ paper/paper.md
@@ -46,7 +46,7 @@ bibliography: paper.bib
 
 Researchers and climate service providers analyze data from large ensembles of Earth System Model (ESM) simulations. These analyses typically include model comparisons with observations, bias-correction and statistical adjustment, computation of various climate indicators and diagnostics, and ensemble statistics. As the number of models contributing to these ensembles grows, so does the complexity of the code required to deal with model idiosyncrasies, outlier detection, unit conversion, etc. In addition, growing ensemble sizes and improvements in the spatiotemporal resolution of ESMs raise the computational costs of running those analyses. `xclim` is designed to meet the operational needs of climate service providers by offering algorithms for over 150 climate indicators, multiple downscaling algorithms, and ensemble statistics.
 
-The development of `xclim` started at Ouranos in 2018 out of the need to deliver data for a pan-Canadian atlas of climate indicators. In-house specialists had different implementations for the same indicators, and there was a desire to adopt a common library that would tie together investments in research and development with operational production capabilities. At the time, the package that was closest to meeting these requirements was `icclim` [@icclim], a library developed in the context of the [European Climate Assessment & Dataset](https://www.ecad.eu/) project, with the purpose of monitoring and analyzing changes in climate extremes. It was not, however, designed to be easily extensible, and we believed the indicators they offered could be written more succinctly and computed more efficiently by relying on objects and primitives from `xarray` [@Hoyer:2017]. `xclim` started as a reimplementation of `icclim` with an `xarray` backend, drawing inspiration from `MetPy` [@metpy], and eventually grew to include other algorithms routinely used in climate data analysis.
+The development of `xclim` started in 2018 at [Ouranos](https://www.ouranos.ca), a consortium on regional climatology and adaptation to climate change based in Montréal, out of the need to deliver data for a pan-Canadian atlas of climate indicators. In-house specialists had different implementations for the same indicators, and there was a desire to adopt a common library that would tie together investments in research and development with operational production capabilities. At the time, the package that was closest to meeting these requirements was `icclim` [@icclim], a library developed in the context of the [European Climate Assessment & Dataset](https://www.ecad.eu/) project, with the purpose of monitoring and analyzing changes in climate extremes. It was not, however, designed to be easily extensible, and we believed the indicators they offered could be written more succinctly and computed more efficiently by relying on objects and primitives from `xarray` [@Hoyer:2017]. `xclim` started as a reimplementation of `icclim` with an `xarray` backend, drawing inspiration from `MetPy` [@metpy], and eventually grew to include other algorithms routinely used in climate data analysis.
 
 `xclim` is meant to be one component in a larger software ecosystem for climate data analysis. Other libraries often used in tandem with `xclim` are `clisops` [@clisops] for spatial subsetting and averaging utilities, and `xESMF` [@xesmf] for spatial regridding.
 
