### 2023-02-27 Ouranosinc/xclim#1305 @huard
Added text in paper to describe Ouranos.
