### 2023-02-27 771e32a
Bump version: 0.40.11-beta → 0.40.12-beta
