### 2023-02-27 02b0ede
Update HISTORY.rst
Co-authored-by: Trevor James Smith <10819524+Zeitsperre@users.noreply.github.com>
