### 2023-02-27 fd4ebc6
Update xclim/indicators/land/_snow.py
english 2/2

Co-authored-by: Trevor James Smith <10819524+Zeitsperre@users.noreply.github.com>
