+++ xclim/indices/_agro.py
@@ -887,23 +887,23 @@ def water_budget(
 
 @declare_units(
     pr="[precipitation]",
-    s_thresh_wet="[length]",
-    s_thresh_dry="[length]",
-    e_thresh_dry="[length]",
+    thresh_wet_start="[length]",
+    thresh_dry_start="[length]",
+    thresh_dry_end="[length]",
 )
 def rain_season(
     pr: xarray.DataArray,
-    s_thresh_wet: Quantified = "25.0 mm",
-    s_window_wet: int = 3,
-    s_window_not_dry: int = 30,
-    s_thresh_dry: Quantified = "1.0 mm",
-    s_window_dry: int = 7,
-    s_method_dry: str = "per_day",
+    thresh_wet_start: Quantified = "25.0 mm",
+    window_wet_start: int = 3,
+    window_not_dry_start: int = 30,
+    thresh_dry_start: Quantified = "1.0 mm",
+    window_dry_start: int = 7,
+    method_dry_start: str = "per_day",
     start_date_min: DayOfYearStr = "05-01",
     start_date_max: DayOfYearStr = "12-31",
-    e_thresh_dry: Quantified = "0.0 mm",
-    e_window_dry: int = 20,
-    e_method_dry: str = "per_day",
+    thresh_dry_end: Quantified = "0.0 mm",
+    window_dry_end: int = 20,
+    method_dry_end: str = "per_day",
     end_date_min: DayOfYearStr = "09-01",
     end_date_max: DayOfYearStr = "12-31",
     freq="AS-JAN",
@@ -914,36 +914,36 @@ def rain_season(
 
     pr: xr.DataArray
         Precipitation data.
-    s_thresh_wet: Quantified
-        Accumulated precipitation threshold associated with `s_window_wet`.
-    s_window_wet: int
-        Number of days when accumulated precipitation is above `s_thresh_wet`. Defines the first condition to start the rain season
-    s_window_not_dry: int
-        Number of days, after `s_window_wet` days, during which no dry period must be found as a second and last condition to start the rain season.
-        A dry sequence is defined with `s_thresh_dry`, `s_window_dry` and `s_method_dry`.
-    s_thresh_dry: Quantified
-        Threshold length defining a dry day in the sequence related to `s_window_dry`.
-    s_window_dry: int
-        Number of days used to define a dry sequence in the start of the season. Daily precipitations lower than `s_thresh_dry`
-        during `s_window_dry` days are considered a dry sequence. The precipitations must be lower than `s_thresh_dry`
-        for either every day in the sequence (`s_method_dry == "per_day"`) or for the total (`s_method_dry == "total"`).
-    s_method_dry: {"per_day", "total"}
-        Method used to define a dry sequence associated with `s_window_dry`. The threshold `s_thresh_dry` is either compared
-        to every daily precipitations (`s_method_dry == "per_day"`) or to total precipitations (`s_method_dry == "total"`)
-        in the sequence `s_window_dry` days.
+    thresh_wet_start: Quantified
+        Accumulated precipitation threshold associated with `window_wet_start`.
+    window_wet_start: int
+        Number of days when accumulated precipitation is above `thresh_wet_start`. Defines the first condition to start the rain season
+    window_not_dry_start: int
+        Number of days, after `window_wet_start` days, during which no dry period must be found as a second and last condition to start the rain season.
+        A dry sequence is defined with `thresh_dry_start`, `window_dry_start` and `method_dry_start`.
+    thresh_dry_start: Quantified
+        Threshold length defining a dry day in the sequence related to `window_dry_start`.
+    window_dry_start: int
+        Number of days used to define a dry sequence in the start of the season. Daily precipitations lower than `thresh_dry_start`
[…]
