+++ docs/references.bib
@@ -1882,11 +1882,15 @@ Publisher: Ubiquity Press},
 }
 
 @article{sivakumar_predicting_1998,
-	title = {Predicting rainy season potential from the onset of rains in Southern Sahelian and Sudanian climatic zones of West Africa},
-	author={Sivakumar, M.V.K.},
-	journal = {Agricultural and Forest Meteorology},
-	volume = {42},
-	number = {4},
-	pages = {295-305},
-	year = {1988},
+title = {Predicting rainy season potential from the onset of rains in Southern Sahelian and Sudanian climatic zones of West Africa},
+journal = {Agricultural and Forest Meteorology},
+volume = {42},
+number = {4},
+pages = {295-305},
+year = {1988},
+issn = {0168-1923},
+doi = {https://doi.org/10.1016/0168-1923(88)90039-1},
+url = {https://www.sciencedirect.com/science/article/pii/0168192388900391},
+author = {Sivakumar, M.V.K.},
+abstract = {Analysis of long-term daily rainfall data for 58 locations in the Southern Sahelian and Sudanian climatic zones of West Africa showed that a significant relationship exists between the date of onset of rains and the length of the growing season. Early onset of rains, relative to the computed mean date of onset for a given location, resulted in a longer growing season. Probabilities of growing season lengths for early, normal and delayed onset of rains have been computed for all the locations. This analysis has important applications in crop planning as well as disaster planning and forms an initial step in concepts such as “Response Farming” or “Weather-responsive Crop Management Tactics” for drought-prone West Africa.}
 }
