### 2023-02-28 56ad771
Apply suggestions from code review
Co-authored-by: Trevor James Smith <10819524+Zeitsperre@users.noreply.github.com>
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
