### 2023-02-28 48319af
docstring reworked
