+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.41.0-beta
+current_version = 0.41.0
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+).(?P<patch>\d+)(\-(?P<release>[a-z]+))?
