+++ HISTORY.rst
@@ -2,7 +2,7 @@
 History
 =======
 
-0.41.0 (unreleased)
+0.41.0 (2023-02-28)
 -------------------
 Contributors to this version: Trevor James Smith (:user:`Zeitsperre`), Pascal Bourgault (:user:`aulemahal`), Ludwig Lierhammer (:user:`ludwiglierhammer`), Éric Dupuis (:user:`coxipi`).
 
