### 2023-02-28 c6db408
Refactoring, corrected definitions, more comments
