+++ xclim/indices/_agro.py
@@ -899,18 +899,17 @@ def rain_season(
     thresh_dry_start: Quantified = "1.0 mm",
     window_dry_start: int = 7,
     method_dry_start: str = "per_day",
-    start_date_min: DayOfYearStr = "05-01",
-    start_date_max: DayOfYearStr = "12-31",
+    date_min_start: DayOfYearStr = "05-01",
+    date_max_start: DayOfYearStr = "12-31",
     thresh_dry_end: Quantified = "0.0 mm",
     window_dry_end: int = 20,
     method_dry_end: str = "per_day",
-    end_date_min: DayOfYearStr = "09-01",
-    end_date_max: DayOfYearStr = "12-31",
+    date_min_end: DayOfYearStr = "09-01",
+    date_max_end: DayOfYearStr = "12-31",
     freq="AS-JAN",
-    coord: str | None = None,
 ):
     """
-    Find the first and last day of the rain season and its length.
+    Find the length of the rain season and the day of year of its start and its end.
 
     pr: xr.DataArray
         Precipitation data.
@@ -931,9 +930,9 @@ def rain_season(
         Method used to define a dry sequence associated with `window_dry_start`. The threshold `thresh_dry_start` is either compared
         to every daily precipitations (`method_dry_start == "per_day"`) or to total precipitations (`method_dry_start == "total"`)
         in the sequence `window_dry_start` days.
-    start_date_min: DayOfYearStr
+    date_min_start: DayOfYearStr
         First day of year when season can start ("mm-dd").
-    start_date_max: DayOfYearStr
+    date_max_start: DayOfYearStr
         Last day of year when season can start ("mm-dd").
     thresh_dry_end: str
         Threshold length defining a dry day in the sequence related to `window_dry_end`.
@@ -945,16 +944,12 @@ def rain_season(
         Method used to define a dry sequence associated with `window_dry_end`. The threshold `thresh_dry_end` is either compared
         to every daily precipitations (`method_dry_end == "per_day"`) or to total precipitations (`method_dry_end == "total"`)
         in the sequence `window_dry` days.
-    end_date_min: DayOfYearStr
+    date_min_end: DayOfYearStr
         First day of year when season can end ("mm-dd").
-    end_date_max: DayOfYearStr
+    date_max_end: DayOfYearStr
         Last day of year when season can end ("mm-dd").
     freq : str
       Resampling frequency.
-    coord : {None, "default", name of a DateTimeAccessor}
-        If `coord == "default"`, the function returns values along `dim` instead of indexes (`coord is None`).
-        If `dim` has a datetime dtype, `coord` can also be a str of the name of the
-        DateTimeAccessor object to use (ex: 'dayofyear').
 
     Returns
     -------
@@ -962,11 +957,14 @@ def rain_season(
     rain_season_end: xr.DataArray, [dimensionless]
     rain_season_length: xr.DataArray, [dimensionless]
 
-    Note
+    Summary
     -------
-    The rain season starts with an intense period of raining (during `window_wet_start` days, there is a precipitation total of `thresh_wet_start`),
-    directly followed by a period of `window_not_dry_start` days where there is no dry sequence (`window_dry_start` days with a
-    minimum total or daily precipitation).
+    The rain season starts at the end of a period of raining (a total precipitation  of `thresh_wet_start` over `window_wet_start` days). This must
+    be directly followed by a period of `window_not_dry_start` days with no dry sequence. The dry sequence is a period of `window_dry_start`
+    days where precipitations are below `thresh_dry_start` (either the total precipitations over the period, or the daily precipitations, depending
[…]
