### 2023-02-28 284bd03
add lmoments3 to base requirements, remove conditional skipping for lmoments3, update CI
