+++ tox.ini
@@ -5,7 +5,7 @@ env_list =
     docs
     notebooks_doctests{-coverage,}
 ;    opt-slow
-    py38-lm3
+    py38
     py39-upstream-doctest
     py310-slow
 requires =
@@ -90,7 +90,6 @@ passenv =
 extras = dev
 deps =
     coverage: coveralls
-    lm3: git+https://github.com/OpenHydrology/lmoments3.git@develop#egg=lmoments3
     upstream: -rrequirements_upstream.txt
 install_command = python -m pip install --no-user {opts} {packages}
 download = True
