### 2023-03-01 a939edb
params can be used as input, and computed beforehand (`get_params`)
