### 2023-03-01 aeb8038
Clearer var names and better doc
