### 2023-03-01 d34c9a5
no longer use generic type for Mapping when parameters are always dicts
