### 2023-03-01 23216ae
Deal with type hinting warnings within class variable signatures
