### 2023-03-01 1180825
docstring fixes, mark note about class types
