### 2023-03-01 cbca142
fitting/rolling/resampling defined in a separate function
