### 2023-03-02 16f4b6b
resampling/rolling/fitting out of SPI
