+++ xclim/indices/_agro.py
@@ -885,23 +885,35 @@ def water_budget(
     return out
 
 
-# this might be useful elsewhere?
-def _get_group(da, freq):
-    if freq is None:
-        _get_group(da, xarray.infer_freq(da.time))
-    else:
-        _, base, _, _ = parse_offset(freq)
-        try:
-            group = {"D": "time.dayofyear", "M": "time.month"}[base]
-        except KeyError():
-            raise ValueError(f"Resampling frequency `{freq}` not supported.")
-    return group
-
-
-def _preprocess_spx(pr, freq, window):
+def _preprocess_spx(da, freq, window):
+    _, base, _, _ = parse_offset(freq or xarray.infer_freq(da.time))
+    try:
+        group = {"D": "time.dayofyear", "M": "time.month"}[base]
+    except KeyError():
+        raise ValueError(f"Standardized index with frequency `{freq}` not supported.")
     if freq:
-        pr = pr.resample(time=freq).mean(keep_attrs=True)
-    if uses_dask(pr) and len(pr.chunks[pr.get_axis_num("time")]) > 1:
+        da = da.resample(time=freq).mean(keep_attrs=True)
+    if window > 1:
+        da = da.rolling(time=window).mean(skipna=False, keep_attrs=True)
+    return da, group
+
+
+def _compute_spx_fit_params(da, cal_range, freq, window, dist, method, group=None):
+    # "WPM" method doesn't seem to work for gamma or pearson3
+    dist_and_methods = {"gamma": ["ML", "APP"], "fisk": ["ML", "APP"]}
+    if dist not in dist_and_methods:
+        raise NotImplementedError(f"The distribution `{dist}` is not supported.")
+    if method not in dist_and_methods[dist]:
+        raise NotImplementedError(
+            f"The method `{method}` is not supported for distribution `{dist}`."
+        )
+    if group is None:
+        da, group = _preprocess_spx(da, freq, window)
+
+    if cal_range:
+        da = da.sel(time=slice(cal_range[0], cal_range[1]))
+
+    if uses_dask(da) and len(da.chunks[da.get_axis_num("time")]) > 1:
         warnings.warn(
             "The input data is chunked on time dimension and must be fully rechunked to"
             " run `fit` on groups ."
@@ -909,22 +921,16 @@ def _preprocess_spx(pr, freq, window):
             " has to handle.",
             stacklevel=2,
         )
-        pr = pr.chunk({"time": -1})
-
-    # Rolling precipitations
-    if window > 1:
-        pr = pr.rolling(time=window).mean(skipna=False, keep_attrs=True)
-    pr.attrs["_processed"] = True
-    return pr
-
-
-def _compute_spx_fit_params(pr, cal_range, freq, window, dist, method, group=None):
-    group = group if group else _get_group(pr, freq)
-    if pr.attrs["_processed_"]:
-        pr = _preprocess_spx(pr, freq, window)
-    if cal_range:
-        pr = pr.sel(time=slice(cal_range[0], cal_range[1]))
-    return pr.groupby(group).map(fit, (dist, method))
+        da = da.chunk({"time": -1})
+    params = da.groupby(group).map(fit, (dist, method))
+    params.attrs.update(
+        {
+            "Calibration period": str(cal_range),
+            "Distribution": dist,
+            "Method": method,
+        }
+    )
+    return params
 
 
 @declare_units(
@@ -947,8 +953,8 @@ def standardized_precipitation_index(
     pr : xarray.DataArray
         Daily precipitation.
     cal_range: Tuple[DateStr, DateStr] | None
-        Dates used to subset `pr` and obtain dataset for calibration. The tuple is formed by two `DateStr`,  i.e. a `str` in format `"YYYY-MM-DD"`.
-        Default option `None` means that the full range of `pr` is used.
[…]
