### 2023-03-02 c89d850
collect API documentation under API pages, restructure API documentation, use a standardized approach to titles, move longer docstrings into the docs pages
