+++ xclim/testing/utils.py
@@ -1,4 +1,8 @@
-"""Testing and tutorial utilities' module."""
+# noqa: D205,D400
+"""
+Testing and Tutorial Utilities' Module
+======================================
+"""
 # Some of this code was copied and adapted from xarray
 from __future__ import annotations
 
