+++ xclim/indices/fire/__init__.py
@@ -1,6 +1,6 @@
 # noqa: D205,D400
 """
-Fire indices submodule
+Fire Indices Submodule
 ======================
 """
 
