### 2023-03-02 f5b0c0c
add new linkcheck exceptions, fix references
