+++ docs/analogues.rst
@@ -40,7 +40,7 @@ target climate:
 
 All methods accept arrays, the first is the reference (n, D) and the second is the candidate (m, D). Where the climate
 indicators vary along D and the distribution dimension along n or m. All methods output a single float. See their
-documentation in :ref:`analogues:Analogue metrics API`.
+documentation in :ref:`analogues:Analogues Metrics API`.
 
 .. warning::
 
