### 2023-03-02 d19605a
(WIP) More restructuring - working hyperlinks for indicators and indices, better flow
