+++ docs/indicators.rst
@@ -14,7 +14,6 @@ Indicators are split into realms (atmos, land, seaIce), according to the variabl
 See :ref:`notebooks/extendxclim:Defining new indicators` for instruction on how to create your own indicators. This page
 lists all indicators with a summary description, click on the names to get to the complete docstring of each indicator.
 
-
 atmos: Atmosphere
 =================
 
@@ -22,7 +21,7 @@ atmos: Atmosphere
 
    <dl>
    {% for indname, ind in indicators['atmos'].items() %}
-     <dt><code>atmos.{{ indname | safe}}</code> : <a class="reference_internal" href="indicators_api.html#xclim.indicators.atmos.{{ indname }}" title="{{ indname }}"><b>{{ ind.title }}</b></a></dt>
+     <dt><code>atmos.{{ indname | safe}}</code> : <a class="reference_internal" href="api.html#xclim.indicators.atmos.{{ indname }}" title="{{ indname }}"><b>{{ ind.title }}</b></a></dt>
      <dd>
      {% if ind.identifier != indname %}<b>Id: </b> {{ ind.identifier }} <br>{% endif %}
      <b>Description: </b> {{ ind.abstract }} <br>
@@ -40,7 +39,7 @@ land: Land surface
 
    <dl>
    {% for indname, ind in indicators['land'].items() %}
-     <dt><code>land.{{ indname | safe}}</code> : <a class="reference_internal" href="indicators_api.html#xclim.indicators.land.{{ indname }}" title="{{ indname }}"><b>{{ ind.title }}</b></a></dt>
+     <dt><code>land.{{ indname | safe}}</code> : <a class="reference_internal" href="api.html#xclim.indicators.land.{{ indname }}" title="{{ indname }}"><b>{{ ind.title }}</b></a></dt>
      <dd>
      {% if ind.identifier != indname %}<b>Id: </b> {{ ind.identifier }} <br>{% endif %}
      <b>Description: </b> {{ ind.abstract }} <br>
@@ -58,7 +57,7 @@ seaIce: Sea ice
 
    <dl>
    {% for indname, ind in indicators['seaIce'].items() %}
-     <dt><code>seaIce.{{ indname | safe}}</code> : <a class="reference_internal" href="indicators_api.html#xclim.indicators.seaIce.{{ indname }}" title="{{ indname }}"><b>{{ ind.title }}</b></a></dt>
+     <dt><code>seaIce.{{ indname | safe}}</code> : <a class="reference_internal" href="api.html#xclim.indicators.seaIce.{{ indname }}" title="{{ indname }}"><b>{{ ind.title }}</b></a></dt>
      <dd>
      {% if ind.identifier != indname %}<b>Id: </b> {{ ind.identifier }} <br>{% endif %}
      <b>Description: </b> {{ ind.abstract }} <br>
