### 2023-03-03 acb8a0e
use new generation method on RtD
