+++ .readthedocs.yml
@@ -12,8 +12,6 @@ build:
   tools:
     python: "mambaforge-4.10"
   jobs:
-#    pre_install:
-#      - pip install llvmlite~=0.38 --ignore-installed
     pre_build:
       - sphinx-apidoc -o docs/ --private --module-first xclim xclim/testing/tests xclim/indicators xclim/indices
       - rm docs/xclim.rst
