### 2023-03-03 7c5fe5a
fix linking of targets to prevent collisions on ReadTheDocs
