+++ docs/explanation.rst
@@ -12,15 +12,15 @@ Vestibulum lectus mauris ultrices eros in cursus turpis massa tincidunt. Amet lu
 Enim ut tellus elementum sagittis vitae et leo. Integer malesuada nunc vel risus commodo viverra maecenas accumsan lacus. Feugiat in fermentum posuere urna nec tincidunt praesent semper. Ac turpis egestas integer eget aliquet. Orci ac auctor augue mauris augue neque gravida in fermentum. Odio ut sem nulla pharetra diam. Nisi lacus sed viverra tellus in. At lectus urna duis convallis convallis. Nullam non nisi est sit amet facilisis magna etiam tempor. Mi eget mauris pharetra et ultrices neque ornare aenean euismod.
 
 Other projects similar to xclim
--------------------------------
-* `icclim` (`Source Code <https://github.com/cerfacs-globc/icclim>`_; `Documentation <https://icclim.readthedocs.io/en/stable/index.html>`_)
-   * `Their comparison page for ICCLIM and xclim <https://icclim.readthedocs.io/en/stable/explanation/xclim_and_icclim.html>`_
-* `climate_indices` (`Source Code <https://github.com/monocongo/climate_indices>`_; `Documentation <https://climate-indices.readthedocs.io/en/latest/index.html>`_)
-   * Discussion about the merging of `climate_indices` and `xclim`: :issue:`1273`.
-* `MetPy` (`Source Code <https://github.com/Unidata/MetPy>`_; `Documentation <https://unidata.github.io/MetPy/latest/index.html>`_)
-* `xcdat` (`Source Code <https://github.com/xCDAT/xcdat>`_; `Documentation <https://xcdat.readthedocs.io/en/latest/>`_)
+===============================
+* `icclim` (`icclim Source Code <https://github.com/cerfacs-globc/icclim>`_; `icclim Documentation <https://icclim.readthedocs.io/en/stable/index.html>`_)
+   - `Their comparison page for ICCLIM and xclim <https://icclim.readthedocs.io/en/stable/explanation/xclim_and_icclim.html>`_
+* `climate_indices` (`climate_indices Source Code <https://github.com/monocongo/climate_indices>`_; `climate_indices Documentation <https://climate-indices.readthedocs.io/en/latest/index.html>`_)
+   - Discussion about the merging of `climate_indices` and `xclim`: :issue:`1273`.
+* `MetPy` (`MetPy Source Code <https://github.com/Unidata/MetPy>`_; `MetPy Documentation <https://unidata.github.io/MetPy/latest/index.html>`_)
+* `xcdat` (`xcdat Source Code <https://github.com/xCDAT/xcdat>`_; `xcdat Documentation <https://xcdat.readthedocs.io/en/latest/>`_)
 
 R-language specific projects
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-* `climdex.pcic` (`Source Code <https://github.com/pacificclimate/climdex.pcic>`_; `R-CRAN Index <https://cran.r-project.org/web/packages/climdex.pcic/index.html>`_)
-* `climind` (`Source Code <https://github.com/ECA-D/climind>`_; `Documentation <https://rdrr.io/cran/ClimInd/man/>`_)
+----------------------------
+* `climdex.pcic` (`climdex.pcic Source Code <https://github.com/pacificclimate/climdex.pcic>`_; `climdex.pci R-CRAN Index <https://cran.r-project.org/web/packages/climdex.pcic/index.html>`_)
+* `climind` (`climind Source Code <https://github.com/ECA-D/climind>`_; `climind Documentation <https://rdrr.io/cran/ClimInd/man/>`_)
