+++ xclim/data/cf.yml
@@ -14,7 +14,7 @@ doc: |
   features that are currently not well covered by the CF Conventions. Currently identified issues frequently relate to
   standard_name and/or cell_methods which both are controlled vocabularies of the CF Conventions.
 
-  .. _clix-meta: https://github.com/clix-meta/clix-meta
+  .. _clix-meta project: https://github.com/clix-meta/clix-meta
 realm: atmos
 references: clix-meta https://github.com/clix-meta/clix-meta
 indicators:
