### 2023-03-06 ed25808
Update docs/explanation.rst
Co-authored-by: Abel Aoun <aoun@cerfacs.fr>
