+++ docs/explanation.rst
@@ -14,7 +14,7 @@ Enim ut tellus elementum sagittis vitae et leo. Integer malesuada nunc vel risus
 Other projects similar to xclim
 ===============================
 * `icclim` (`icclim Source Code <https://github.com/cerfacs-globc/icclim>`_; `icclim Documentation <https://icclim.readthedocs.io/en/stable/index.html>`_)
-   - `Their comparison page for ICCLIM and xclim <https://icclim.readthedocs.io/en/stable/explanation/xclim_and_icclim.html>`_
+   - `Their comparison page for icclim and xclim <https://icclim.readthedocs.io/en/stable/explanation/xclim_and_icclim.html>`_
 * `climate_indices` (`climate_indices Source Code <https://github.com/monocongo/climate_indices>`_; `climate_indices Documentation <https://climate-indices.readthedocs.io/en/latest/index.html>`_)
    - Discussion about the merging of `climate_indices` and `xclim`: :issue:`1273`.
 * `MetPy` (`MetPy Source Code <https://github.com/Unidata/MetPy>`_; `MetPy Documentation <https://unidata.github.io/MetPy/latest/index.html>`_)
