+++ CHANGES.rst
@@ -1,6 +1,6 @@
-=======
-History
-=======
+=========
+Changelog
+=========
 
 0.42.0 (unreleased)
 -------------------
