### 2023-03-06 03aca77
note about indices and indicators
