+++ xclim/testing/utils.py
@@ -414,7 +414,7 @@ def list_input_variables(
 def publish_release_notes(
     style: str = "md", file: os.PathLike | StringIO | TextIO | None = None
 ) -> str | None:
-    """Format release history in Markdown or ReStructuredText.
+    """Format release notes in Markdown or ReStructuredText.
 
     Parameters
     ----------
@@ -429,15 +429,15 @@ def publish_release_notes(
 
     Notes
     -----
-    This function is solely for development purposes.
+    This function is solely for development and packaging purposes.
     """
-    history_file = Path(__file__).parent.parent.parent.joinpath("CHANGES.rst")
+    changes_file = Path(__file__).parent.parent.parent.joinpath("CHANGES.rst")
 
-    if not history_file.exists():
-        raise FileNotFoundError("History file not found in xclim file tree.")
+    if not changes_file.exists():
+        raise FileNotFoundError("Changelog file not found in xclim folder tree.")
 
-    with open(history_file) as hf:
-        history = hf.read()
+    with open(changes_file) as hf:
+        changes = hf.read()
 
     if style == "rst":
         hyperlink_replacements = {
@@ -455,34 +455,34 @@ def publish_release_notes(
         raise NotImplementedError()
 
     for search, replacement in hyperlink_replacements.items():
-        history = re.sub(search, replacement, history)
+        changes = re.sub(search, replacement, changes)
 
     if style == "md":
-        history = history.replace("=======\nHistory\n=======", "# History")
+        changes = changes.replace("=========\nChangelog\n=========", "# Changelog")
 
         titles = {r"\n(.*?)\n([\-]{1,})": "-", r"\n(.*?)\n([\^]{1,})": "^"}
         for title_expression, level in titles.items():
-            found = re.findall(title_expression, history)
+            found = re.findall(title_expression, changes)
             for grouping in found:
                 fixed_grouping = (
                     str(grouping[0]).replace("(", r"\(").replace(")", r"\)")
                 )
                 search = rf"({fixed_grouping})\n([\{level}]{'{' + str(len(grouping[1])) + '}'})"
                 replacement = f"{'##' if level=='-' else '###'} {grouping[0]}"
-                history = re.sub(search, replacement, history)
+                changes = re.sub(search, replacement, changes)
 
         link_expressions = r"[\`]{1}([\w\s]+)\s<(.+)>`\_"
-        found = re.findall(link_expressions, history)
+        found = re.findall(link_expressions, changes)
         for grouping in found:
             search = rf"`{grouping[0]} <.+>`\_"
             replacement = f"[{str(grouping[0]).strip()}]({grouping[1]})"
-            history = re.sub(search, replacement, history)
+            changes = re.sub(search, replacement, changes)
 
     if not file:
-        return history
+        return changes
     if isinstance(file, (Path, os.PathLike)):
         file = Path(file).open("w")
-    print(history, file=file)
+    print(changes, file=file)
 
 
 def show_versions(
