+++ CHANGES.rst
@@ -8,20 +8,13 @@ Contributors to this version: Trevor James Smith (:user:`Zeitsperre`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^
+* Added `xclim` to the `ouranos Zenodo community <https://zenodo.org/communities/ouranos/>`_ . (:pull:`1313`).
 * Documentation adjustments. (:issue:`1305`, :pull:`1308`):
     * The CONTRIBUTING page has been moved to the top level of the repository.
     * Information concerning the licensing of xclim is clearly indicated in README.
     * `sphinx-autodoc-typehints` is now used to simplify call signatures generated in documentation.
     * The SDBA module API is now found with the rest of the User API documentation.
 * Removed `Mapping` abstract base class types in call signatures (`dict` variables were always expected). (:pull:`1308`).
-
-0.42.0 (unreleased)
--------------------
-Contributors to this version: Trevor James Smith (:user:`Zeitsperre`).
-
-Internal changes
-^^^^^^^^^^^^^^^^
-* Added `xclim` to the `ouranos Zenodo community <https://zenodo.org/communities/ouranos/>`_ . (:pull:`1313`).
 
 0.41.0 (2023-02-28)
 -------------------
