### 2023-03-07 9bba321
make note about Purpose section
