+++ docs/explanation.rst
@@ -5,11 +5,10 @@ Why use xclim?
 Purpose
 =======
 
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Pharetra magna ac placerat vestibulum lectus mauris ultrices. Vitae tortor condimentum lacinia quis vel eros donec ac odio. Et ligula ullamcorper malesuada proin libero. Nisl tincidunt eget nullam non nisi est sit amet. Congue nisi vitae suscipit tellus mauris a diam. In vitae turpis massa sed elementum tempus. Magna eget est lorem ipsum dolor sit amet consectetur. Tempor orci dapibus ultrices in iaculis nunc sed augue. Etiam sit amet nisl purus in mollis nunc. Pellentesque eu tincidunt tortor aliquam nulla facilisi cras fermentum odio. Etiam erat velit scelerisque in dictum non consectetur. Lectus magna fringilla urna porttitor rhoncus dolor purus non enim. Aliquet eget sit amet tellus cras adipiscing enim eu. Eleifend quam adipiscing vitae proin sagittis nisl rhoncus mattis rhoncus.
+.. important::
 
-Vestibulum lectus mauris ultrices eros in cursus turpis massa tincidunt. Amet luctus venenatis lectus magna fringilla urna porttitor rhoncus. Vestibulum sed arcu non odio euismod lacinia. Amet justo donec enim diam vulputate ut pharetra. Vitae suscipit tellus mauris a diam maecenas sed enim. Et netus et malesuada fames ac turpis egestas integer eget. Ut ornare lectus sit amet est. Nunc aliquet bibendum enim facilisis gravida neque convallis a. Volutpat odio facilisis mauris sit amet massa. Suspendisse in est ante in nibh. Pharetra vel turpis nunc eget lorem dolor sed viverra ipsum. Tristique senectus et netus et malesuada fames ac turpis egestas. Tortor pretium viverra suspendisse potenti nullam ac. Felis eget velit aliquet sagittis. Ipsum consequat nisl vel pretium lectus quam.
-
-Enim ut tellus elementum sagittis vitae et leo. Integer malesuada nunc vel risus commodo viverra maecenas accumsan lacus. Feugiat in fermentum posuere urna nec tincidunt praesent semper. Ac turpis egestas integer eget aliquet. Orci ac auctor augue mauris augue neque gravida in fermentum. Odio ut sem nulla pharetra diam. Nisi lacus sed viverra tellus in. At lectus urna duis convallis convallis. Nullam non nisi est sit amet facilisis magna etiam tempor. Mi eget mauris pharetra et ultrices neque ornare aenean euismod.
+    The content of this section is being developed in the forthcoming paper submission to JOSS.
+    This section will be updated when the wording has been finalized in :pull:`250`
 
 Other projects similar to xclim
 ===============================
