+++ .pre-commit-config.yaml
@@ -22,7 +22,7 @@ repos:
   - id: pretty-format-json
     args: ['--autofix', '--no-ensure-ascii', '--no-sort-keys']
 - repo: https://github.com/pappasam/toml-sort
-  rev: v0.22.2
+  rev: v0.22.4
   hooks:
   - id: toml-sort-fix
 - repo: https://github.com/pycqa/flake8
@@ -32,16 +32,16 @@ repos:
     additional_dependencies: ['flake8-rst-docstrings']
     args: ['--config=setup.cfg']
 - repo: https://github.com/psf/black
-  rev: 22.12.0
+  rev: 23.1.0
   hooks:
   - id: black
 - repo: https://github.com/PyCQA/isort
-  rev: 5.11.4
+  rev: 5.12.0
   hooks:
   - id: isort
     args: ['--settings-file=pyproject.toml', "--add_imports='from __future__ import annotations'"]
 - repo: https://github.com/nbQA-dev/nbQA
-  rev: 1.6.1
+  rev: 1.6.3
   hooks:
   - id: nbqa-black
     args: ['--target-version=py38']
@@ -65,7 +65,7 @@ repos:
   - id: blackdoc
     exclude: '(xclim/indices/__init__.py|docs/installation.rst)'
     additional_dependencies:
-    - 'black==22.12.0'
+    - 'black==23.1.0'
 - repo: https://github.com/adrienverge/yamllint.git
   rev: v1.29.0
   hooks:
