### 2023-03-07 856c4f0
Update wording, some grammar, add a few sections about other submodules, add internationalization section
