### 2023-03-08 #73 @Batalex
Thanks, @aguspesce, for your effort!

@Zeitsperre I was thinking of reviewing Xclim a bit after our reviewers. The more, the merrier 😄
