### 2023-03-08 1531dfd
Working version of purpose and add explanations for similarities with some projects
