### 2023-03-08 3208975
add OGC exception
