### 2023-03-08 5a1ca6b
fix makefile calls
