+++ CHANGES.rst
@@ -9,11 +9,16 @@ Contributors to this version: Trevor James Smith (:user:`Zeitsperre`).
 Internal changes
 ^^^^^^^^^^^^^^^^
 * Added `xclim` to the `ouranos Zenodo community <https://zenodo.org/communities/ouranos/>`_ . (:pull:`1313`).
-* Documentation adjustments. (:issue:`1305`, :pull:`1308`):
+* Significant documentation adjustments. (:issue:`1305`, :pull:`1308`):
     * The CONTRIBUTING page has been moved to the top level of the repository.
     * Information concerning the licensing of xclim is clearly indicated in README.
     * `sphinx-autodoc-typehints` is now used to simplify call signatures generated in documentation.
     * The SDBA module API is now found with the rest of the User API documentation.
+    * `HISTORY.rst` has been renamed `CHANGES.rst`, to follow `dask`-like conventions.
+    * Hyperlink targets for individual `indices` and `indicators` now point to their entries under `API` or `Indices`.
+    * Module-level docstrings have migrated from the library scripts directly into the documentation RestructuredText files.
+    * The documentation now includes a page explaining the reasons for developing `xclim` and a section briefly detailing similar and related projects.
+    * Markdown explanations in some Jupyter Notebooks have been edited for clarity
 * Removed `Mapping` abstract base class types in call signatures (`dict` variables were always expected). (:pull:`1308`).
 
 0.41.0 (2023-02-28)
