### 2023-03-08 Ouranosinc/xclim#1282 @Zeitsperre
@NickleDave

Changes that were just merged to the main branch (#1308) included modifications to the `usage` page (one of your initial comments). If you're interested in taking another look at that, please let me know. Otherwise, I'll consider this issue closed in the next few days.

Thanks again!
