+++ tox.ini
@@ -3,14 +3,14 @@ min_version = 4.0
 env_list =
     black
     docs
-    notebooks_doctests{-coverage,}
+    notebooks_doctests
 ;    opt-slow
     py38
     py39-upstream-doctest
     py310-slow
 requires =
     pip >= 21.0
-opts = -v
+opts = -vv
 
 [testenv:black]
 description = Run code quality compliance tests under {basepython}
@@ -25,6 +25,7 @@ deps =
     nbqa
     pydocstyle
     yamllint
+commands_pre =
 commands =
     pydocstyle --config=setup.cfg xclim
     flake8 --config=setup.cfg xclim
@@ -42,6 +43,7 @@ setenv =
     !notebooks: SKIP_NOTEBOOKS = 1
     PYTHONPATH = {toxinidir}
     READTHEDOCS = 1
+commands_pre =
 commands =
     make docs
 commands_post =
@@ -94,8 +96,8 @@ deps =
 install_command = python -m pip install --no-user {opts} {packages}
 download = True
 commands_pre =
-    pip list
-    pip check
+    python -m pip list
+    python -m pip check
 commands =
     doctest: pytest --no-cov --rootdir xclim/testing/tests/ --xdoctest xclim --ignore=xclim/testing/tests/
     !slow: pytest xclim -m "not slow"
