### 2023-03-08 56d6ef7
adjust ttest_1samp calls to support scipy>=1.9
