### 2023-03-08 60e9a70
adjust warnings, make note of potential API changes
