### 2023-03-08 36f6cce
test against older scipy on python3.8
