+++ tox.ini
@@ -91,6 +91,7 @@ passenv =
     XCLIM_*
 extras = dev
 deps =
+    py38: scipy<1.9
     coverage: coveralls
     upstream: -rrequirements_upstream.txt
 install_command = python -m pip install --no-user {opts} {packages}
