### 2023-03-08 1c601e2
try using deepcopy
