### 2023-03-09 2264d66
remove unnecessary deepcopy calls
