### 2023-03-09 61412f3
change call signature and add warning for both functions using mf_flag
