### 2023-03-10 Ouranosinc/xclim#1282 @NickleDave
Hi @Zeitsperre, I took a quick look and see in the PR where revisions were made in places where there were minor typos / unclear language. LGTM!!! Thank you for noting this and integrating it into the review.
