### 2023-03-10 16084d6
expose yearly_interpolated_doy and yearly_random_doy, update CHANGES.rst
