### 2023-03-10 ab2f34d
move indices module docstring into indices.rst for better documentation structure
