### 2023-03-10 eebd710
Apply suggestions from code review
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
