### 2023-03-10 79c82de
fix rendering bug in Indicators
