### 2023-03-10 93c0a28
Remove deprecated indicators (first_day_below, first_day_above, snow_cover_duration, continuous_snow_cover_start, continuous_snow_cover_end) and indices (snow_cover_duration, continuous_snow_cover_start, continuous_snow_cover_end), catch existing DeprecationWarnings, migrate deprecated test calls to new functions, note methods and indices that may require internal logic revisions
