### 2023-03-13 0ca4b63
Remove unnecessary FutureWarning checks (see #1319)
