### 2023-03-13 0e06e2c
Remove deprecated indicators (`land.freq_analysis`, `land.stats`, `land.fit`)
