### 2023-03-13 e23a5f7
update CHANGES.rst to reflect removals
