+++ xclim/testing/tests/test_generic_indicators.py
@@ -41,11 +41,11 @@ class TestReturnLevel:
             ndq_series, mode="max", t=[2, 5], dist="gamma", season="DJF"
         )
 
-        assert out.description in [
-            "Streamflow frequency analysis for the maximal winter 1-day flow "
-            "estimated using the gamma distribution."
-        ]
-        assert out.name == "q1maxwinter"
+        assert out.description == (
+            "Frequency analysis for the maximal winter 1-day value estimated using the "
+            "gamma distribution."
+        )
+        assert out.name == "fa_1maxwinter"
         assert out.shape == (2, 2, 3)  # nrt, nx, ny
         np.testing.assert_array_equal(out.isnull(), False)
 
@@ -75,13 +75,6 @@ class TestReturnLevel:
         )
         assert np.isnan(out.values[:, 0, 0]).all()
 
-    def test_wrong_variable(self, pr_series):
-        with pytest.raises(ValidationError):
-            with pytest.warns(DeprecationWarning):
-                generic.return_level(
-                    pr_series(np.random.rand(100)), mode="max", t=2, dist="gamma"
-                )
-
 
 class TestStats:
     """See other tests in test_land::TestStats"""
@@ -93,7 +86,7 @@ class TestStats:
 
     def test_ndq(self, ndq_series):
         out = generic.stats(ndq_series, freq="YS", op="min", season="MAM")
-        assert out.attrs["units"] == "m^3 s-1"
+        assert out.attrs["units"] == "m3 s-1"
 
     def test_missing(self, ndq_series):
         a = ndq_series
