### 2023-03-14 #73 @Zeitsperre
Hey there,

Just letting people know that the comments to date have all been addressed. The main development branch (and `latest` documentation) reflect these changes. Our next release is slated for the end of March, so if more comments arrive this week, we should be able to address them next week, in time for the release.

Thanks again!
