### 2023-03-14 #73 @Batalex
Hello @Zeitsperre,
Thanks your responsiveness and communication on Xclim release schedule. I am following up with @jmunroe to give you the best estimate possible as to when you will have the second review.

I really appreciate your patience on this submission.
