### 2023-03-15 1a373b9
First raw implementation, add a switch in first_run
