### 2023-03-16 377227d
redo 'realization' to input_core_dims and output_core_dims
