+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.41.3-beta
+current_version = 0.41.4-beta
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+).(?P<patch>\d+)(\-(?P<release>[a-z]+))?
