### 2023-03-20 0e4bd77
Adjust call signature in last_spring_frost indice
