### 2023-03-20 ea5f576
add operators to relevant indices
