### 2023-03-20 e693be1
add operators to season_length indices
