+++ .github/workflows/main.yml
@@ -39,8 +39,10 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: "3.8"
-      - name: Install tox
-        run: pip install tox~=4.0
+      - name: Install pylint and tox
+        run: pip install pylint tox~=4.0
+      - name: Run pylint
+        run: pylint --rcfile=pylintrc --disable=import-error --exit-zero xclim
       - name: Run linting suite
         run: tox -e black
 
@@ -62,26 +64,35 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
-  test-py39-upstream-coverage:
+  test-pypi:
     needs: black
+    name: test-${{ matrix.tox-env }}
     if: |
       contains(github.event.pull_request.labels.*.name, 'approved') ||
       (github.event.review.state == 'approved') ||
       (github.event_name == 'push')
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        - tox-env: py38-coverage
+          python-version: "3.8"
+        - tox-env: py39-coverage
+          python-version: "3.9"
+        - tox-env: notebooks_doctests
+          python-version: "3.8"
     steps:
       - uses: actions/checkout@v3
-      - name: Set up Python3.9
+      - name: Set up Python${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
-          python-version: "3.9"
+          python-version: ${{ matrix.tox-env }}
       - name: Install tox
         run: pip install tox~=4.0
       - name: Test with tox
-        run: tox -e py39-upstream-coverage
+        run: tox -e ${{ matrix.tox-env }}
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          COVERALLS_FLAG_NAME: run-py39-upstream
+          COVERALLS_FLAG_NAME: run-{{ matrix.tox-env }}
           COVERALLS_PARALLEL: true
           COVERALLS_SERVICE_NAME: github
 
@@ -143,58 +154,9 @@ jobs:
           COVERALLS_PARALLEL: true
           COVERALLS_SERVICE_NAME: github
 
-  test-py38-coverage:
-    needs: black
-    if: |
-      contains(github.event.pull_request.labels.*.name, 'approved') ||
-      (github.event.review.state == 'approved') ||
-      (github.event_name == 'push')
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python3.8
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.8"
-      - name: Install tox
-        run: pip install tox~=4.0
-      - name: Test with tox
-        run: tox -e py38-coverage
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          COVERALLS_FLAG_NAME: run-py38
-          COVERALLS_PARALLEL: true
-          COVERALLS_SERVICE_NAME: github
-
-  py38-notebooks_doctests:
-    needs: black
-    if: |
-      contains(github.event.pull_request.labels.*.name, 'approved') ||
-      (github.event.review.state == 'approved') ||
-      (github.event_name == 'push')
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python3.8
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.8"
-      - name: Install pylint and tox
-        run: pip install pylint tox~=4.0
-      - name: Run pylint
-        run: pylint --rcfile=pylintrc --disable=import-error --exit-zero xclim
[…]
