+++ .github/workflows/main.yml
@@ -74,12 +74,12 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        - tox-env: py38-coverage
+        - tox-env: "py38-coverage"
           python-version: "3.8"
-        - tox-env: py39-coverage
+        - tox-env: "py39-coverage"
           python-version: "3.9"
-        - tox-env: notebooks_doctests
-          python-version: "3.8"
+        - tox-env: "notebooks_doctests"
+          python-version: "3.10"
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python${{ matrix.python-version }}
