+++ .github/workflows/main.yml
@@ -66,7 +66,7 @@ jobs:
 
   test-pypi:
     needs: black
-    name: test-${{ matrix.tox-env }}
+    name: test-Python${{ matrix.tox-env }}
     if: |
       contains(github.event.pull_request.labels.*.name, 'approved') ||
       (github.event.review.state == 'approved') ||
@@ -86,7 +86,7 @@ jobs:
       - name: Set up Python${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
-          python-version: ${{ matrix.tox-env }}
+          python-version: ${{ matrix.python-version }}
       - name: Install tox
         run: pip install tox~=4.0
       - name: Test with tox
@@ -104,6 +104,11 @@ jobs:
       (github.event.review.state == 'approved') ||
       (github.event_name == 'push')
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        include:
+          - tox-env: py310
+            python-version: "3.10"
     defaults:
       run:
         shell: bash -l {0}
@@ -116,23 +121,22 @@ jobs:
           micromamba-version: 'latest'
           environment-file: environment.yml
           extra-specs: |
-            python="3.10"
+            python=${{ matrix.python-version }}
       - name: Conda and Mamba versions
         run: |
           conda --version
           echo "micromamba: $(micromamba --version)"
-      - name: Install dependencies
+      - name: Install xclim
         run: |
           python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git@5d11981792#egg=sbck"
           python -m pip install --no-user .
-      - name: Test
+      - name: Check versions
         run: |
-          pip list
+          conda list
+          xclim show_version_info
           pip check
+      - name: Test with pytest
           pytest xclim --numprocesses=logical --durations=10 --cov=xclim --cov-report=term-missing
-        env:
-          CONDA_EXE: mamba
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 #      - name: Install tox
 #        shell: bash -l {0}
 #        run: |
@@ -151,7 +155,7 @@ jobs:
           coveralls
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          COVERALLS_FLAG_NAME: run-py310-opt-slow
+          COVERALLS_FLAG_NAME: run-{{ matrix.tox-env }}-opt-slow
           COVERALLS_PARALLEL: true
           COVERALLS_SERVICE_NAME: github
 
