+++ .github/workflows/main.yml
@@ -136,6 +136,7 @@ jobs:
           xclim show_version_info
           pip check
       - name: Test with pytest
+        run: |
           pytest xclim --numprocesses=logical --durations=10 --cov=xclim --cov-report=term-missing
 #      - name: Install tox
 #        shell: bash -l {0}
