### 2023-03-20 d86d0ca
use strategy for all stages
