+++ .github/workflows/upstream.yml
@@ -42,14 +42,15 @@ jobs:
             (github.event_name == 'schedule' || github.event_name == 'workflow_dispatch')
             || needs.detect-ci-trigger.outputs.triggered == 'true'
         )
-    defaults:
-      run:
-        shell: bash -l {0}
+
     strategy:
       fail-fast: false
       matrix:
         python-version:
           - "3.10"
+    defaults:
+      run:
+        shell: bash -l {0}
     steps:
       - uses: actions/checkout@v3
         with:
