+++ .github/workflows/main.yml
@@ -29,16 +29,20 @@ jobs:
       ((github.event_name == 'pull_request') && (github.event.action != 'labeled')) ||
       (github.event.review.state == 'approved') ||
       (github.event_name == 'push')
+    strategy:
+      matrix:
+        python-version:
+          - "3.8"
     steps:
       - name: Cancel previous runs
         uses: styfle/cancel-workflow-action@0.11.0
         with:
           access_token: ${{ github.token }}
       - uses: actions/checkout@v3
-      - name: Set up Python3.8
+      - name: Set up Python${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
-          python-version: "3.8"
+          python-version: ${{ matrix.python-version }}
       - name: Install pylint and tox
         run: pip install pylint tox~=4.0
       - name: Run pylint
@@ -51,16 +55,22 @@ jobs:
     if: |
       (github.event_name == 'pull_request') && !contains(github.event.pull_request.labels.*.name, 'approved')
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        tox-env:
+          - "py38"
+        python-version:
+          - "3.8"
     steps:
       - uses: actions/checkout@v3
-      - name: Set up Python3.8
+      - name: Set up Python${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
-          python-version: "3.8"
+          python-version: ${{ matrix.python-version }}
       - name: Install tox
         run: pip install tox~=4.0
       - name: Test with tox
-        run: tox -e py38
+        run: tox -e ${{ matrix.tox-env }}
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
