+++ .github/workflows/main.yml
@@ -111,7 +111,7 @@ jobs:
           echo "micromamba: $(micromamba --version)"
       - name: Install dependencies
         run: |
-          python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git@5d11981792#egg=sbck"
+          python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck"
           python -m pip install --no-user .
       - name: Test
         run: |
