### 2023-03-21 2a54094
add eofs to CI build
