+++ .github/workflows/main.yml
@@ -109,8 +109,9 @@ jobs:
         run: |
           conda --version
           echo "micromamba: $(micromamba --version)"
-      - name: Install dependencies
+      - name: Install optional dependencies
         run: |
+          python -m pip install --no-user eofs
           python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck"
           python -m pip install --no-user .
       - name: Test
