+++ .github/workflows/main.yml
@@ -111,6 +111,7 @@ jobs:
           echo "micromamba: $(micromamba --version)"
       - name: Install optional dependencies
         run: |
+          python -m pip install --no-user pybind11
           python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck"
           python -m pip install --no-user .
       - name: Test
