### 2023-03-21 bee89c6
add eigen3 to upstream base libraries
