### 2023-03-22 4d33e20
SBCK in environment.yml and no longer in upstream requirements, add eofs to base requirements, remove optional handling for eofs, add Eigen3 install documentation, add `sbck` tox flag
