+++ tox.ini
@@ -68,8 +68,6 @@ commands =
 ;[testenv:opt-{slow,not_slow}]
 ;description = Run tests with optional requirements (SBCK (experimental), eofs) and pytest under {basepython} (Anaconda distribution)
 ;conda_env = environment-dev.yml
-;commands_pre =
-;    python -m pip install --no-user "git+https://github.com/Ouranosinc/SBCK.git@easier-install#egg=sbck&subdirectory=python"
 ;commands =
 ;    pip check
 ;    !slow: pytest xclim -m "not slow" --durations=10
@@ -94,6 +92,8 @@ deps =
     py38: scipy<1.9
     coverage: coveralls
     upstream: -rrequirements_upstream.txt
+    sbck: pybind11
+    sbck: git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck
 install_command = python -m pip install --no-user {opts} {packages}
 download = True
 commands_pre =
