### 2023-03-22 e3572cd
use new convention for VCS install
