+++ .github/workflows/main.yml
@@ -116,7 +116,7 @@ jobs:
       - name: Install optional dependencies
         run: |
           python -m pip install --no-user pybind11
-          python -m pip install --no-user "git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck"
+          python -m pip install --no-user "sbck @ git+https://github.com/yrobink/SBCK-python.git@master"
           python -m pip install --no-user .
       - name: Test
         run: |
