+++ tox.ini
@@ -93,7 +93,7 @@ deps =
     coverage: coveralls
     upstream: -rrequirements_upstream.txt
     sbck: pybind11
-    sbck: git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck
+    sbck: sbck @ git+https://github.com/yrobink/SBCK-python.git@master
 install_command = python -m pip install --no-user {opts} {packages}
 download = True
 commands_pre =
