+++ environment.yml
@@ -68,4 +68,4 @@ dependencies:
     - pip
     - pip:
         - sphinxcontrib-svg2pdfconverter
-        - git+https://github.com/yrobink/SBCK-python.git@master#egg=sbck
+        - sbck @ git+https://github.com/yrobink/SBCK-python.git@master
