### 2023-03-22 b54861c
Add resample before/after switch in more indicators
