### 2023-03-23 f2f05e1
Move long abstract: indicator -> index
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
