+++ xclim/indicators/atmos/_precip.py
@@ -640,10 +640,6 @@ rain_season = PrecipWithIndexing(
         "Number of steps of the original series in the season, between 'start' and 'end'.",
     ],
     units="",
-    abstract="The rain season begins when two conditions are met: 1) There must be a number of wet days with"
-    "precipitations above or equal to a given threshold; 2) There must be another sequence following, where, for a given period in time, there are no"
-    "dry sequence (i.e. a certain number of days where precipitations are below or equal to a certain threshold). The rain season ends"
-    "when there is a dry sequence.",
     cell_methods="",
     compute=indices.rain_season,
 )
