### 2023-03-23 cc35876
Added back shorter abstract to indicator
