### 2023-03-23 5277954
rollback eofs to optional dependency, remove redundant steps in CI, add eofs flag in tox
