+++ pyproject.toml
@@ -36,7 +36,6 @@ dependencies = [
   "cftime>=1.4.1",
   "Click>=8.1",
   "dask[array]>=2.6",
-  "eofs",
   "jsonpickle",
   "lmoments3>=1.0.5",
   "numba",
