+++ tox.ini
@@ -92,6 +92,7 @@ deps =
     py38: scipy<1.9
     coverage: coveralls
     upstream: -rrequirements_upstream.txt
+    eofs: eofs
     sbck: pybind11
     sbck: sbck @ git+https://github.com/yrobink/SBCK-python.git@master
 install_command = python -m pip install --no-user {opts} {packages}
