### 2023-03-23 ffcab0a
Spacing in doc and revert unnecessary change in run_length
