+++ .github/workflows/upstream.yml
@@ -4,8 +4,15 @@ on:
     branches:
       - master
   pull_request:
+    types:
+      - opened
+      - reopened
+      - synchronize
     branches:
       - master
+  pull_request_review:
+    types:
+      - submitted
   schedule:
     - cron: "0 0 * * *" # Daily “At 00:00” UTC
   workflow_dispatch: # allows you to trigger the workflow run manually
@@ -15,34 +22,15 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  detect-ci-trigger:
-    name: detect upstream-dev ci trigger
-    runs-on: ubuntu-latest
-    if: |
-      github.repository == 'pydata/xarray'
-      && (github.event_name == 'push' || github.event_name == 'pull_request')
-    outputs:
-      triggered: ${{ steps.detect-trigger.outputs.trigger-found }}
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 2
-      - uses: xarray-contrib/ci-trigger@v1
-        id: detect-trigger
-        with:
-          keyword: "[test-upstream]"
-
   upstream-dev:
     name: upstream-dev
     runs-on: ubuntu-latest
-    needs: detect-ci-trigger
     if: |
-        always()
-        && (
-            (github.event_name == 'schedule' || github.event_name == 'workflow_dispatch')
-            || needs.detect-ci-trigger.outputs.triggered == 'true'
-        )
-
+      github.event_name == 'schedule' ||
+      github.event_name == 'workflow_dispatch' ||
+      contains(github.event.pull_request.labels.*.name, 'approved') ||
+      (github.event.review.state == 'approved') ||
+      (github.event_name == 'push') }}
     strategy:
       fail-fast: false
       matrix:
@@ -73,7 +61,6 @@ jobs:
           python -m pip install -r requirements_upstream.txt
       - name: Install xclim
         run: |
-          python -m pip install --no-user --no-deps "git+https://github.com/yrobink/SBCK-python.git@5d11981792#egg=sbck"
           python -m pip install --no-user --no-deps .
       - name: Check versions
         run: |
