+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.41.5-beta
+current_version = 0.41.6-beta
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+).(?P<patch>\d+)(\-(?P<release>[a-z]+))?
