### 2023-03-23 2200bd2
Bump version: 0.41.7-beta → 0.41.8-beta
