### 2023-03-23 291ef82
Reverted to old rle_with_holes which had less problems in corner cases
