### 2023-03-24 9aa36f3
Support new heat indices
