### 2023-03-24 5ffda6f
Change class to `Precip`
