### 2023-03-24 34583b2
Fix sdba_encode_cf bug
