+++ xclim/testing/tests/test_sdba/test_adjustment.py
@@ -489,36 +489,41 @@ class TestQM:
 
     @pytest.mark.parametrize("use_dask", [True, False])
     def test_add_dims(self, use_dask):
-        if use_dask:
-            chunks = {"location": -1}
-        else:
-            chunks = None
-        ref = (
-            open_dataset(
-                "sdba/ahccd_1950-2013.nc", chunks=chunks, drop_variables=["lat", "lon"]
+        with set_options(sdba_encode_cf=use_dask):
+            if use_dask:
+                chunks = {"location": -1}
+            else:
+                chunks = None
+            ref = (
+                open_dataset(
+                    "sdba/ahccd_1950-2013.nc",
+                    chunks=chunks,
+                    drop_variables=["lat", "lon"],
+                )
+                .sel(time=slice("1981", "2010"))
+                .tasmax
             )
-            .sel(time=slice("1981", "2010"))
-            .tasmax
-        )
-        ref = convert_units_to(ref, "K")
-        ref = ref.isel(location=1, drop=True).expand_dims(location=["Amos"])
+            ref = convert_units_to(ref, "K")
+            ref = ref.isel(location=1, drop=True).expand_dims(location=["Amos"])
 
-        dsim = open_dataset(
-            "sdba/CanESM2_1950-2100.nc", chunks=chunks, drop_variables=["lat", "lon"]
-        ).tasmax
-        hist = dsim.sel(time=slice("1981", "2010"))
-        sim = dsim.sel(time=slice("2041", "2070"))
+            dsim = open_dataset(
+                "sdba/CanESM2_1950-2100.nc",
+                chunks=chunks,
+                drop_variables=["lat", "lon"],
+            ).tasmax
+            hist = dsim.sel(time=slice("1981", "2010"))
+            sim = dsim.sel(time=slice("2041", "2070"))
 
-        # With add_dims, "does it run" test
-        group = Grouper("time.dayofyear", window=5, add_dims=["location"])
-        EQM = EmpiricalQuantileMapping.train(ref, hist, group=group)
-        EQM.adjust(sim).load()
+            # With add_dims, "does it run" test
+            group = Grouper("time.dayofyear", window=5, add_dims=["location"])
+            EQM = EmpiricalQuantileMapping.train(ref, hist, group=group)
+            EQM.adjust(sim).load()
 
-        # Without, sanity test.
-        group = Grouper("time.dayofyear", window=5)
-        EQM2 = EmpiricalQuantileMapping.train(ref, hist, group=group)
-        scen2 = EQM2.adjust(sim).load()
-        assert scen2.sel(location=["Kugluktuk", "Vancouver"]).isnull().all()
+            # Without, sanity test.
+            group = Grouper("time.dayofyear", window=5)
+            EQM2 = EmpiricalQuantileMapping.train(ref, hist, group=group)
+            scen2 = EQM2.adjust(sim).load()
+            assert scen2.sel(location=["Kugluktuk", "Vancouver"]).isnull().all()
 
 
 class TestPrincipalComponents:
