### 2023-03-24 8c9d546
use open_dataset fixture in sdba tests, use noqas
