### 2023-03-27 #73 @jmunroe
Thank so much for your patience with me. I have now gone through the package carefully following the PyOpenSci review instructions.  This package is an incredibly thorough treatment of climate based metrics and I found it a real pleasure to review. A few issue I identified are presented below but none of the them are serious and I recommend this package be accepted by PyOpenSci.

I hope that in acceptance by PyOpenSci (and hopefully by JOSS as well), xclim will continue to be extended and adopted by a wide international community of climate scientists and professionals.  This is an important body of work that should empower those trying to understand, mitigate, and adapt to our changing climate.

  > However, please see the commments regarding installation documentation below.
  > Nice documentation! I really appreciated that the documenation was not *just* about the API for the function but included significant detail on what the actually calculation or metric did (with the associated reference).
   > Thanks for already moving `CONTRIBUTING.rst` to the top-level directory.

- [x] Package has a `README.rst` file in the root directory.

  > The README does not include explicit installation instructions. It does refer to the package documentation which does contain installation instructions.
  > Not applicable for xclim
  > Descriptive links to all vingettes are not in the README, but they are easily findable in the included pacakge documentation.
  > I am not aware of another Python package that does exactly the same thing xclim in the scientific ecosystem.  There are potentially related packages (such as MetPy for analyzing weather data, or climpred for computing metrcs for earth systems forecasts) in the broader ecosystem that could be referenced in the documentation just in case a potential user would be better served by a different project.
  >
  > Another potential Python package to reference is NCAR's GeoCAT (which is a part of NCAR overall strategy to migrate from away from NCL to Python). Historically, my understanding is non-Python tools like NCL were more widely used to do similar things to xclim.
  >
  > All three of these packages (MetPy, climpred, and GeoCAT) are based on the same xarray/dask foundations like xclim.

  > I focused on the worked examples provided in the documentation (consisting of ~10 .ipynb notebooks). xclim appears to work as described and the indicators and indices that I examined all looked reasonable and are very well documented.
  > I am unable to independently verify the correct implementation of every indicator/idice included with this software. However, the significant amount of detail in the documentation and the substantial references to peer-review literature gives me a high degree of confidence in the implementation.
  > Try to verify locally with `pytest` on Python 3.10.  There were a few failed tests in a local dev environment (33 failed, 1359 passed).  I have not attempted to debug any of these failing tests.
  >
  > I also note that on GitHub CI with more recent fixes that that the tag 0.40.0 I was reviewing, it appears the entire testing suite passes without failures.
  >
  > I am more suspicious of my own development environment than xclim itself and I am satisfied if the testing suite passes on GitHub actions, as it appears to do.
  >
  > `$ pytest xclim --numprocesses=logical --durations=10 --cov=xclim --cov-report=term-missing`
  >
  > ```
  >
  > =============================================== short test summary info ================================================
  > FAILED xclim/testing/tests/test_atmos.py::TestWaterBudget::test_convert_units - AttributeError: 'Dataset' object has no attribute 'rsus'. Did you mean: 'rsds'?
  > FAILED xclim/testing/tests/test_atmos.py::TestWaterBudget::test_nan_values - AttributeError: 'Dataset' object has no attribute 'rsus'. Did you mean: 'rsds'?
  > FAILED xclim/testing/tests/test_atmos.py::TestUTCI::test_universal_thermal_climate_index - AttributeError: 'Dataset' object has no attribute 'rsus'. Did you mean: 'rsds'?
  > FAILED xclim/testing/tests/test_atmos.py::TestPotentialEvapotranspiration::test_convert_units - AttributeError: 'Dataset' object has no attribute 'rsus'. Did you mean: 'rsds'?
  > FAILED xclim/testing/tests/test_atmos.py::TestPotentialEvapotranspiration::test_nan_values - AttributeError: 'Dataset' object has no attribute 'rsus'. Did you mean: 'rsds'?
  > FAILED xclim/testing/tests/test_atmos.py::test_wind_chill_index - AssertionError:
  > FAILED xclim/testing/tests/test_cffwis.py::TestCFFWIS::test_fire_weather_ufunc_overwintering - AssertionError:
  > FAILED xclim/testing/tests/test_ffdi.py::TestFFDI::test_ffdi_indicators[xlim-True] - KeyError: 'rh'
  > FAILED xclim/testing/tests/test_ffdi.py::TestFFDI::test_ffdi_indicators[xlim-False] - KeyError: 'rh'
  > FAILED xclim/testing/tests/test_ffdi.py::TestFFDI::test_ffdi_indicators[discrete-True] - KeyError: 'rh'
  > FAILED xclim/testing/tests/test_ffdi.py::TestFFDI::test_ffdi_indicators[discrete-False] - KeyError: 'rh'
  > FAILED xclim/testing/tests/test_indices.py::TestTG::test_simple[tg_mean-283.1391] - AssertionError:
  > FAILED xclim/testing/tests/test_indices.py::TestTG::test_simple[tg_min-266.1117] - AssertionError:
  > FAILED xclim/testing/tests/test_indices.py::TestTG::test_simple[tg_max-292.125] - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::TestDaysWithSnow::test_simple - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::test_days_over_precip_doy_thresh - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::test_days_over_precip_thresh - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::test_days_over_precip_thresh__seasonal_indexer - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::test_fraction_over_precip_doy_thresh - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::test_fraction_over_precip_thresh - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::test_dry_spell - AssertionError:
  > FAILED xclim/testing/tests/test_precip.py::test_dry_spell_frequency_op - RuntimeError: NetCDF: HDF error
  > FAILED xclim/testing/tests/test_run_length.py::test_run_bounds_data[True] - AssertionError:
  > FAILED xclim/testing/tests/test_run_length.py::test_keep_longest_run_data[True] - AssertionError:
  > FAILED xclim/testing/tests/test_run_length.py::test_run_bounds_data[False] - AssertionError:
  > FAILED xclim/testing/tests/test_run_length.py::test_keep_longest_run_data[False] - AssertionError:
  > FAILED xclim/testing/tests/test_snow.py::TestSndMaxDoy::test_no_snow - AssertionError:
  > FAILED xclim/testing/tests/test_temperature.py::TestWarmSpellDurationIndex::test_warm_spell_duration_index - AssertionError:
  > FAILED xclim/testing/tests/test_temperature.py::TestFreezeThawSpell::test_freezethaw_spell_frequency - AssertionError:
  > FAILED xclim/testing/tests/test_temperature.py::TestFreezeThawSpell::test_freezethaw_spell_mean_length - AssertionError:
  >
  > FAILED xclim/testing/tests/test_temperature.py::TestFreezeThawSpell::test_freezethaw_spell_max_length - AssertionError:
  > FAILED xclim/testing/tests/test_temperature.py::test_corn_heat_units - AssertionError:
  > FAILED xclim/testing/tests/test_sdba/test_properties.py::TestProperties::test_spatial_correlogram - AssertionError:
  > =================== 33 failed, 1359 passed, 67 skipped, 2 xfailed, 96 warnings in 341.70s (0:05:41) ====================
  >
  > ```

    > Both pylint and black are configured through GitHub actions for CI.

> For the JOSS paper information, I am using the `joss_submission` tag from GitHub.

---

##### Installation notes

The reviewer instructions appears leave it up to the reviewer on how to install the package.  I cloned the repo:

`git clone https://github.com/Ouranosinc/xclim.git`

And then, to ensure I was looking at the same version as was submitted for review, I switched to the `v0.40.0` tag:

`git checkout tags/v0.40.0`

Following the instructions at the `xclim` documentation under Installation, I created a separate conda environment to install the required dependencies:

```
conda create -n my_xclim_env python=3.8 --file=environment.yml
conda activate my_xclim_env
pip install ".[dev]"
```

And there I hit my first issue:

```
CondaValueError: could not parse 'name: xclim' in: environment.yml
```

The fix (at least for conda 22.11.1) is that `--file` is an option to pass to `conda env create` and not `conda create`. This needs to be fixed in the install instructions.

The instructions refer to 'Extra dependencies' such as `flox`, `SBCK`, `eigen`, `eofs`, `pybind`.  Since I used conda's `environment.yml` the extras `eofs`, `eigen`, `pybind11` were already included.

I confess I tend to get confused when there is the option of using either `environment.yml` and `requirements_*.txt` files. So, I skipped the instructions following 'Extra Dependencies' in the documentation.
I assume there must be situtations when I should and should not install these extra dependencies but as a new user of the package, I don't what those situations are yet.
Since theses installation instructions are right near the top of the documenation, perhaps it would be better for the maintainers to make those choices for me?  For example, I am now wondering "should I be installing `flox`?". Since it is 'highly recommended', would it not make more sense to have it as part of the default instructions?

##### Basic Usage

To evaluate

I installed Jupyter lab and created a notebook to test the instructions given under the *Basic Usage* section of the documentation.  The very first example

```
# ds = xr.open_dataset("your_file.nc")
ds = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc")
ds.tas
```

My initial reading of this code made me think that this ERA5 dataset was something I need to first download locally (I did not distinguish between `xr.open_dataset` and `open_dataset` in my very first glance at the code).
After some review, I see now that there companion GitHub repo that was available that had testing data and the `xclim.testing` API automatically makes a locally cached copy of this file.  I think it would be clearer if this very first example was written out as
```
# ds = xr.open_dataset("your_file.nc")
ds = xclim.testing.open_dataset("ERA5/daily_surface_cancities_1990-1993.nc")
ds.tas
```
so that it was clear that the `open_dataset` was utility method of the testing framework for xclim.

I liked how this initital documentation oriented the user quickly to the differences between 'indicators' and 'indices' which, especially for the new-user, may be something that could be confusing.

In the example of *Health checks and metadata attributes* there is a typo:
```
gdd = xclim.atmos.growing_degree_days(tas=ds6h.tas, thresh="10.0 degC", freq="MS")
```
```
gdd = xclim.atmos.growing_degree_days(tas=ds6h.air, thresh="10.0 degC", freq="MS")
```

The final basic usage examples on *Graphics* are could be improved by adding some descriptions in the text of what those three visualizations are actually showing. This could be as brief as taking the comment from each code snippet and using that as a text description.
While in-code comments are generally fine, these last few examples on graphics feel tacked on given the strong narrative text established in the beginning of the *Basic Usage* section of the documentation.

##### Examples

###### Workflow Examples

Minor spelling error in the docs:

"Finally, `xarray` is tightly in*(t)*egrated with `dask`, a package that can automatically parallelize operations.

Under *Subsetting and selecting data with xarray*, the documentation reads:

> Usually, xclim users are encouraged to use the subsetting utilities of the clisops package. Here, we will reduce the size of our data using the methods implemented in xarray

This is confusing because, as the first example workflow, the user has not yet been shown to use the clisops package. Should there be a subsub-section immediately before such as *Subsetting and selecting data with cliops* to demonstrate that recommended workflow?

Under *Different ways of resampling* there is setting of a matplotlib style

```
# import plotting stuff

%matplotlib inline
plt.style.use("seaborn")
plt.rcParams["figure.figsize"] = (11, 5)
```

that leads to the warning

```
/tmp/ipykernel_7039/887583071.py:5: MatplotlibDeprecationWarning: The seaborn styles shipped by Matplotlib are deprecated since 3.6, as they no longer correspond to the styles shipped by seaborn. However, they will remain available as 'seaborn-v0_8-<style>'. Alternatively, directly use the seaborn API instead.
  plt.style.use("seaborn")
```

I think the offending line should be changed to
```
plt.sytle.use("seaborn-v0_8")
```
(and elsewhere in the documenation where seaborn styles are used)

Since the example is supposed to contrast the differences between the two sampling methods, I think they should be on the same colorbar scale.  There may be a more general solution for this with matplotlib, but explicitly setting the range is a quick fix:
```
hw_before.sel(time="2010-07-01").plot(vmin=0, vmax=7)
plt.title("Resample, then run length")
plt.figure()
hw_after.sel(time="2010-07-01").plot(vmin=0, vmax=7)
plt.title("Run length, then resample")
```

Under *Spatially varying thresholds*, the code comments showing the thresholds do not

```
# The tasmin threshold is 15°C for the northern half of the domain and 20°C for the southern half.
# (notice that the lat coordinate is in decreasing order : from north to south)
thresh_tasmin = xr.DataArray(
    [7] * 24 + [11] * 24, dims=("lat",), coords={"lat": ds5.lat}, attrs={"units": "°C"}
)
# The tasmax threshold is 16°C for the western half of the domain and 19°C for the eastern half.
thresh_tasmax = xr.DataArray(
    [17] * 24 + [21] * 24, dims=("lon",), coords={"lon": ds5.lon}, attrs={"units": "°C"}
)
```

don't appear to match the values used in the code. I assume the code comments just need to be updated.

###### Ensemble-Reductinon Techniques

Only because the utility of the rest of the package is so high, I was a bit confused with the need to manually create a 2D array of criteria (values) and realizations (runs/simulations).  While I think I understand why all dimensions/variables, other the realization id, need to be flattened into a 2D array to apply an ensemble reduction technique, would it be possible to create a helper function that interates through all non-realization 'dimensions', and all variables to generalize the code snippet given below?

```
# Create 2d xr.DataArray containing criteria values
crit = None
for h in ds_crit.horizon:
    for v in ds_crit.data_vars:
        if crit is None:
            crit = ds_crit[v].sel(horizon=h)
        else:
            crit = xr.concat((crit, ds_crit[v].sel(horizon=h)), dim="criteria")
crit.name = "criteria"
```
Is this "criteria" array effectively the equivalent of creating a feature matrix used in data science?

###### Ensemble-Reductinon Techniques

Since, at the moment, "all frequency analysis functions are hard-coded to operate along the `time` dimension", perhaps a brief reference to `xr.DataArray.rename` on how to rename a dimension in xarray would be useful?

I really liked how the worked example for `xclim.indices.stats.frequency_analysis` show both how to use it and fully explains the steps that are bundled together in that convenience function.

###### Extending xclim

 This level of "developer documentation" is well explained in the "user manual" section of the examples documentation clearly shows that xclim has been written in a away that is extensible.
 My previous experience with some other packages that are designed to be extended or have some sort of "plug-in" architecture, the documentation of how to actually use that extensibility is buried in the internal developer-focused documentation (if it exists at all). The authors of xclim should be commended for their clarity of presentation here that will allow others to extend the indicator/indice framework introduced by xclim.

###### Statistical Downscaling and Bias-Adjustment

The content is clear and well explained. Just a couple of very small typos that could be fixed:

>  A more complex example could have **bias** distribution varying strongly across months. To perform the adjustment with different factors for each months, one can pass group='time.month'. Moreover, to reduce the risk of sharp change in the adjustment at the interface of the months, interp='linear' can be passed to adjust and the adjustment factors will be interpolated linearly. Ex: the factors for the 1st of May will be the average of those for **A**pril and those for **M**ay.

and in the following notebook on advanced SDBA:

> The previous notebook covered the most common utilities of xclim.sdba for **conventional** cases

Again, kudos for the depth and richness of examples presented here.

###### Spatial Analogues examples

This is really neat to see this functionality so accessible.Soon after completing this review, I definitely have to run this kind of computation on places that I have personal connections with. And with xclim, I think that will be straightforward to do!
