### 2023-03-27 795d774
Add vignettes to README.rst
