### 2023-03-27 Ouranosinc/xclim#1335 @aulemahal
Nice!
I didn't read everything but indeed, back in the day,  @tlvu warned us of using another git repo for the testing data.
On one hand, I'm not convinced we need to be able to run tests on older versions. In theory, at the time of release, all tests were passing, no? On the other one, I realize we use the testing data in the notebooks and not being able to reproduce _those_ seems more problematic to me.

Thus indeed, I guess that tagging a testdata version would help solve this!
(Long, I tagged you in cas you had any advice?  This comment refers to the first box of section **Functionality** above)
