### 2023-03-27 dddf37d
add climpred and GeoCAT to similar projects,
