+++ docs/installation.rst
@@ -6,13 +6,14 @@ Installation
 
 Stable release
 --------------
-To install xclim via pip, run this command in your terminal:
+
+To install `xclim` via `pip`, run this command in your terminal:
 
 .. code-block:: shell
 
     $ pip install xclim
 
-This is the preferred method to install xclim, as it will always install the most recent stable release.
+This is the preferred method to install `xclim`, as it will always install the most recent stable release.
 
 If you don't have `pip`_ installed, this `Python installation guide`_ can guide you through the process.
 
@@ -21,8 +22,9 @@ If you don't have `pip`_ installed, this `Python installation guide`_ can guide 
 
 Anaconda release
 ----------------
+
 For ease of installation across operating systems, we also offer an Anaconda Python package hosted on conda-forge.
-This version tends to be updated at around the same frequency as the pip library, but can lag by a few days at times.
+This version tends to be updated at around the same frequency as the PyPI-hosted library, but can lag by a few days at times.
 
 `xclim` can be installed from conda-forge wth the following:
 
@@ -32,15 +34,41 @@ This version tends to be updated at around the same frequency as the pip library
 
 .. _extra-dependencies:
 
-Extra dependencies
+Extra Dependencies
 ------------------
-To improve performance of `xclim`, we highly recommend you also install `flox`_ (see: :doc:`flox API <flox:api>`).
-This package integrates into xarray and significantly improves the performance of the grouping and resampling algorithms, especially when using `dask` on large datasets.
 
-We also recommend using the subsetting tools found in `clisops`_ (see: :doc:`clisops.core.subset API <clisops:api>`) for spatial manipulation of geospatial data.
+Speedups and Helper Libraries
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+To improve performance of `xclim`, we highly recommend you also install `flox`_ (see: :doc:`flox API <flox:api>`).
+This package seamlessly integrates into `xarray` and significantly improves the performance of the grouping and resampling algorithms, especially when using `dask` on large datasets.
+
+For grid subsetting, we also recommend using the tools found in `clisops`_ (see: :doc:`clisops.core.subset API <clisops:api>`) for spatial manipulation of geospatial data. `clisops` began as a component of `xclim` and is designed to alongside `xclim` and the `Pangeo`_ stack (`xarray`, `dask`, `jupyter`). In order to install `clisops`, the `GDAL`_ system libraries must be available.
+
+On Debian/Ubuntu, `GDAL` can be installed via `apt`:
+
+.. code-block:: shell
+
+    $ sudo apt-get install libgdal-dev
+
+If on Anaconda Python, `GDAL` will be installed if needed as a `clisops` dependency.
+
+Both of these libraries are available on PyPI and conda-forge:
+
+.. code-block:: shell
+
+    $ pip install flox clisops
+    # Or, alternatively:
+    $ conda install -c conda-forge flox clisops
+
+.. _GDAL: https://gdal.org/download.html#binaries
+.. _Pangeo: https://pangeo.io/
+
+Upstream Dependencies
+^^^^^^^^^^^^^^^^^^^^^
 
 `xclim` is regularly tested against the main development branches of a handful of key base libraries (`cftime`, `flox`, `pint`, `xarray`).
-For convenience, these libraries can be installed alongside `xclim` using the following `pip`-installable recipe:
+For convenience, these libraries can be installed alongside `xclim` using the following `pip`-install command:
 
 .. code-block:: shell
 
@@ -55,11 +83,15 @@ Or, alternatively:
 .. _flox: https://github.com/xarray-contrib/flox
[…]
