### 2023-03-27 641facd
specify what is installed in environment.yml
