+++ CHANGES.rst
@@ -12,7 +12,7 @@ New features and enhancements
     * ``xclim.core.calendar.yearly_interpolated_doy``
     * ``xclim.core.calendar.yearly_random_doy``
 * `scipy` is no longer pinned below v1.9 and `lmoments3>=1.0.5` is now a core dependency and installed by default with `pip`. (:issue:`1142`, :pull:`1171`).
-* Fix bug on number of bins in ``xclim.sdba.propeties.spatial_correlogram``. (:pull:`1334`)
+* Fix bug on number of bins in ``xclim.sdba.propeties.spatial_correlogram``. (:pull:`1336`)
 
 Bug fixes
 ^^^^^^^^^
