### 2023-03-27 f1ffaa9
add some seaborn-based graphical examples
