+++ pyproject.toml
@@ -83,6 +83,7 @@ dev = [
   "nc-time-axis",
   "netCDF4 >=1.4",
   "pooch",
+  "seaborn",
   "sphinx",
   "sphinx-autodoc-typehints",
   "sphinx-codeautolink",
