### 2023-03-28 Ouranosinc/xclim#1335 @tlvu
My previous worries about splitting the testdata with the code https://github.com/Ouranosinc/xclim-testdata/pull/1#issuecomment-704281014

So tagging the testdata should solve this reproducibility issue.

But to ensure smooth dev workflow, the code should allow overriding the tag with a branch name.  During dev cycle, both the testdata and the code would most probably move together.  Without the override capability, we will have to continuously tag the testdata so it can be used with the code and this can get tedious.

However, with this tag override capability, we must not forget to tag the final version of the testdata and bump that tag on the code side before merging.  The tag should be the default value when no override is used.
