### 2023-03-28 69cb07c
update ubuntu image names, add `XCLIM_TESTDATA_BRANCH` to upstream.yml
