+++ .github/workflows/label_on_approval.yml
@@ -8,7 +8,7 @@ on:
 jobs:
   approved:
     if: github.event.review.state == 'approved'
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/github-script@v6
         with:
