+++ .github/workflows/first_pull_request.yml
@@ -5,7 +5,7 @@ on:
 
 jobs:
   welcome:
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/github-script@v6
         with:
