+++ .github/workflows/bump-version.yml
@@ -26,7 +26,7 @@ on:
 
 jobs:
   bump_patch_version:
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
         with:
