+++ .github/workflows/testdata_version.yml
@@ -18,11 +18,11 @@ jobs:
       - uses: actions/checkout@v3
       - name: Set latest testing data tag
         run: |
-          TAG=( \
+          TAG="$( \
           git -c 'versionsort.suffix=-' \
           ls-remote --exit-code --refs --sort='version:refname' --tags https://github.com/Ouranosinc/xclim '*.*.*' \
           | tail --lines=1 \
-          | cut --delimiter='/' --fields=3)
+          | cut --delimiter='/' --fields=3)"
           echo "TAG=${TAG}" >> $GITHUB_ENV
       - name: Report tag
         run: echo $TAG
