### 2023-03-29 72c1ddd
report testing data used in main.yml
