+++ .github/workflows/testdata_version.yml
@@ -19,18 +19,24 @@ jobs:
         run: |
           TAG="$( \
           git -c 'versionsort.suffix=-' \
-          ls-remote --exit-code --refs --sort='version:refname' --tags https://github.com/Ouranosinc/xclim '*.*.*' \
+          ls-remote --exit-code --refs --sort='version:refname' --tags https://github.com/Ouranosinc/xclim-testdata '*.*.*' \
           | tail --lines=1 \
           | cut --delimiter='/' --fields=3)"
           echo "TAG=${TAG}" >> $GITHUB_ENV
-      - name: Report tag
-        run: echo $TAG
       - name: Gather CI Test Data Version
         run: |
           TESTDATA_CI="$(grep -E "XCLIM_TESTDATA_BRANCH" .github/workflows/main.yml | cut -d ' ' -f4)"
           echo "TESTDATA_CI=${TESTDATA_CI}" >> $GITHUB_ENV
-      - name: Report Testing Data in CI
-        run: echo $TESTDATA_CI
+      - name: Report Versions Found
+        run: |
+          echo "Latest xclim-testdata tag: ${{ env.TAG }}"
+          echo "Tag for xclim-testdata in CI: ${{ env.TESTDATA_CI }}"
+      - name: Compare Versions
+        if: ( ${{ env.TAG }} == ${{ env.TESTDATA_CI }} )
+        uses: actions/github-script@v6
+        with:
+          script: |
+            core.setFailed('Configured `xclim-testdata` tag is not `latest`.')
       - name: Find Comment
         uses: peter-evans/find-comment@v2
         id: fc
@@ -39,12 +45,6 @@ jobs:
           comment-author: 'github-actions[bot]'
           body-includes: |
             > **Warning**
-      - name: Check Versions
-        if: ( ${{ env.TAG }} == ${{ env.TESTDATA_CI }} )
-        uses: actions/github-script@v6
-        with:
-          script: |
-            core.setFailed('Configured `xclim-testdata` tag is not the latest.')
       - name: Update Failure Comment
         if: ${{ failure() }}
         uses: peter-evans/create-or-update-comment@v2
@@ -55,7 +55,7 @@ jobs:
             > **Warning**
             > It appears that this PR modifies the `Ouranosinc/xclim-testdata` branch to a tag that is not the latest in the repository.
 
-            Please make sure to modify this value to match the latest tag before merging.
+            Please make sure to modify this value to match the latest tag (`${{ env.TAG }}`) before merging.
             If this PR depends on a new testing dataset, be sure to tag a new version of `Ouranosinc/xclim-testdata` with your changes merged to `main`.
           edit-mode: replace
       - name: Update Success Comment
@@ -66,7 +66,5 @@ jobs:
           issue-number: ${{ github.event.pull_request.number }}
           body: |
             > **Warning**
-            > It appears that this PR updates the `Ouranosinc/xclim-testdata` branch to the most recent tag.
-
-            No further action is required.
+            > It appears that this PR updates the `Ouranosinc/xclim-testdata` branch to the most recent tag. No further action is required.
           edit-mode: replace
