+++ .github/workflows/testdata_version.yml
@@ -15,36 +15,33 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - name: Set latest testing data tag
+      - name: Find xclim-testdata Tag and CI Testing Branch
         run: |
-          TAG="$( \
+          XCLIM_TESTDATA_TAG="$( \
           git -c 'versionsort.suffix=-' \
           ls-remote --exit-code --refs --sort='version:refname' --tags https://github.com/Ouranosinc/xclim-testdata '*.*.*' \
           | tail --lines=1 \
           | cut --delimiter='/' --fields=3)"
-          echo "TAG=${TAG}" >> $GITHUB_ENV
-      - name: Gather CI Test Data Version
-        run: |
-          TESTDATA_CI="$(grep -E "XCLIM_TESTDATA_BRANCH" .github/workflows/main.yml | cut -d ' ' -f4)"
-          echo "TESTDATA_CI=${TESTDATA_CI}" >> $GITHUB_ENV
+          echo "XCLIM_TESTDATA_TAG=${XCLIM_TESTDATA_TAG}" >> $GITHUB_ENV
+          XCLIM_TESTDATA_BRANCH="$(grep -E "XCLIM_TESTDATA_BRANCH" .github/workflows/main.yml | cut -d ' ' -f4)"
+          echo "XCLIM_TESTDATA_BRANCH=${XCLIM_TESTDATA_BRANCH}" >> $GITHUB_ENV
       - name: Report Versions Found
         run: |
-          echo "Latest xclim-testdata tag: ${{ env.TAG }}"
-          echo "Tag for xclim-testdata in CI: ${{ env.TESTDATA_CI }}"
-      - name: Compare Versions
-        if: ( ${{ env.TAG }} == ${{ env.TESTDATA_CI }} )
-        uses: actions/github-script@v6
-        with:
-          script: |
-            core.setFailed('Configured `xclim-testdata` tag is not `latest`.')
+          echo "Latest xclim-testdata tag: ${{ env.XCLIM_TESTDATA_TAG }}"
+          echo "Tag for xclim-testdata in CI: ${{ env.XCLIM_TESTDATA_BRANCH }}"
       - name: Find Comment
         uses: peter-evans/find-comment@v2
         id: fc
         with:
           issue-number: ${{ github.event.pull_request.number }}
           comment-author: 'github-actions[bot]'
-          body-includes: |
-            > **Warning**
+          body-includes: It appears that this PR modifies the `Ouranosinc/xclim-testdata` branch
+      - name: Compare Versions
+        if: ( ${{ env.XCLIM_TESTDATA_TAG }} == ${{ env.XCLIM_TESTDATA_BRANCH }} )
+        uses: actions/github-script@v6
+        with:
+          script: |
+            core.setFailed('Configured `xclim-testdata` tag is not `latest`.')
       - name: Update Failure Comment
         if: ${{ failure() }}
         uses: peter-evans/create-or-update-comment@v2
@@ -55,7 +52,8 @@ jobs:
             > **Warning**
             > It appears that this PR modifies the `Ouranosinc/xclim-testdata` branch to a tag that is not the latest in the repository.
 
-            Please make sure to modify this value to match the latest tag (`${{ env.TAG }}`) before merging.
+            Please make sure to modify this value to match the latest tag (`${{ env.XCLIM_TESTDATA_TAG }}`) before merging.
+
             If this PR depends on a new testing dataset, be sure to tag a new version of `Ouranosinc/xclim-testdata` with your changes merged to `main`.
           edit-mode: replace
       - name: Update Success Comment
@@ -66,5 +64,5 @@ jobs:
           issue-number: ${{ github.event.pull_request.number }}
           body: |
[…]
