+++ .github/workflows/testdata_version.yml
@@ -35,10 +35,9 @@ jobs:
         with:
           issue-number: ${{ github.event.pull_request.number }}
           comment-author: 'github-actions[bot]'
-          body-includes: It appears that this PR modifies the `Ouranosinc/xclim-testdata` branch
+          body-includes: It appears that this PR modifies the `XCLIM_TESTDATA_BRANCH` environment variable
       - name: Compare Versions
-#        if: ( ${{ env.XCLIM_TESTDATA_TAG }} != ${{ env.XCLIM_TESTDATA_BRANCH }} )
-        if:  ${{ failure() }}
+        if: ( ${{ env.XCLIM_TESTDATA_TAG }} != ${{ env.XCLIM_TESTDATA_BRANCH }} )
         uses: actions/github-script@v6
         with:
           script: |
@@ -51,11 +50,11 @@ jobs:
           issue-number: ${{ github.event.pull_request.number }}
           body: |
             > **Warning**
-            > It appears that this PR modifies the `Ouranosinc/xclim-testdata` branch to a tag that is not the latest in the repository.
+            > It appears that this PR modifies the `XCLIM_TESTDATA_BRANCH` environment variable to a tag that is not the latest in the `Ouranosinc/xclim-testdata` repository.
 
-            Please make sure to modify this value to match the latest tag (`${{ env.XCLIM_TESTDATA_TAG }}`) before merging.
+            Please be sure to modify this value to match the most recent tag (`${{ env.XCLIM_TESTDATA_TAG }}`) before merging.
 
-            If this PR depends on a new testing dataset, be sure to tag a new version of `Ouranosinc/xclim-testdata` with your changes merged to `main`.
+            If this PR depends on changes in a new testing dataset branch, be sure to tag a new version of `Ouranosinc/xclim-testdata` with your changes merged to `main`.
           edit-mode: replace
       - name: Update Success Comment
         if: ${{ success() }}
@@ -64,6 +63,8 @@ jobs:
           comment-id: ${{ steps.fc.outputs.comment-id }}
           issue-number: ${{ github.event.pull_request.number }}
           body: |
-            > **Warning**
-            > It appears that this PR modifies the `Ouranosinc/xclim-testdata` branch to the most recent tag (`${{ env.XCLIM_TESTDATA_TAG }}`). No further action is required.
+            > **Note**
+            > It appears that this PR modifies the `XCLIM_TESTDATA_BRANCH` environment variable to the most recent tag (`${{ env.XCLIM_TESTDATA_TAG }}`).
+
+            No further action is required.
           edit-mode: replace
