+++ .github/workflows/testdata_version.yml
@@ -1,4 +1,4 @@
-name: "Update Testing Data Tag"
+name: "Verify Testing Data"
 
 on:
   pull_request:
@@ -6,12 +6,12 @@ on:
       - opened
       - reopened
       - synchronize
-#    paths:
-#      - .github/workflows/main.yml
+    paths:
+      - .github/workflows/main.yml
 
 jobs:
   use-latest-tag:
-    name: Use Latest Testing Data
+    name: Check Latest xclim-testdata Tag
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
