+++ .github/workflows/publish-pypi.yml
@@ -1,4 +1,4 @@
-name: "Publish Python 🐍 distributions 📦 to PyPI"
+name: Publish Python 🐍 distributions 📦 to PyPI
 
 on:
   release:
