+++ .github/workflows/tag-testpypi.yml
@@ -1,4 +1,4 @@
-name: "Publish Python 🐍 distributions 📦 to TestPyPI"
+name: Publish Python 🐍 distributions 📦 to TestPyPI
 
 on:
   push:
