+++ .github/workflows/label_on_approval.yml
@@ -1,4 +1,4 @@
-name: "Label Pull Request on Pull Request review approval"
+name: Label Pull Request on Pull Request review approval
 
 on:
   pull_request_review:
@@ -7,6 +7,7 @@ on:
 
 jobs:
   approved:
+    name: Label on Approval
     if: github.event.review.state == 'approved'
     runs-on: ubuntu-latest
     steps:
