### 2023-03-29 cb91d25
fix comparison logic
