### 2023-03-29 f39ef18
fix comparison logic
