### 2023-03-29 ca019a7
add a warning for users of stable versions using main branch of xclim-testdata
