### 2023-03-30 f6c3beb
Revert to "last run" returns "last element", fix test
