### 2023-03-30 f345583
Remove mention of non-existant option
