### 2023-03-30 b8a5991
remove eigen and pybind11 (needed for sbck only)
