+++ docs/installation.rst
@@ -13,8 +13,6 @@ To install `xclim` via `pip`, run this command in your terminal:
 
     $ pip install xclim
 
-This is the preferred method to install `xclim`, as it will always install the most recent stable release.
-
 If you don't have `pip`_ installed, this `Python installation guide`_ can guide you through the process.
 
 .. _pip: https://pip.pypa.io/
