### 2023-03-30 Ouranosinc/xclim#1335 @Zeitsperre
@tlvu

Thanks for the suggestion on how best to proceed for this. We now have a testing data tagging scheme and some GitHub Actions to prevent us from accidentally breaking it during some development branch. It's nothing fancy, but we should be able to more easily test older versions of `xclim` going forward.
