### 2023-03-30 ddeaaa1
Add make_criteria helper
