### 2023-03-31 b1bb065
Add french RAIN_SEASON and units days
