### 2023-03-31 72d7877
Revert changes test_locales
