### 2023-03-31 134633e
upd chngs - complete version preserving coords and allowing different dims on vars
