### 2023-03-31 f566018
try loading dataset to prevent crash
