### 2023-03-31 057ed99
try loading dataset to prevent crash
