### 2023-03-31 78dfaf9
Bump version: 0.42.0-beta → 0.42.0
