### 2023-03-31 Ouranosinc/xclim#1335 @Zeitsperre
Except for #1342, we managed to address all major comments in under a work-week. Nicely done, team!
