### 2023-04-03 227ac1c
Ignore dimension order when uncriterionning the criteria
