+++ xclim/testing/tests/test_sdba/test_properties.py
@@ -13,7 +13,7 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1980"), location="Vancouver")
             .pr
-        )
+        ).load()
 
         out_year = sdba.properties.mean(sim)
         np.testing.assert_array_almost_equal(out_year.values, [3.0016028e-05])
@@ -31,7 +31,7 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1980"), location="Vancouver")
             .pr
-        )
+        ).load()
 
         out_year = sdba.properties.var(sim)
         np.testing.assert_array_almost_equal(out_year.values, [2.5884779e-09])
@@ -72,7 +72,7 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1980"), location="Vancouver")
             .pr
-        )
+        ).load()
 
         out_year = sdba.properties.quantile(sim, q=0.2)
         np.testing.assert_array_almost_equal(out_year.values, [2.8109431013945154e-07])
@@ -94,7 +94,8 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1952"), location="Vancouver")
             .pr
-        )
+        ).load()
+
         tmean = (
             sdba.properties.spell_length_distribution(sim, op="<", group="time.month")
             .sel(month=1)
@@ -121,7 +122,8 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1952"), location="Vancouver")
             .tasmax
-        )
+        ).load()
+
         tmean = sdba.properties.spell_length_distribution(
             simt, op=">=", group="time.month", method="quantile", thresh=0.9
         ).sel(month=6)
@@ -168,7 +170,8 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1952"), location="Vancouver")
             .pr
-        )
+        ).load()
+
         out = sdba.properties.acf(sim, lag=1, group="time.month").sel(month=1)
         np.testing.assert_array_almost_equal(out.values, [0.11242357313756905])
 
@@ -183,7 +186,8 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1952"), location="Vancouver")
             .tasmax
-        )
+        ).load()
+
         amp = sdba.properties.annual_cycle_amplitude(simt)
         relamp = sdba.properties.relative_annual_cycle_amplitude(simt)
         phase = sdba.properties.annual_cycle_phase(simt)
@@ -216,7 +220,8 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1952"), location="Vancouver")
             .tasmax
-        )
+        ).load()
+
         # Initial annual cycle was this with window = 1
         amp = sdba.properties.mean_annual_range(simt, window=1)
         relamp = sdba.properties.mean_annual_relative_range(simt, window=1)
@@ -258,12 +263,14 @@ class TestProperties:
             open_dataset("sdba/CanESM2_1950-2100.nc")
             .sel(time=slice("1950", "1952"), location="Vancouver")
             .tasmax
-        )
+        ).load()
+
         sim = (
             open_dataset("sdba/CanESM2_1950-2100.nc")
[…]
