+++ xclim/sdba/base.py
@@ -292,7 +292,10 @@ class Grouper(Parametrizable):
             return da[self.dim].rename("group")
 
         ind = da.indexes[self.dim]
-        i = getattr(ind, self.prop)
+        if self.prop == "week":
+            i = ind.isocalendar().week
+        else:
+            i = getattr(ind, self.prop)
 
         if not np.issubdtype(i.dtype, np.integer):
             raise ValueError(
