+++ pyproject.toml
@@ -40,7 +40,8 @@ dependencies = [
   "lmoments3>=1.0.5",
   "numba",
   "numpy>=1.16",
-  "pandas>=0.23",
+  "pandas>=0.23,<2.0; python_version == '3.8'",
+  "pandas>=0.23; python_version >= '3.9'",
   "pint>=0.10",
   "pyyaml",
   "scikit-learn>=0.21.3",
