### 2023-04-03 ffa3485
Fix dtype handling
