### 2023-04-03 adca72e
add h5nc to pip too
