+++ pyproject.toml
@@ -59,10 +59,12 @@ dev = [
   "coverage[toml]",
   "flake8",
   "flake8-rst-docstrings",
+  "h5netcdf",
   "ipython",
   "mypy",
   "nbqa",
   "nbval",
+  "netCDF4 >=1.4",
   "pre-commit >=2.9",
   "pydocstyle >=5.1.1",
   "pybtex",
@@ -82,7 +84,6 @@ dev = [
   "matplotlib",
   "nbsphinx",
   "nc-time-axis",
-  "netCDF4 >=1.4",
   "pooch",
   "sphinx",
   "sphinx-autodoc-typehints",
