+++ .pre-commit-config.yaml
@@ -23,7 +23,7 @@ repos:
   - id: pretty-format-json
     args: ['--autofix', '--no-ensure-ascii', '--no-sort-keys']
 - repo: https://github.com/pappasam/toml-sort
-  rev: v0.22.4
+  rev: v0.23.0
   hooks:
   - id: toml-sort-fix
 - repo: https://github.com/pycqa/flake8
@@ -33,7 +33,7 @@ repos:
     additional_dependencies: ['flake8-rst-docstrings']
     args: ['--config=setup.cfg']
 - repo: https://github.com/psf/black
-  rev: 23.1.0
+  rev: 23.3.0
   hooks:
   - id: black
 - repo: https://github.com/PyCQA/isort
@@ -42,7 +42,7 @@ repos:
   - id: isort
     args: ['--settings-file=pyproject.toml', "--add_imports='from __future__ import annotations'"]
 - repo: https://github.com/nbQA-dev/nbQA
-  rev: 1.6.3
+  rev: 1.7.0
   hooks:
   - id: nbqa-black
     args: ['--target-version=py38']
@@ -66,9 +66,9 @@ repos:
   - id: blackdoc
     exclude: '(xclim/indices/__init__.py|docs/installation.rst)'
     additional_dependencies:
-    - 'black==23.1.0'
+    - 'black==23.3.0'
 - repo: https://github.com/adrienverge/yamllint.git
-  rev: v1.29.0
+  rev: v1.30.0
   hooks:
   - id: yamllint
     args: ['--config-file=.yamllint.yaml']
