+++ .pre-commit-config.yaml
@@ -26,16 +26,21 @@ repos:
   rev: v0.23.0
   hooks:
   - id: toml-sort-fix
+- repo: https://github.com/adrienverge/yamllint.git
+  rev: v1.30.0
+  hooks:
+  - id: yamllint
+    args: ['--config-file=.yamllint.yaml']
+- repo: https://github.com/psf/black
+  rev: 23.3.0
+  hooks:
+  - id: black
 - repo: https://github.com/pycqa/flake8
   rev: 6.0.0
   hooks:
   - id: flake8
     additional_dependencies: ['flake8-rst-docstrings']
     args: ['--config=setup.cfg']
-- repo: https://github.com/psf/black
-  rev: 23.3.0
-  hooks:
-  - id: black
 - repo: https://github.com/PyCQA/isort
   rev: 5.12.0
   hooks:
@@ -67,11 +72,6 @@ repos:
     exclude: '(xclim/indices/__init__.py|docs/installation.rst)'
     additional_dependencies:
     - 'black==23.3.0'
-- repo: https://github.com/adrienverge/yamllint.git
-  rev: v1.30.0
-  hooks:
-  - id: yamllint
-    args: ['--config-file=.yamllint.yaml']
 - repo: meta
   hooks:
   - id: check-hooks-apply
