### 2023-04-04 #73 @Zeitsperre
Hello @Batalex and @jmunroe

I hope you both are well! We managed to hammer out as many of the smaller review comments as we could, and have marked the comments concerning the behaviour of `frequency_analysis` as something that necessitates some deliberation in the coming months.

This has all been merged to our main development branch in time for our latest release ([version 0.42.0](https://xclim.readthedocs.io/en/stable/changes.html#v0-42-0-2023-04-03))

If you could please confirm that these changes meet your expectations, we would greatly appreciate it. Please let us know if there are any other modifications required.

Best,
