### 2023-04-04 #73 @Batalex
Hello @Zeitsperre,
I am impressed by Xclim's reactivity!

In addition to Agustina and James' remarks, I'd like to add a little comment. We recommend separating unit tests from source code or, at least, excluding them from the final build.

As for the review, @aguspesce, @jmunroe, could you please make a final check as to whether you are satisfied with how your comments were dealt with?
You can go through the [changelog](https://xclim.readthedocs.io/en/stable/changes.html), the Xclim's team was nice enough to highlight the changes brought by this submission.

Once you are done, you may check the `The author has responded to my review and made changes to my satisfaction. I recommend approving this package.` item in your review.
