### 2023-04-04 3201bc4
Simplified rle_with_holes and some corner cases better treated
