### 2023-04-04 6782e89
More specific test for `rle_with_holes`
