### 2023-04-05 #73 @Zeitsperre
> In addition to Agustina and James' remarks, I'd like to add a little comment. We recommend separating unit tests from source code or, at least, excluding them from the final build.

Addressing this comment would be a bit tricky; As a second layer of caution, we currently run tests on the `sdist` when building our binary packages on conda-forge. Previously, using `setuptools`, we could more easily exclude the `xclim/testing/tests` folder when building the `wheel`, but it seems like the new backend we are using (`flit`) doesn't easily allow that level of control.

I'll open an issue at https://github.com/pypa/flit to see if that feature is maybe just not documented (i.e. Optional exclusion of files/folders depending on build distribution (`sdist` |`bdist_wheel`)). If this is blocking acceptance to pyOpenSci, I can try getting this done for the next release.

Thanks again!
