### 2023-04-05 Ouranosinc/xclim#1348 @Zeitsperre
[PyOpenSci] Split `tests` folder out of `xclim` source code
The PyOpenSci editor has one final suggestion for us:

> We recommend separating unit tests from source code or, at least, excluding them from the final build.

This is tricky to deal with; As a second layer of caution, we currently run tests on the `sdist` when building our binary packages on conda-forge. Previously, using `setuptools`, we could more easily exclude the `xclim/testing/tests` folder when building the `wheel`, but it seems like the new backend we are using (`flit`) doesn't easily allow that level of control.

One thing we could is to start by splitting the `tests` folder out of the package until we can figure out how best to indicate to `flit` which files should not be specifically in the `wheel`. This would probably not be too difficult to do in time for the next release.

Original comment: https://github.com/pyOpenSci/software-submission/issues/73#issuecomment-1496466252
