### 2023-04-05 #73 @lwasser
Hi everyone! 👋 this has been a wonderful review to watch unfold!

A note about tests - [for reference, here is our language around testing in our packaging guide](https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-structure.html#the-src-layout-and-testing).

For many scientific packages, it is important to have tests included in distribution archives. This is particularly important if you have complex wheel builds supporting multiple architectures, etc.. In this case it sounds like it's important to the conda-forge build. While we suggest that unit tests live outside of the package directory, we do not require it.

We suggest tests live outside of the package directory to prevent users from running tests on the package file tree rather than in the installed version. And also it makes it easy to exclude tests from distributions that do not need to have tests included.

i wanted to add this note only because it is NOT a hard requirement for pyOpenSci acceptance. We do understand that in some cases tests may need to be a part of the distribution and moving them might be significant burden to the maintainers.

it would be interesting to see if there is  a bit more flexibility with hatchling or pdm-build as a backend. i'm not 100% sure as i haven't tried to exclude specific parts of the package dir with them. And, I don't want to cause anyone here more / unnecessary work!  So if this request is a pain point for xclim, then it is ok to move forward without this specific change.

Just for information gathering purposes, i am going to ask the hatchling dev about that specific feature as well. :)
