### 2023-04-05 Ouranosinc/xclim#1348 @huard
If I remember right, code for tests is in the distribution so it can be used in the documentation.
