### 2023-04-05 Ouranosinc/xclim#1348 @Zeitsperre
It's confusing:

- The code to run the tests is part of the package, under `xclim/testing/*.py`
- The tests themselves are under `xclim/testing/tests/**.py`

The proposal here would be to move `xclim/testing/tests/` to the top-level and see if we can selectively exclude them when building the `wheel` (and **not** the source distribution).
