### 2023-04-05 e814f56
split tests folder from xclim package, move testing helpers to xclim/testing, adjust pytest options and calls
