+++ xclim/testing/__init__.py
@@ -1,4 +1,5 @@
 """Helpers for testing xclim."""
 from __future__ import annotations
 
+from . import helpers
 from .utils import *
