+++ tox.ini
@@ -27,13 +27,13 @@ deps =
     yamllint
 commands_pre =
 commands =
-    pydocstyle --config=setup.cfg xclim
-    flake8 --config=setup.cfg xclim
-    black --check xclim
+    pydocstyle --config=setup.cfg xclim tests
+    flake8 --config=setup.cfg xclim tests
+    black --check xclim tests
     nbqa black --check docs
     blackdoc --check --exclude=xclim/indices/__init__.py xclim
     blackdoc --check docs
-    isort --check xclim
+    isort --check xclim tests
     yamllint --config-file=.yamllint.yaml xclim
 commands_post =
 
@@ -62,7 +62,7 @@ allowlist_externals =
 description = Run notebooks and doctests with pytest under {basepython}
 commands =
     pytest --no-cov --nbval --dist=loadscope docs/notebooks --ignore=docs/notebooks/example.ipynb
-    pytest --rootdir xclim/testing/tests/ --xdoctest xclim --ignore=xclim/testing/tests/
+    pytest --rootdir=tests/ --xdoctest xclim
 
 # Requires tox-conda compatible with tox@v4.0
 ;[testenv:opt-{slow,not_slow}]
@@ -101,7 +101,7 @@ commands_pre =
     python -m pip list
     python -m pip check
 commands =
-    doctest: pytest --no-cov --rootdir xclim/testing/tests/ --xdoctest xclim --ignore=xclim/testing/tests/
+    doctest: pytest --no-cov --rootdir=tests/ --xdoctest xclim
     !slow: pytest xclim -m "not slow"
     slow: pytest xclim
 commands_post =
