### 2023-04-05 8e51870
fix circular imports, adjust notebook tests
