+++ xclim/testing/__init__.py
@@ -1,5 +1,6 @@
 """Helpers for testing xclim."""
 from __future__ import annotations
 
-from . import helpers
+from . import diagnostics, helpers
+from .sdba_utils import *
 from .utils import *
