+++ tox.ini
@@ -61,7 +61,7 @@ allowlist_externals =
 [testenv:notebooks_doctests{-coverage,}]
 description = Run notebooks and doctests with pytest under {basepython}
 commands =
-    pytest --no-cov --nbval --dist=loadscope docs/notebooks --ignore=docs/notebooks/example.ipynb
+    pytest --no-cov --nbval --dist=loadscope --rootdir=tests/ docs/notebooks --ignore=docs/notebooks/example.ipynb
     pytest --rootdir=tests/ --xdoctest xclim
 
 # Requires tox-conda compatible with tox@v4.0
