+++ tests/conftest.py
@@ -19,17 +19,11 @@ import xclim
 from xclim import __version__ as __xclim_version__
 from xclim.core import indicator
 from xclim.core.calendar import max_doy
-from xclim.testing.helpers import (
-    PREFETCH_TESTING_DATA,
-    TESTDATA_BRANCH,
-    add_example_file_paths,
-    generate_atmos,
-    populate_testing_data,
-)
+from xclim.testing import helpers
 from xclim.testing.utils import _default_cache_dir  # noqa
 from xclim.testing.utils import open_dataset as _open_dataset
 
-if not __xclim_version__.endswith("-dev") and TESTDATA_BRANCH == "main":
+if not __xclim_version__.endswith("-dev") and helpers.TESTDATA_BRANCH == "main":
     # This is fine on GitHub Workflows and ReadTheDocs
     if not os.getenv("CI") and not os.getenv("READTHEDOCS"):
         warnings.warn(
@@ -39,10 +33,10 @@ if not __xclim_version__.endswith("-dev") and TESTDATA_BRANCH == "main":
             UserWarning,
         )
 
-if re.match(r"^v\d+\.\d+\.\d+", TESTDATA_BRANCH):
-    if parse_version(TESTDATA_BRANCH) > parse_version(__xclim_version__):
+if re.match(r"^v\d+\.\d+\.\d+", helpers.TESTDATA_BRANCH):
+    if parse_version(helpers.TESTDATA_BRANCH) > parse_version(__xclim_version__):
         warnings.warn(
-            f"`xclim` version ({__xclim_version__}) predates `xclim-testdata` version ({TESTDATA_BRANCH}). "
+            f"`xclim` version ({__xclim_version__}) predates `xclim-testdata` version ({helpers.TESTDATA_BRANCH}). "
             f"It is very likely that the testing data is incompatible with this build of `xclim`.",
             UserWarning,
         )
@@ -540,7 +534,7 @@ def cmip3_day_tas(threadsafe_data_dir):
     ds = _open_dataset(
         "cmip3/tas.sresb1.giss_model_e_r.run1.atm.da.nc",
         cache_dir=threadsafe_data_dir,
-        branch=TESTDATA_BRANCH,
+        branch=helpers.TESTDATA_BRANCH,
         engine="h5netcdf",
     )
     yield ds.tas
@@ -550,7 +544,7 @@ def cmip3_day_tas(threadsafe_data_dir):
 @pytest.fixture(scope="session")
 def open_dataset(threadsafe_data_dir):
     def _open_session_scoped_file(
-        file: str | os.PathLike, branch: str = TESTDATA_BRANCH, **xr_kwargs
+        file: str | os.PathLike, branch: str = helpers.TESTDATA_BRANCH, **xr_kwargs
     ):
         xr_kwargs.setdefault("engine", "h5netcdf")
         return _open_dataset(
@@ -568,7 +562,7 @@ def add_imports(xdoctest_namespace, threadsafe_data_dir) -> None:
     ns["xr"] = xclim.testing  # xr.open_dataset(...) -> xclim.testing.open_dataset(...)
     ns["xclim"] = xclim
     ns["open_dataset"] = partial(
-        _open_dataset, cache_dir=threadsafe_data_dir, branch=TESTDATA_BRANCH
+        _open_dataset, cache_dir=threadsafe_data_dir, branch=helpers.TESTDATA_BRANCH
     )  # Needed for modules where xarray is imported as `xr`
 
 
@@ -607,7 +601,7 @@ def atmosds(threadsafe_data_dir) -> xr.Dataset:
     return _open_dataset(
         threadsafe_data_dir.joinpath("atmosds.nc"),
[…]
