### 2023-04-05 #73 @Zeitsperre
@lwasser

Thanks for looking into things with the `hatch` backend. Ultimately, I really like the simplicity of `flit`, so I've gone with the approach of splitting the `tests` folder out of the package. After factoring the functions that were mixed into the tests themselves back into the package, the `wheel` is significantly smaller (~140 KB!).

Assuming that changes are accepted, this will be part of the next scheduled release (v0.43.0) in around a month.

Thanks again!
