### 2023-04-05 #73 @lwasser
ok thank you for making that structure change @Zeitsperre !! i totally get it. a lot of folks love the simplicity of flit. it's good for us to know what each tool can and can't do so that is why i thought i'd dig just a bit to understand more :)

and yay for a smaller wheel! PyPI / warehouse maintainers will appreciate that! ✨
