### 2023-04-11 #73 @tlogan2000
This is excellent news! Thanks to all the reviewers and to @Zeitsperre for preparing and coordinating the xclim submission.  Great job.
