### 2023-04-12 #73 @Zeitsperre
This is such fantastic news! Thanks so much to @lwasser, @NickleDave, @Batalex, @aguspesce, and @jmunroe for all of your effort and time guiding us to build a better package!

I was scheduled to deliver a brief internal presentation on news concerning xclim today, so I'll be sharing this news as well :)

We should be opening pull requests soon and adding all associated information and badges to xclim's README and documentation either today or tomorrow, and following up with JOSS by the end of the week. Again, thanks so much for all your work!
