### 2023-05-19 6c6d473
Add JOSS Article, Review badge and Citation information (#250)
### What kind of change does this PR introduce?

This PR finalizes the submission and publishing steps for the [Journal
of Open Source
Software](https://joss.readthedocs.io/en/latest/index.html). The paper
is expected to have a length of 250 - 1000 words and demands that the
software is *feature-complete*. As such, the submission process should
not be started until we have at the very least added all necessary
indicators from our backlog and/or stabilized our API (i.e.: v1.0-alpha
or release-candidate).

Updates (May 2023): 
With x
[…]
