+++ AUTHORS.rst
@@ -12,6 +12,7 @@ Co-Developers
 
 * Abel Aoun <aoun.abel@gmail.com> `@bzah <https://github.com/bzah>`_
 * Pascal Bourgault <bourgault.pascal@ouranos.ca> `@aulemahal <https://github.com/aulemahal>`_
+* Éric Dupuis <dupuis.eric@ouranos.ca> `@coxipi <https://github.com/coxipi>`_
 * David Huard <huard.david@ouranos.ca> `@huard <https://github.com/huard>`_
 * Juliette Lavoie <lavoie.juliette@ouranos.ca> `@juliettelavoie <https://github.com/juliettelavoie>`_
 * Gabriel Rondeau-Genesse <rondeau-genesse.gabriel@ouranos.ca> `@RondeauG <https://github.com/RondeauG>`_
@@ -22,17 +23,19 @@ Contributors
 
 * Raquel Alegre <raquel.alegre@gmail.com> `@raquelalegre <https://github.com/raquelalegre>`_
 * Clair Barnes <clair.barnes.16@ucl.ac.uk> `@clairbarnes <https://github.com/clairbarnes>`_
-* Sébastien Biner <biner.sebastien@ouranos.ca> `@sbiner <https://github.com/sbiner>`_
+* Alexis Beaupré-Laperrière `@Beauprel <https://github.com/Beauprel>`_
+* Sébastien Biner `@sbiner <https://github.com/sbiner>`_
 * David Caron `@davidcaron <https://github.com/davidcaron>`_
 * Carsten Ehbrecht <ehbrecht@dkrz.de> `@cehbrecht <https://github.com/cehbrecht>`_
 * Jeremy Fyke `@jeremyfyke <https://github.com/jeremyfyke>`_
 * Tom Keel <thomas.keel.18@ucl.ac.uk> `@Thomasjkeel <https://github.com/Thomasjkeel>`_
+* Marie-Pier Labonté <labonte.marie-pier@ouranos.ca> `@marielabonte <https://github.com/marielabonte>`_
+* Ludwig Lierhammer <ludwig.lierhammer@hereon.de> `@ludwiglierhammer <https://github.com/ludwiglierhammer>`_
 * Jwen Fai Low `@jwenfai <https://github.com/jwenfai>`_
-* Marie-Pier Labonté `@marielabonte <https://github.com/marielabonte>`_
 * Jamie Quinn <jamiejquinn@jamiejquinn.com> `@JamieJQuinn <https://github.com/JamieJQuinn>`_
-* Philippe Roy <roy.philippe@ouranos.ca> `@Balinus <https://github.com/Balinus>`_
 * Yannick Rousseau
+* Philippe Roy `@Balinus <https://github.com/Balinus>`_
 * Dougie Squire <dougiesquire@gmail.com> `@dougiesquire <https://github.com/dougiesquire>`_
 * Ag Stephens <ag.stephens@stfc.ac.uk> `@agstephens <https://github.com/agstephens>`_
 * Maliko Tanguy <malngu@ceh.ac.uk> `@malngu <https://github.com/malngu>`_
-* Alexis Beaupré-Laperrière `@Beauprel <https://github.com/Beauprel>`_
+* Christopher Whelan `@qwhelan <https://github.com/qwhelan>`_
