### 2023-05-19 #73 @Zeitsperre
Hi @Batalex and @lwasser

The article was just published yesterday in JOSS! Here's the review: https://github.com/openjournals/joss-reviews/issues/5415. I think we're good to close this issue!
