### 2023-06-07 #73 @lwasser
hey there @Zeitsperre just following up on this review! i hope the JOSS component is going well. I have a small request - can you take 5-10 minutes to [fill out our post-review survey please?](https://forms.gle/vwENjvjyiwUFzJRz5) i'd greatly appreciate it. The most important part for us it how the review impacted / improved etc your package (if it did) and feedback on the process. many thanks in advance for doing this!!
